Fast, parallel implementation of grid data convolution

A fast, parallel implementation of convolutions for grid-type data (matrices, rasters and other grid based objects).

pfocal(
  data,
  kernel,
  edge_value = 0,
  transform_function = "MULTIPLY",
  reduce_function = "SUM",
  mean_divider = "ONE",
  variance = FALSE,
  na.rm = NA,
  mp = TRUE,
  debug_use_r_implementation = FALSE,
  ...
)

Arguments

data

[matrix-type] Grid to compute onto.

kernel

[matrix] Computation kernel (neighborhood).

edge_value

[numeric] Numeric value, NA or NaN.The value to give at the edge of the kernel matrix when the moving window overflows the data grid.

transform_function

[character] The function to apply to the cell values covered by the kernel. For possible values, see pfocal_info_transform(). Default to "MULTIPLY".

reduce_function

[character] The function to apply to the kernel values after the function passed in transform_function has been applied (the function that summarize the data). For possible values, see pfocal_info_reduce(). Default to "SUM".

mean_divider

[character] Optional, allows to specify how the final value at each cell is divided by a value that can be function of the intermediate data resulting of applying transform_function. For possible values, see pfocal_info_mean_divisor(). Default to "ONE" (for no division).

variance

[logical] Whether to return the "variance" of the intermediate values at each point (for more details please see pfocal_info_variance()). Default to FALSE (just returns the value at each point).

na.rm

[NA OR logical] The behavior to adopt for dealing with missing values, default to NA (faster option). For possible values see pfocal_info_nan_policy().

mp

[logical] Whether to use the open_mp implementation, default to TRUE.

debug_use_r_implementation

[logical] Used for debugging purposes whether to use the slow R implementation instead of the fast C++ code. Default to FALSE.

...

None used at the moment .

Value

The updated, convoluted grid.

Details

Note that the memory allocation for the output is of size sizeof(double) * ncol * nrow and for the intermediate values, sizeof(double) * (ncol + kernel_ncol) * (nrow + kernel_nrow/2).

Examples

data <- matrix(nrow = 10, ncol = 10, data = runif(10 * 10))
kernel <- matrix(1 / 9, nrow = 3, ncol = 3)
pfocal(data = data, kernel = kernel)
#>            [,1]      [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
#>  [1,] 0.1062627 0.1725753 0.2110088 0.2970513 0.2860302 0.2721521 0.1743829
#>  [2,] 0.2488431 0.3391178 0.2950473 0.3379222 0.3712159 0.4506350 0.3743086
#>  [3,] 0.3207893 0.4100867 0.2461755 0.2296780 0.3181074 0.4965297 0.5426139
#>  [4,] 0.4819781 0.6445239 0.3956522 0.3604478 0.4235078 0.6420612 0.7092524
#>  [5,] 0.4638728 0.6857039 0.4852314 0.4849188 0.5145584 0.5951594 0.6472768
#>  [6,] 0.4786395 0.6804435 0.6054782 0.5484440 0.5669300 0.5535241 0.5227003
#>  [7,] 0.3625062 0.4959709 0.5194919 0.5086416 0.5491981 0.4949362 0.4239478
#>  [8,] 0.3524099 0.4630270 0.5645277 0.5579906 0.6039318 0.5959396 0.5376112
#>  [9,] 0.3327112 0.4331481 0.5353715 0.6050064 0.6287885 0.5998641 0.5710144
#> [10,] 0.2534251 0.3240274 0.3949877 0.4194637 0.4470595 0.4183162 0.4227777
#>            [,8]      [,9]     [,10]
#>  [1,] 0.2483783 0.2741507 0.2328561
#>  [2,] 0.3819133 0.3292371 0.2505575
#>  [3,] 0.5033639 0.4565013 0.3069434
#>  [4,] 0.5834784 0.4452955 0.2204873
#>  [5,] 0.5131422 0.4834106 0.2673433
#>  [6,] 0.3932662 0.3619006 0.2236086
#>  [7,] 0.3354868 0.3043464 0.2090381
#>  [8,] 0.5184546 0.4021553 0.2328641
#>  [9,] 0.5689197 0.5301223 0.2743323
#> [10,] 0.4433176 0.4203965 0.2251052