Fast methods for common kernel computations

Methods wrapping pfocal to implement common kernel computations with default argument values.

pfocal_fast_gaussian_radius(
  data,
  vertical_radius,
  vertical_sd = 1,
  horizontal_radius = vertical_radius,
  horizontal_sd = vertical_sd,
  tail_included = TRUE,
  na.rm = NA,
  mp = TRUE,
  debug_use_r_implementation = FALSE,
  ...,
  transform_function = "MULTIPLY",
  reduce_function = "SUM",
  mean_divider = "ONE",
  variance = FALSE
)

pfocal_fast_gaussian_confidence(
  data,
  vertical_r0 = 0.05,
  vertical_sd = 1,
  horizontal_r0 = vertical_r0,
  horizontal_sd = vertical_sd,
  tail_included = TRUE,
  na.rm = NA,
  mp = TRUE,
  debug_use_r_implementation = FALSE,
  ...,
  transform_function = "MULTIPLY",
  reduce_function = "SUM",
  mean_divider = "ONE",
  variance = FALSE
)

pfocal_fast_binomial(
  data,
  vertical_radius,
  horizontal_radius = vertical_radius,
  na.rm = NA,
  mp = TRUE,
  debug_use_r_implementation = FALSE,
  ...,
  transform_function = "MULTIPLY",
  reduce_function = "SUM",
  mean_divider = "ONE",
  variance = FALSE
)

pfocal_fast_abs_rectangle(
  data,
  height,
  width = height,
  value = 1,
  na.rm = NA,
  mp = TRUE,
  debug_use_r_implementation = FALSE,
  ...,
  transform_function = "MULTIPLY",
  reduce_function = "SUM",
  mean_divider = "ONE",
  variance = FALSE
)

pfocal_fast_separated(
  data,
  kernel_list,
  na.rm = NA,
  mp = TRUE,
  debug_use_r_implementation = FALSE,
  ...,
  transform_function = "MULTIPLY",
  reduce_function = "SUM",
  mean_divider = "ONE",
  variance = FALSE
)

Arguments

data

[matrix-type] Grid to compute onto.

vertical_radius

[numeric] The kernel's radius in the vertical dimension.

vertical_sd

[numeric] The kernel's standard deviation in the vertical dimension.

horizontal_radius

[numeric] The kernel's radius in the horizontal dimension.

horizontal_sd

[numeric] The kernel's standard deviation in the horizontal dimension.

tail_included

[logical] Whether or not to include the kernel tail.

na.rm

[NA OR logical] The behavior to adopt for dealing with missing values, default to NA (faster option). For possible values see pfocal_info_nan_policy().

mp

[logical] Whether to use the open_mp implementation, default to TRUE.

debug_use_r_implementation

[logical] Used for debugging purposes whether to use the slow R implementation instead of the fass C++ code. Default to FALSE.

...

None used at the moment .

transform_function

[character] The function to apply to the cell values covered by the kernel. For possible values, see pfocal_info_transform(). Default to "MULTIPLY".

reduce_function

[character] The function to apply to the kernel values after the function passed in transform_function has been applied (the function that summarize the data). For possible values, see pfocal_info_reduce(). Default to "SUM".

mean_divider

[character] Optional, allows to specify how the final value at each cell is divided by a value that can be function of the intermediate data resulting of applying transform_function. For possible values, see pfocal_info_mean_divisor(). Default to "ONE" (for no division).

variance

[logical] Whether to return the "variance" of the intermediate values at each point (for more details please see pfocal_info_variance()). Default to FALSE (just returns the value at each point).

vertical_r0

[numeric] The kernel's r0 (exponential) in the vertical dimension.

horizontal_r0

[numeric] The kernel's r0 (exponential) in the horizontal dimension.

height

[numeric] For rectangular kernels, the height of the rectangle.

width

[numeric] For rectangular kernels, the width of the rectangle.

value

[numeric] For single value matrices, the value.

kernel_list

[list] A list of kernels computed from functions in kernel-gaussian, kernel-binomial, kernel-circular, kernel-distance, kernel-exponential.

Value

The updated, convoluted grid.

Examples

data <- matrix(nrow = 10, ncol = 10, data = runif(10 * 10))

pfocal_fast_gaussian_radius(data, vertical_radius = 2)
#>            [,1]      [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
#>  [1,] 0.1468319 0.2375491 0.2856298 0.2740692 0.2365602 0.2193475 0.2262842
#>  [2,] 0.1919972 0.3193363 0.3832555 0.3717418 0.3291658 0.2960351 0.2902011
#>  [3,] 0.2251195 0.3716962 0.4389056 0.4237920 0.3838619 0.3408731 0.3069767
#>  [4,] 0.2688753 0.4380226 0.5130873 0.5067973 0.4686673 0.4095699 0.3616069
#>  [5,] 0.2995788 0.4582188 0.5170113 0.5159838 0.4993059 0.4493516 0.3827250
#>  [6,] 0.3111043 0.4540954 0.5092588 0.5288997 0.5376335 0.4935817 0.4320169
#>  [7,] 0.3525895 0.4639631 0.4993887 0.5281313 0.5468682 0.5124200 0.4510040
#>  [8,] 0.3354378 0.4167945 0.4410717 0.4811798 0.5131739 0.4896240 0.4450528
#>  [9,] 0.3073256 0.3508543 0.3534707 0.3865441 0.4190688 0.4079191 0.3711178
#> [10,] 0.2541948 0.2775893 0.2700386 0.2957508 0.3180025 0.3078594 0.2864927
#>            [,8]      [,9]     [,10]
#>  [1,] 0.2582285 0.2709717 0.2078243
#>  [2,] 0.3296516 0.3502452 0.2718252
#>  [3,] 0.3262002 0.3492158 0.2777882
#>  [4,] 0.3741637 0.3841296 0.2949808
#>  [5,] 0.3726478 0.3815517 0.3039787
#>  [6,] 0.4271734 0.4232505 0.3200876
#>  [7,] 0.4438221 0.4528511 0.3606582
#>  [8,] 0.4528756 0.4654471 0.3690950
#>  [9,] 0.3803756 0.4093094 0.3377095
#> [10,] 0.3093405 0.3466269 0.2908990
pfocal_fast_gaussian_confidence(data)
#>            [,1]      [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
#>  [1,] 0.1588881 0.2598217 0.3414980 0.3532138 0.3106297 0.2921300 0.3239027
#>  [2,] 0.2638220 0.3999961 0.4593624 0.4411827 0.3803488 0.3410276 0.3366807
#>  [3,] 0.2652902 0.4264075 0.4713933 0.4234954 0.3609827 0.3149419 0.2822213
#>  [4,] 0.2345550 0.4171263 0.4855384 0.4493957 0.4086407 0.3636132 0.3153192
#>  [5,] 0.2476948 0.4335367 0.5291145 0.5400095 0.5288408 0.4626534 0.3813067
#>  [6,] 0.3152343 0.4760093 0.5541898 0.5931571 0.5965588 0.5149859 0.4092227
#>  [7,] 0.3846471 0.5072722 0.5414510 0.5805840 0.5986632 0.5511151 0.4785495
#>  [8,] 0.3953579 0.4687683 0.4729145 0.5124241 0.5662758 0.5809280 0.5628128
#>  [9,] 0.3766348 0.4029282 0.3908474 0.4344220 0.4965685 0.5144826 0.5019926
#> [10,] 0.2931179 0.2955948 0.2809518 0.3122647 0.3354634 0.3166022 0.2839015
#>            [,8]      [,9]     [,10]
#>  [1,] 0.3759850 0.3720633 0.2724934
#>  [2,] 0.3758610 0.4023346 0.3215175
#>  [3,] 0.3161717 0.3676585 0.2997981
#>  [4,] 0.3409661 0.3812400 0.3057122
#>  [5,] 0.3721187 0.3762148 0.2971644
#>  [6,] 0.3571410 0.3249802 0.2448061
#>  [7,] 0.4401227 0.4034053 0.2940161
#>  [8,] 0.5804389 0.5824140 0.4460248
#>  [9,] 0.5508468 0.6077997 0.5025258
#> [10,] 0.3239975 0.4093239 0.3736288
pfocal_fast_binomial(data, vertical_radius = 2)
#>            [,1]      [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
#>  [1,]  40.80705  67.48401  87.45806  90.16751  79.53095  73.93789  82.24133
#>  [2,]  66.67960 102.63353 118.39382 112.94928  97.48932  86.33180  86.37683
#>  [3,]  68.89288 110.34333 122.35727 109.62063  93.29669  80.37988  73.35846
#>  [4,]  60.35718 106.75330 124.83624 116.40404 104.80216  91.94551  80.80195
#>  [5,]  63.36697 110.65231 135.46241 139.18608 134.40532 117.12036  96.96463
#>  [6,]  80.78023 122.60302 143.56843 153.79716 153.21200 132.49830 105.64450
#>  [7,]  97.23803 129.61878 139.48602 148.77241 153.73996 142.09006 123.82418
#>  [8,] 101.37497 121.97665 122.55410 131.96753 146.17846 149.06767 145.11367
#>  [9,]  95.66182 105.11689 100.85360 111.41061 127.42902 131.18109 129.07787
#> [10,]  72.98834  75.98342  71.26694  78.76149  85.84427  80.78599  73.31274
#>            [,8]      [,9]     [,10]
#>  [1,]  95.38519  94.87043  69.14252
#>  [2,]  97.54598 104.27482  82.19385
#>  [3,]  82.72499  95.24934  77.83676
#>  [4,]  87.43165  97.74810  78.44560
#>  [5,]  93.48463  94.93635  74.30848
#>  [6,]  91.88675  83.90613  62.42491
#>  [7,] 113.87874 104.20516  75.34609
#>  [8,] 149.59249 149.65328 113.74436
#>  [9,] 142.34224 157.27993 129.12894
#> [10,]  84.54292 105.95487  95.44459
pfocal_fast_abs_rectangle(data, height = 2)
#>             [,1]       [,2]      [,3]     [,4]      [,5]      [,6]      [,7]
#>  [1,] 0.07777468 0.08730181 0.7946454 1.331379 0.9945366 0.7974241 0.8467682
#>  [2,] 0.78155405 1.51309428 1.8477772 2.440795 1.9657768 1.6368422 1.6910914
#>  [3,] 0.78557109 2.08416082 2.1330607 1.966895 1.6605333 1.3297070 1.1120170
#>  [4,] 0.23035080 1.37469067 2.1526650 1.638839 1.0172599 1.1103460 0.9455022
#>  [5,] 0.20792233 1.34296010 2.3783730 2.100281 1.8006988 2.1059034 1.4864822
#>  [6,] 0.46628395 1.31482877 2.2751575 2.342549 2.5854867 2.7441590 1.5998584
#>  [7,] 1.07727840 2.32742057 2.2365350 2.278925 2.8229843 2.4405984 1.6547449
#>  [8,] 1.27587506 2.55721844 2.0425504 2.175056 2.4281175 2.3990248 2.2816621
#>  [9,] 1.45584187 1.95174503 1.4193837 1.575840 1.7903559 2.7491495 2.6265730
#> [10,] 1.83737365 2.21969255 1.2012125 1.642578 2.0084695 2.2756594 1.6818531
#>            [,8]      [,9]     [,10]
#>  [1,] 1.4796464 1.8354665 1.3099890
#>  [2,] 2.0405366 2.2346394 2.1950562
#>  [3,] 0.7615571 1.2554029 1.7809790
#>  [4,] 0.5281074 1.4169252 1.6966287
#>  [5,] 1.4313826 1.9707974 2.0569811
#>  [6,] 1.4142575 1.4744275 1.5401681
#>  [7,] 1.1753260 0.8120331 0.7223288
#>  [8,] 2.0704725 2.3474432 2.1656436
#>  [9,] 2.6657712 3.2999185 3.4284019
#> [10,] 1.5304874 2.2712273 3.2070486

pfocal_fast_separated(data, 
                      kernel_list = list(binomial_kernel(vertical_radius = 2, 
                                                         horizontal_radius = 2), 
                                         distance_kernel(vertical_radius = 2,
                                                         horizontal_radius = 2)))
#>           [,1]     [,2]     [,3]     [,4]     [,5]     [,6]     [,7]     [,8]
#>  [1,] 1449.562 1883.918 2384.005 2531.531 2505.716 2427.365 2360.423 2197.248
#>  [2,] 1882.280 2444.254 3135.324 3385.740 3356.645 3244.096 3160.307 2924.507
#>  [3,] 2559.279 3392.752 4401.214 4803.148 4783.230 4622.729 4475.085 4099.002
#>  [4,] 2917.228 3867.443 5030.797 5434.113 5307.102 5021.921 4756.024 4244.222
#>  [5,] 3056.683 4057.676 5319.456 5717.442 5563.423 5280.043 5004.179 4434.139
#>  [6,] 3059.588 4106.861 5498.739 5947.920 5890.153 5758.055 5599.244 5052.582
#>  [7,] 3000.637 4076.149 5579.097 6026.343 6021.967 6004.756 5942.610 5431.620
#>  [8,] 2788.603 3796.251 5254.178 5589.463 5543.741 5555.757 5547.955 5099.682
#>  [9,] 2073.381 2818.553 3954.557 4192.189 4181.991 4273.367 4357.514 4045.877
#> [10,] 1510.486 2051.866 2891.458 3064.748 3111.204 3292.967 3496.244 3362.762
#>           [,9]    [,10]
#>  [1,] 1637.635 1336.060
#>  [2,] 2145.242 1736.239
#>  [3,] 3001.478 2366.192
#>  [4,] 3029.327 2333.556
#>  [5,] 3151.871 2389.081
#>  [6,] 3656.881 2770.633
#>  [7,] 3947.791 3016.265
#>  [8,] 3665.483 2835.336
#>  [9,] 2882.364 2241.980
#> [10,] 2455.027 1930.119