Fast, parallel implementation of grid data convolution

A fast, parallel implementation of convolutions for grid-type data (matrices, rasters and other grid based objects).

pfocal(
  data,
  kernel,
  edge_value = 0,
  transform_function = "MULTIPLY",
  reduce_function = "SUM",
  mean_divider = "ONE",
  variance = FALSE,
  na.rm = NA,
  mp = TRUE,
  debug_use_r_implementation = FALSE,
  ...
)

Arguments

data

[matrix-type] Grid to compute onto.

kernel

[matrix] Computation kernel (neighborhood).

edge_value

[numeric] Numeric value, NA or NaN.The value to give at the edge of the kernel matrix when the moving window overflows the data grid.

transform_function

[character] The function to apply to the cell values covered by the kernel. For possible values, see pfocal_info_transform(). Default to "MULTIPLY".

reduce_function

[character] The function to apply to the kernel values after the function passed in transform_function has been applied (the function that summarize the data). For possible values, see pfocal_info_reduce(). Default to "SUM".

mean_divider

[character] Optional, allows to specify how the final value at each cell is divided by a value that can be function of the intermediate data resulting of applying transform_function. For possible values, see pfocal_info_mean_divisor(). Default to "ONE" (for no division).

variance

[logical] Whether to return the "variance" of the intermediate values at each point (for more details please see pfocal_info_variance()). Default to FALSE (just returns the value at each point).

na.rm

[NA OR logical] The behavior to adopt for dealing with missing values, default to NA (faster option). For possible values see pfocal_info_nan_policy().

mp

[logical] Whether to use the open_mp implementation, default to TRUE.

debug_use_r_implementation

[logical] Used for debugging purposes whether to use the slow R implementation instead of the fass C++ code. Default to FALSE.

...

None used at the moment .

Value

The updated, convoluted grid.

Details

Note that the memory allocation for the output is of size sizeof(double) * ncol * nrow and for the intermediate values, sizeof(double) * (ncol + kernel_ncol) * (nrow + kernel_nrow/2).

Examples

data <- matrix(nrow = 10, ncol = 10, data = runif(10 * 10))
kernel <- matrix(1 / 9, nrow = 3, ncol = 3)
pfocal(data = data, kernel = kernel)
#>            [,1]      [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
#>  [1,] 0.1948394 0.2865527 0.3706410 0.3699840 0.3558263 0.3685958 0.3689619
#>  [2,] 0.3573151 0.4782142 0.5832891 0.5753912 0.6150051 0.6067652 0.5580258
#>  [3,] 0.5141320 0.6132048 0.6139798 0.5214981 0.6269414 0.6170730 0.6020319
#>  [4,] 0.5140002 0.6516766 0.6299383 0.5224553 0.5659367 0.6117839 0.4976528
#>  [5,] 0.4425154 0.5828819 0.5003813 0.4395601 0.4077494 0.4810120 0.4149793
#>  [6,] 0.3182816 0.4416853 0.4330330 0.4134871 0.4267701 0.4829421 0.4688869
#>  [7,] 0.3666398 0.4776358 0.4518536 0.4747674 0.5101742 0.4527671 0.4170171
#>  [8,] 0.4416207 0.5366382 0.5301366 0.4738220 0.5346747 0.4231089 0.3726607
#>  [9,] 0.4691752 0.6361447 0.5473028 0.4324514 0.4146132 0.3848340 0.3745357
#> [10,] 0.2554898 0.3484111 0.3107547 0.2354886 0.2612759 0.2622723 0.2952310
#>            [,8]      [,9]     [,10]
#>  [1,] 0.4612462 0.3894858 0.2909095
#>  [2,] 0.5716402 0.4720183 0.3344693
#>  [3,] 0.5193736 0.4321488 0.2697632
#>  [4,] 0.4117480 0.3279598 0.2493183
#>  [5,] 0.3978819 0.3898426 0.2719312
#>  [6,] 0.4479001 0.4417122 0.2688617
#>  [7,] 0.3603016 0.4252044 0.2352046
#>  [8,] 0.3420778 0.3970538 0.2546919
#>  [9,] 0.3738343 0.4398021 0.3162151
#> [10,] 0.3064275 0.3786362 0.2786425