Fast, parallel implementation of grid data convolution

A fast, parallel implementation of convolutions for grid-type data (matrices, rasters and other grid based objects).

pfocal(
  data,
  kernel,
  edge_value = 0,
  transform_function = "MULTIPLY",
  reduce_function = "SUM",
  mean_divider = "ONE",
  variance = FALSE,
  na.rm = NA,
  mp = TRUE,
  debug_use_r_implementation = FALSE,
  ...
)

Arguments

data

[matrix-type] Grid to compute onto.

kernel

[matrix] Computation kernel (neighborhood).

edge_value

[numeric] Numeric value, NA or NaN.The value to give at the edge of the kernel matrix when the moving window overflows the data grid.

transform_function

[character] The function to apply to the cell values covered by the kernel. For possible values, see pfocal_info_transform(). Default to "MULTIPLY".

reduce_function

[character] The function to apply to the kernel values after the function passed in transform_function has been applied (the function that summarize the data). For possible values, see pfocal_info_reduce(). Default to "SUM".

mean_divider

[character] Optional, allows to specify how the final value at each cell is divided by a value that can be function of the intermediate data resulting of applying transform_function. For possible values, see pfocal_info_mean_divisor(). Default to "ONE" (for no division).

variance

[logical] Whether to return the "variance" of the intermediate values at each point (for more details please see pfocal_info_variance()). Default to FALSE (just returns the value at each point).

na.rm

[NA OR logical] The behavior to adopt for dealing with missing values, default to NA (faster option). For possible values see pfocal_info_nan_policy().

mp

[logical] Whether to use the open_mp implementation, default to TRUE.

debug_use_r_implementation

[logical] Used for debugging purposes whether to use the slow R implementation instead of the fast C++ code. Default to FALSE.

...

None used at the moment .

Value

The updated, convoluted grid.

Details

Note that the memory allocation for the output is of size sizeof(double) * ncol * nrow and for the intermediate values, sizeof(double) * (ncol + kernel_ncol) * (nrow + kernel_nrow/2).

Examples

data <- matrix(nrow = 10, ncol = 10, data = runif(10 * 10))
kernel <- matrix(1 / 9, nrow = 3, ncol = 3)
pfocal(data = data, kernel = kernel)
#>            [,1]      [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
#>  [1,] 0.1278350 0.2148824 0.2757514 0.3460528 0.3757782 0.3887825 0.3251569
#>  [2,] 0.2462112 0.4066415 0.3702850 0.4903852 0.5214922 0.6284631 0.5835202
#>  [3,] 0.2655487 0.4523665 0.3434845 0.3784921 0.3821423 0.5324814 0.5625475
#>  [4,] 0.2123278 0.4594822 0.3763402 0.4058706 0.3801588 0.6234789 0.7635371
#>  [5,] 0.1957089 0.4618305 0.5003889 0.5009420 0.4453545 0.5395191 0.6197214
#>  [6,] 0.3239715 0.4967121 0.4773295 0.4666097 0.4773539 0.4648077 0.5499887
#>  [7,] 0.3728782 0.4783573 0.4634961 0.4497486 0.5431471 0.4834718 0.4531069
#>  [8,] 0.3594847 0.4375457 0.4005234 0.4057043 0.5568193 0.5474371 0.5734818
#>  [9,] 0.3034348 0.4608076 0.4850337 0.4414681 0.5662080 0.6007939 0.5789432
#> [10,] 0.1967163 0.3530603 0.3588763 0.3562770 0.4044385 0.4081574 0.4179173
#>            [,8]      [,9]     [,10]
#>  [1,] 0.2912977 0.2896579 0.2156103
#>  [2,] 0.4910165 0.4247974 0.2633825
#>  [3,] 0.4542495 0.3708373 0.2249578
#>  [4,] 0.6412071 0.4745720 0.2271170
#>  [5,] 0.5963068 0.5051357 0.3300885
#>  [6,] 0.5811209 0.6488053 0.4259221
#>  [7,] 0.4509633 0.5144708 0.3723061
#>  [8,] 0.5202173 0.5195521 0.3149975
#>  [9,] 0.4352290 0.3177897 0.1877887
#> [10,] 0.3357618 0.2591684 0.1508972