Fast, parallel implementation of grid data convolution

A fast, parallel implementation of convolutions for grid-type data (matrices, rasters and other grid based objects).

pfocal(
  data,
  kernel,
  edge_value = 0,
  transform_function = "MULTIPLY",
  reduce_function = "SUM",
  mean_divider = "ONE",
  variance = FALSE,
  na.rm = NA,
  mp = TRUE,
  debug_use_r_implementation = FALSE,
  ...
)

Arguments

data

[matrix-type] Grid to compute onto.

kernel

[matrix] Computation kernel (neighborhood).

edge_value

[numeric] Numeric value, NA or NaN.The value to give at the edge of the kernel matrix when the moving window overflows the data grid.

transform_function

[character] The function to apply to the cell values covered by the kernel. For possible values, see pfocal_info_transform(). Default to "MULTIPLY".

reduce_function

[character] The function to apply to the kernel values after the function passed in transform_function has been applied (the function that summarize the data). For possible values, see pfocal_info_reduce(). Default to "SUM".

mean_divider

[character] Optional, allows to specify how the final value at each cell is divided by a value that can be function of the intermediate data resulting of applying transform_function. For possible values, see pfocal_info_mean_divisor(). Default to "ONE" (for no division).

variance

[logical] Whether to return the "variance" of the intermediate values at each point (for more details please see pfocal_info_variance()). Default to FALSE (just returns the value at each point).

na.rm

[NA OR logical] The behavior to adopt for dealing with missing values, default to NA (faster option). For possible values see pfocal_info_nan_policy().

mp

[logical] Whether to use the open_mp implementation, default to TRUE.

debug_use_r_implementation

[logical] Used for debugging purposes whether to use the slow R implementation instead of the fast C++ code. Default to FALSE.

...

None used at the moment .

Value

The updated, convoluted grid.

Details

Note that the memory allocation for the output is of size sizeof(double) * ncol * nrow and for the intermediate values, sizeof(double) * (ncol + kernel_ncol) * (nrow + kernel_nrow/2).

Examples

data <- matrix(nrow = 10, ncol = 10, data = runif(10 * 10))
kernel <- matrix(1 / 9, nrow = 3, ncol = 3)
pfocal(data = data, kernel = kernel)
#>            [,1]      [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
#>  [1,] 0.2533098 0.3084419 0.3185849 0.2401917 0.3411557 0.2806674 0.3648612
#>  [2,] 0.3164349 0.3805883 0.4082054 0.3271450 0.4362716 0.3505900 0.4640531
#>  [3,] 0.3175922 0.4292610 0.5222930 0.3948508 0.4605427 0.3966614 0.4671038
#>  [4,] 0.3206128 0.3970048 0.4726342 0.3686938 0.4459701 0.4614853 0.5265974
#>  [5,] 0.3331641 0.4115058 0.4830221 0.3881504 0.4697357 0.4429280 0.5074521
#>  [6,] 0.3617782 0.4360711 0.4556576 0.4388206 0.5320050 0.4890340 0.5066934
#>  [7,] 0.3903083 0.4747682 0.4230336 0.4291086 0.4770323 0.4179326 0.4340502
#>  [8,] 0.4901454 0.6627977 0.5795527 0.5727733 0.5785116 0.5580233 0.4502386
#>  [9,] 0.4727276 0.6642159 0.5586605 0.5218522 0.5288282 0.5333116 0.4732406
#> [10,] 0.3191731 0.4909455 0.4371620 0.4321864 0.4331412 0.4004498 0.2888696
#>            [,8]      [,9]     [,10]
#>  [1,] 0.4205809 0.4806399 0.3384026
#>  [2,] 0.5058110 0.6162341 0.4070726
#>  [3,] 0.4697774 0.5813912 0.4147547
#>  [4,] 0.5529982 0.5383365 0.3492440
#>  [5,] 0.6320822 0.5832406 0.4087580
#>  [6,] 0.6581145 0.5811060 0.3831643
#>  [7,] 0.5579984 0.5332295 0.3711589
#>  [8,] 0.4981883 0.5471225 0.4291281
#>  [9,] 0.4363904 0.5120504 0.3946519
#> [10,] 0.2358503 0.3370835 0.2907867