Fast, parallel implementation of grid data convolution

A fast, parallel implementation of convolutions for grid-type data (matrices, rasters and other grid based objects).

pfocal(
  data,
  kernel,
  edge_value = 0,
  transform_function = "MULTIPLY",
  reduce_function = "SUM",
  mean_divider = "ONE",
  variance = FALSE,
  na.rm = NA,
  mp = TRUE,
  debug_use_r_implementation = FALSE,
  ...
)

Arguments

data

[matrix-type] Grid to compute onto.

kernel

[matrix] Computation kernel (neighborhood).

edge_value

[numeric] Numeric value, NA or NaN.The value to give at the edge of the kernel matrix when the moving window overflows the data grid.

transform_function

[character] The function to apply to the cell values covered by the kernel. For possible values, see pfocal_info_transform(). Default to "MULTIPLY".

reduce_function

[character] The function to apply to the kernel values after the function passed in transform_function has been applied (the function that summarize the data). For possible values, see pfocal_info_reduce(). Default to "SUM".

mean_divider

[character] Optional, allows to specify how the final value at each cell is divided by a value that can be function of the intermediate data resulting of applying transform_function. For possible values, see pfocal_info_mean_divisor(). Default to "ONE" (for no division).

variance

[logical] Whether to return the "variance" of the intermediate values at each point (for more details please see pfocal_info_variance()). Default to FALSE (just returns the value at each point).

na.rm

[NA OR logical] The behavior to adopt for dealing with missing values, default to NA (faster option). For possible values see pfocal_info_nan_policy().

mp

[logical] Whether to use the open_mp implementation, default to TRUE.

debug_use_r_implementation

[logical] Used for debugging purposes whether to use the slow R implementation instead of the fast C++ code. Default to FALSE.

...

None used at the moment .

Value

The updated, convoluted grid.

Details

Note that the memory allocation for the output is of size sizeof(double) * ncol * nrow and for the intermediate values, sizeof(double) * (ncol + kernel_ncol) * (nrow + kernel_nrow/2).

Examples

data <- matrix(nrow = 10, ncol = 10, data = runif(10 * 10))
kernel <- matrix(1 / 9, nrow = 3, ncol = 3)
pfocal(data = data, kernel = kernel)
#>            [,1]      [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
#>  [1,] 0.2765424 0.3599251 0.2941120 0.3286113 0.4426338 0.3980712 0.3087253
#>  [2,] 0.3168103 0.4426898 0.3732492 0.4580694 0.5501923 0.5243751 0.4000301
#>  [3,] 0.2845465 0.4549509 0.4860392 0.5427751 0.5423978 0.5865642 0.5128843
#>  [4,] 0.2900089 0.5156576 0.5493873 0.6654795 0.5107766 0.5148680 0.4851258
#>  [5,] 0.3992647 0.6500310 0.5954870 0.6277452 0.5190585 0.5690022 0.5602769
#>  [6,] 0.3712093 0.5703237 0.4663944 0.4663053 0.4049012 0.4246638 0.4778152
#>  [7,] 0.3029970 0.4131019 0.3393517 0.2584605 0.3206444 0.4545736 0.5520460
#>  [8,] 0.2877789 0.4331028 0.4229989 0.4339124 0.4353855 0.5094636 0.5345806
#>  [9,] 0.2744883 0.4360350 0.4609578 0.5819933 0.5760168 0.5818270 0.5017845
#> [10,] 0.2202117 0.3672462 0.3718712 0.5140636 0.4769126 0.4195257 0.2836512
#>            [,8]      [,9]     [,10]
#>  [1,] 0.1692359 0.3340057 0.2700495
#>  [2,] 0.3270452 0.5490202 0.4381293
#>  [3,] 0.4705261 0.4997051 0.3259200
#>  [4,] 0.5080391 0.4901928 0.2452843
#>  [5,] 0.4712634 0.3666382 0.1667384
#>  [6,] 0.4861413 0.4868235 0.2945980
#>  [7,] 0.6179425 0.5442263 0.3737573
#>  [8,] 0.6059258 0.4889393 0.3186044
#>  [9,] 0.5736774 0.4847613 0.3483654
#> [10,] 0.3052583 0.2638246 0.2012535