Fast, parallel implementation of grid data convolution

A fast, parallel implementation of convolutions for grid-type data (matrices, rasters and other grid based objects).

pfocal(
  data,
  kernel,
  edge_value = 0,
  transform_function = "MULTIPLY",
  reduce_function = "SUM",
  mean_divider = "ONE",
  variance = FALSE,
  na.rm = NA,
  mp = TRUE,
  debug_use_r_implementation = FALSE,
  ...
)

Arguments

data

[matrix-type] Grid to compute onto.

kernel

[matrix] Computation kernel (neighborhood).

edge_value

[numeric] Numeric value, NA or NaN.The value to give at the edge of the kernel matrix when the moving window overflows the data grid.

transform_function

[character] The function to apply to the cell values covered by the kernel. For possible values, see pfocal_info_transform(). Default to "MULTIPLY".

reduce_function

[character] The function to apply to the kernel values after the function passed in transform_function has been applied (the function that summarize the data). For possible values, see pfocal_info_reduce(). Default to "SUM".

mean_divider

[character] Optional, allows to specify how the final value at each cell is divided by a value that can be function of the intermediate data resulting of applying transform_function. For possible values, see pfocal_info_mean_divisor(). Default to "ONE" (for no division).

variance

[logical] Whether to return the "variance" of the intermediate values at each point (for more details please see pfocal_info_variance()). Default to FALSE (just returns the value at each point).

na.rm

[NA OR logical] The behavior to adopt for dealing with missing values, default to NA (faster option). For possible values see pfocal_info_nan_policy().

mp

[logical] Whether to use the open_mp implementation, default to TRUE.

debug_use_r_implementation

[logical] Used for debugging purposes whether to use the slow R implementation instead of the fass C++ code. Default to FALSE.

...

None used at the moment .

Value

The updated, convoluted grid.

Details

Note that the memory allocation for the output is of size sizeof(double) * ncol * nrow and for the intermediate values, sizeof(double) * (ncol + kernel_ncol) * (nrow + kernel_nrow/2).

Examples

data <- matrix(nrow = 10, ncol = 10, data = runif(10 * 10))
kernel <- matrix(1 / 9, nrow = 3, ncol = 3)
pfocal(data = data, kernel = kernel)
#>            [,1]      [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
#>  [1,] 0.1083793 0.1732469 0.2329018 0.2218910 0.2004005 0.2101597 0.2802599
#>  [2,] 0.1953625 0.2636951 0.3750431 0.3500862 0.3308420 0.2661256 0.3338821
#>  [3,] 0.1968097 0.2931036 0.3332414 0.4015795 0.3638567 0.4278449 0.4387729
#>  [4,] 0.2119934 0.3780375 0.4090606 0.5588351 0.5213036 0.5005088 0.4017272
#>  [5,] 0.2988300 0.5233571 0.4601292 0.5825960 0.5668663 0.5851424 0.5354249
#>  [6,] 0.3592735 0.6323677 0.4742856 0.5341685 0.5097863 0.5266686 0.5202234
#>  [7,] 0.4296024 0.6534912 0.5536382 0.4608408 0.4655697 0.4867697 0.6680716
#>  [8,] 0.3535535 0.5310843 0.5200641 0.4711375 0.5157755 0.5461755 0.6649087
#>  [9,] 0.4285388 0.5930045 0.6696354 0.5752321 0.5696812 0.5262838 0.6091056
#> [10,] 0.3014657 0.4085388 0.4533731 0.4268703 0.4005177 0.3318784 0.3283014
#>            [,8]      [,9]     [,10]
#>  [1,] 0.3682415 0.2963188 0.1740710
#>  [2,] 0.4416556 0.4133839 0.2711415
#>  [3,] 0.5313476 0.4687754 0.3121139
#>  [4,] 0.4463141 0.5295260 0.3996420
#>  [5,] 0.5276686 0.5756678 0.3915351
#>  [6,] 0.4747698 0.5427460 0.3677724
#>  [7,] 0.5846334 0.5712429 0.3049532
#>  [8,] 0.5422201 0.5087687 0.2554762
#>  [9,] 0.5226237 0.5467575 0.2934199
#> [10,] 0.2765769 0.3103035 0.1566351