Fast, parallel implementation of grid data convolution

A fast, parallel implementation of convolutions for grid-type data (matrices, rasters and other grid based objects).

pfocal(
  data,
  kernel,
  edge_value = 0,
  transform_function = "MULTIPLY",
  reduce_function = "SUM",
  mean_divider = "ONE",
  variance = FALSE,
  na.rm = NA,
  mp = TRUE,
  debug_use_r_implementation = FALSE,
  ...
)

Arguments

data

[matrix-type] Grid to compute onto.

kernel

[matrix] Computation kernel (neighborhood).

edge_value

[numeric] Numeric value, NA or NaN.The value to give at the edge of the kernel matrix when the moving window overflows the data grid.

transform_function

[character] The function to apply to the cell values covered by the kernel. For possible values, see pfocal_info_transform(). Default to "MULTIPLY".

reduce_function

[character] The function to apply to the kernel values after the function passed in transform_function has been applied (the function that summarize the data). For possible values, see pfocal_info_reduce(). Default to "SUM".

mean_divider

[character] Optional, allows to specify how the final value at each cell is divided by a value that can be function of the intermediate data resulting of applying transform_function. For possible values, see pfocal_info_mean_divisor(). Default to "ONE" (for no division).

variance

[logical] Whether to return the "variance" of the intermediate values at each point (for more details please see pfocal_info_variance()). Default to FALSE (just returns the value at each point).

na.rm

[NA OR logical] The behavior to adopt for dealing with missing values, default to NA (faster option). For possible values see pfocal_info_nan_policy().

mp

[logical] Whether to use the open_mp implementation, default to TRUE.

debug_use_r_implementation

[logical] Used for debugging purposes whether to use the slow R implementation instead of the fast C++ code. Default to FALSE.

...

None used at the moment .

Value

The updated, convoluted grid.

Details

Note that the memory allocation for the output is of size sizeof(double) * ncol * nrow and for the intermediate values, sizeof(double) * (ncol + kernel_ncol) * (nrow + kernel_nrow/2).

Examples

data <- matrix(nrow = 10, ncol = 10, data = runif(10 * 10))
kernel <- matrix(1 / 9, nrow = 3, ncol = 3)
pfocal(data = data, kernel = kernel)
#>            [,1]      [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
#>  [1,] 0.1958098 0.3206499 0.3487664 0.3235708 0.3011532 0.3132454 0.3215998
#>  [2,] 0.3013453 0.4390012 0.3963681 0.4579114 0.5296564 0.6161093 0.5751568
#>  [3,] 0.4112487 0.5862149 0.4653923 0.5201099 0.5787229 0.6879237 0.5946150
#>  [4,] 0.4163720 0.5418477 0.3148405 0.4216282 0.5144756 0.5921457 0.5612621
#>  [5,] 0.3663208 0.4815634 0.3361933 0.4323761 0.4602090 0.5050664 0.4614014
#>  [6,] 0.3059491 0.3787757 0.2984738 0.4736395 0.5696358 0.5502850 0.4609972
#>  [7,] 0.3133900 0.3728172 0.2908643 0.3670187 0.4900378 0.6048988 0.5768984
#>  [8,] 0.4121678 0.5154702 0.3779665 0.3218317 0.4294886 0.5061533 0.5666186
#>  [9,] 0.4147289 0.5626658 0.5010566 0.3492998 0.4180416 0.4577712 0.6261846
#> [10,] 0.2651875 0.3856510 0.3754708 0.3000541 0.3024794 0.2627354 0.3727062
#>            [,8]      [,9]     [,10]
#>  [1,] 0.3884572 0.3381846 0.2379981
#>  [2,] 0.6329447 0.5875132 0.4354585
#>  [3,] 0.5493775 0.5826875 0.4350864
#>  [4,] 0.5406349 0.6485029 0.4648597
#>  [5,] 0.4205578 0.4525038 0.2817817
#>  [6,] 0.3850954 0.3908708 0.2372258
#>  [7,] 0.4338814 0.3223391 0.1427345
#>  [8,] 0.4618387 0.4420496 0.2644095
#>  [9,] 0.5078552 0.4911761 0.2666693
#> [10,] 0.3260588 0.3477493 0.2002411