Fast, parallel implementation of grid data convolution

A fast, parallel implementation of convolutions for grid-type data (matrices, rasters and other grid based objects).

pfocal(
  data,
  kernel,
  edge_value = 0,
  transform_function = "MULTIPLY",
  reduce_function = "SUM",
  mean_divider = "ONE",
  variance = FALSE,
  na.rm = NA,
  mp = TRUE,
  debug_use_r_implementation = FALSE,
  ...
)

Arguments

data

[matrix-type] Grid to compute onto.

kernel

[matrix] Computation kernel (neighborhood).

edge_value

[numeric] Numeric value, NA or NaN.The value to give at the edge of the kernel matrix when the moving window overflows the data grid.

transform_function

[character] The function to apply to the cell values covered by the kernel. For possible values, see pfocal_info_transform(). Default to "MULTIPLY".

reduce_function

[character] The function to apply to the kernel values after the function passed in transform_function has been applied (the function that summarize the data). For possible values, see pfocal_info_reduce(). Default to "SUM".

mean_divider

[character] Optional, allows to specify how the final value at each cell is divided by a value that can be function of the intermediate data resulting of applying transform_function. For possible values, see pfocal_info_mean_divisor(). Default to "ONE" (for no division).

variance

[logical] Whether to return the "variance" of the intermediate values at each point (for more details please see pfocal_info_variance()). Default to FALSE (just returns the value at each point).

na.rm

[NA OR logical] The behavior to adopt for dealing with missing values, default to NA (faster option). For possible values see pfocal_info_nan_policy().

mp

[logical] Whether to use the open_mp implementation, default to TRUE.

debug_use_r_implementation

[logical] Used for debugging purposes whether to use the slow R implementation instead of the fast C++ code. Default to FALSE.

...

None used at the moment .

Value

The updated, convoluted grid.

Details

Note that the memory allocation for the output is of size sizeof(double) * ncol * nrow and for the intermediate values, sizeof(double) * (ncol + kernel_ncol) * (nrow + kernel_nrow/2).

Examples

data <- matrix(nrow = 10, ncol = 10, data = runif(10 * 10))
kernel <- matrix(1 / 9, nrow = 3, ncol = 3)
pfocal(data = data, kernel = kernel)
#>            [,1]      [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
#>  [1,] 0.3156019 0.4726093 0.4848703 0.4114565 0.3729299 0.3029049 0.2575937
#>  [2,] 0.4153477 0.6035902 0.6066184 0.4880694 0.5134313 0.4510332 0.4185651
#>  [3,] 0.3302532 0.5368302 0.5244319 0.5114826 0.5786746 0.4770457 0.4569317
#>  [4,] 0.3200975 0.4614275 0.4145992 0.4392676 0.5864822 0.5231739 0.6074905
#>  [5,] 0.3928170 0.5834449 0.5224721 0.5677892 0.6651386 0.5679218 0.6915743
#>  [6,] 0.4763510 0.6760366 0.6470862 0.5709081 0.6341687 0.5257694 0.6732157
#>  [7,] 0.5429191 0.7712220 0.7673488 0.6292543 0.6392028 0.5114247 0.5855194
#>  [8,] 0.4006821 0.6158753 0.6319249 0.5636270 0.5799111 0.5638211 0.6071525
#>  [9,] 0.3513010 0.5101977 0.4912728 0.4746109 0.5875326 0.5997010 0.6358313
#> [10,] 0.1774774 0.2904481 0.2600429 0.2727515 0.3582923 0.4140514 0.4619662
#>            [,8]      [,9]     [,10]
#>  [1,] 0.2043015 0.1927642 0.1629134
#>  [2,] 0.3590714 0.3393155 0.2855144
#>  [3,] 0.3834366 0.4755752 0.3704761
#>  [4,] 0.5712348 0.6196074 0.4054945
#>  [5,] 0.5896486 0.6992165 0.4106926
#>  [6,] 0.6006772 0.5965949 0.3399281
#>  [7,] 0.5135992 0.6010390 0.3953434
#>  [8,] 0.5492205 0.5049165 0.3160749
#>  [9,] 0.5281424 0.6018616 0.3908468
#> [10,] 0.3902954 0.3992881 0.2465761