Fast, parallel implementation of grid data convolution

A fast, parallel implementation of convolutions for grid-type data (matrices, rasters and other grid based objects).

pfocal(
  data,
  kernel,
  edge_value = 0,
  transform_function = "MULTIPLY",
  reduce_function = "SUM",
  mean_divider = "ONE",
  variance = FALSE,
  na.rm = NA,
  mp = TRUE,
  debug_use_r_implementation = FALSE,
  ...
)

Arguments

data

[matrix-type] Grid to compute onto.

kernel

[matrix] Computation kernel (neighborhood).

edge_value

[numeric] Numeric value, NA or NaN.The value to give at the edge of the kernel matrix when the moving window overflows the data grid.

transform_function

[character] The function to apply to the cell values covered by the kernel. For possible values, see pfocal_info_transform(). Default to "MULTIPLY".

reduce_function

[character] The function to apply to the kernel values after the function passed in transform_function has been applied (the function that summarize the data). For possible values, see pfocal_info_reduce(). Default to "SUM".

mean_divider

[character] Optional, allows to specify how the final value at each cell is divided by a value that can be function of the intermediate data resulting of applying transform_function. For possible values, see pfocal_info_mean_divisor(). Default to "ONE" (for no division).

variance

[logical] Whether to return the "variance" of the intermediate values at each point (for more details please see pfocal_info_variance()). Default to FALSE (just returns the value at each point).

na.rm

[NA OR logical] The behavior to adopt for dealing with missing values, default to NA (faster option). For possible values see pfocal_info_nan_policy().

mp

[logical] Whether to use the open_mp implementation, default to TRUE.

debug_use_r_implementation

[logical] Used for debugging purposes whether to use the slow R implementation instead of the fass C++ code. Default to FALSE.

...

None used at the moment .

Value

The updated, convoluted grid.

Details

Note that the memory allocation for the output is of size sizeof(double) * ncol * nrow and for the intermediate values, sizeof(double) * (ncol + kernel_ncol) * (nrow + kernel_nrow/2).

Examples

data <- matrix(nrow = 10, ncol = 10, data = runif(10 * 10))
kernel <- matrix(1 / 9, nrow = 3, ncol = 3)
pfocal(data = data, kernel = kernel)
#>            [,1]      [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
#>  [1,] 0.2294177 0.3374350 0.3995439 0.4038812 0.4694884 0.4206824 0.4174320
#>  [2,] 0.3206794 0.5371461 0.6241917 0.6984007 0.7055374 0.6452846 0.6458447
#>  [3,] 0.2192409 0.4324056 0.5958448 0.6753069 0.6915167 0.6285155 0.6391552
#>  [4,] 0.2799303 0.5283290 0.5892020 0.6196136 0.5319518 0.4639639 0.5930845
#>  [5,] 0.4092764 0.5920840 0.5627313 0.4573556 0.4234460 0.3232175 0.4389913
#>  [6,] 0.5853576 0.7859758 0.6660482 0.4664849 0.3536303 0.2468310 0.2691085
#>  [7,] 0.4266946 0.6262131 0.5417777 0.4389766 0.3530201 0.2745728 0.2643539
#>  [8,] 0.3202015 0.5272392 0.4206099 0.4346575 0.3765947 0.3571164 0.3220867
#>  [9,] 0.2545318 0.4749417 0.3848803 0.4664738 0.3742851 0.4186620 0.3483362
#> [10,] 0.2427187 0.3609021 0.2394244 0.2801881 0.2637389 0.3112004 0.2303207
#>            [,8]      [,9]     [,10]
#>  [1,] 0.3827382 0.3010640 0.1804497
#>  [2,] 0.5859968 0.4726702 0.2462588
#>  [3,] 0.5528161 0.4616197 0.2287857
#>  [4,] 0.5835519 0.5668037 0.2708484
#>  [5,] 0.4903107 0.5785088 0.3539513
#>  [6,] 0.4164221 0.5101941 0.3914373
#>  [7,] 0.3821243 0.5375678 0.4459234
#>  [8,] 0.3524155 0.5105952 0.4263148
#>  [9,] 0.3671389 0.5163164 0.4139597
#> [10,] 0.1955553 0.2843952 0.2350053