Fast methods for common kernel computations

Methods wrapping pfocal to implement common kernel computations with default argument values.

pfocal_fast_gaussian_radius(
  data,
  vertical_radius,
  vertical_sd = 1,
  horizontal_radius = vertical_radius,
  horizontal_sd = vertical_sd,
  tail_included = TRUE,
  na.rm = NA,
  mp = TRUE,
  debug_use_r_implementation = FALSE,
  ...,
  transform_function = "MULTIPLY",
  reduce_function = "SUM",
  mean_divider = "ONE",
  variance = FALSE
)

pfocal_fast_gaussian_confidence(
  data,
  vertical_r0 = 0.05,
  vertical_sd = 1,
  horizontal_r0 = vertical_r0,
  horizontal_sd = vertical_sd,
  tail_included = TRUE,
  na.rm = NA,
  mp = TRUE,
  debug_use_r_implementation = FALSE,
  ...,
  transform_function = "MULTIPLY",
  reduce_function = "SUM",
  mean_divider = "ONE",
  variance = FALSE
)

pfocal_fast_binomial(
  data,
  vertical_radius,
  horizontal_radius = vertical_radius,
  na.rm = NA,
  mp = TRUE,
  debug_use_r_implementation = FALSE,
  ...,
  transform_function = "MULTIPLY",
  reduce_function = "SUM",
  mean_divider = "ONE",
  variance = FALSE
)

pfocal_fast_abs_rectangle(
  data,
  height,
  width = height,
  value = 1,
  na.rm = NA,
  mp = TRUE,
  debug_use_r_implementation = FALSE,
  ...,
  transform_function = "MULTIPLY",
  reduce_function = "SUM",
  mean_divider = "ONE",
  variance = FALSE
)

pfocal_fast_separated(
  data,
  kernel_list,
  na.rm = NA,
  mp = TRUE,
  debug_use_r_implementation = FALSE,
  ...,
  transform_function = "MULTIPLY",
  reduce_function = "SUM",
  mean_divider = "ONE",
  variance = FALSE
)

Arguments

data

[matrix-type] Grid to compute onto.

vertical_radius

[numeric] The kernel's radius in the vertical dimension.

vertical_sd

[numeric] The kernel's standard deviation in the vertical dimension.

horizontal_radius

[numeric] The kernel's radius in the horizontal dimension.

horizontal_sd

[numeric] The kernel's standard deviation in the horizontal dimension.

tail_included

[logical] Whether or not to include the kernel tail.

na.rm

[NA OR logical] The behavior to adopt for dealing with missing values, default to NA (faster option). For possible values see pfocal_info_nan_policy().

mp

[logical] Whether to use the open_mp implementation, default to TRUE.

debug_use_r_implementation

[logical] Used for debugging purposes whether to use the slow R implementation instead of the fast C++ code. Default to FALSE.

...

None used at the moment .

transform_function

[character] The function to apply to the cell values covered by the kernel. For possible values, see pfocal_info_transform(). Default to "MULTIPLY".

reduce_function

[character] The function to apply to the kernel values after the function passed in transform_function has been applied (the function that summarize the data). For possible values, see pfocal_info_reduce(). Default to "SUM".

mean_divider

[character] Optional, allows to specify how the final value at each cell is divided by a value that can be function of the intermediate data resulting of applying transform_function. For possible values, see pfocal_info_mean_divisor(). Default to "ONE" (for no division).

variance

[logical] Whether to return the "variance" of the intermediate values at each point (for more details please see pfocal_info_variance()). Default to FALSE (just returns the value at each point).

vertical_r0

[numeric] The kernel's r0 (exponential) in the vertical dimension.

horizontal_r0

[numeric] The kernel's r0 (exponential) in the horizontal dimension.

height

[numeric] For rectangular kernels, the height of the rectangle.

width

[numeric] For rectangular kernels, the width of the rectangle.

value

[numeric] For single value matrices, the value.

kernel_list

[list] A list of kernels computed from functions in kernel-gaussian, kernel-binomial, kernel-circular, kernel-distance, kernel-exponential.

Value

The updated, convoluted grid.

Examples

data <- matrix(nrow = 10, ncol = 10, data = runif(10 * 10))

pfocal_fast_gaussian_radius(data, vertical_radius = 2)
#>            [,1]      [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
#>  [1,] 0.2962715 0.3576833 0.3343944 0.3325408 0.3430960 0.3294631 0.3152591
#>  [2,] 0.3533463 0.4446554 0.4251419 0.4151132 0.4293940 0.4203846 0.4058580
#>  [3,] 0.3740837 0.5000629 0.4973191 0.4855832 0.4896905 0.4703542 0.4551027
#>  [4,] 0.4114849 0.5626028 0.5698538 0.5436273 0.5314882 0.5082200 0.5049119
#>  [5,] 0.3799130 0.5428707 0.5694140 0.5361821 0.5071526 0.4932952 0.5116653
#>  [6,] 0.3809064 0.5473090 0.5862069 0.5527042 0.5086183 0.4970096 0.5370405
#>  [7,] 0.3638105 0.5328380 0.6010197 0.5869644 0.5228384 0.4865099 0.5265711
#>  [8,] 0.3274997 0.4666586 0.5293547 0.5235379 0.4693877 0.4525003 0.5068100
#>  [9,] 0.2793512 0.3969936 0.4632432 0.4789326 0.4275882 0.3944424 0.4342633
#> [10,] 0.2185767 0.3057707 0.3626159 0.3833261 0.3429854 0.3047616 0.3265931
#>            [,8]      [,9]     [,10]
#>  [1,] 0.3005911 0.2722566 0.2021824
#>  [2,] 0.3893083 0.3535650 0.2653056
#>  [3,] 0.4475291 0.4114811 0.3097588
#>  [4,] 0.5160186 0.4831916 0.3614218
#>  [5,] 0.5281844 0.4813265 0.3508955
#>  [6,] 0.5685993 0.5194591 0.3700243
#>  [7,] 0.5590241 0.5121721 0.3697986
#>  [8,] 0.5306017 0.4710537 0.3306905
#>  [9,] 0.4518222 0.4008120 0.2815974
#> [10,] 0.3380982 0.3052793 0.2187266
pfocal_fast_gaussian_confidence(data)
#>            [,1]      [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
#>  [1,] 0.3596999 0.4127055 0.3803716 0.3885714 0.4091304 0.4084362 0.4065166
#>  [2,] 0.4586921 0.5232806 0.4659509 0.4736267 0.5227285 0.5415728 0.5410097
#>  [3,] 0.4566855 0.5777255 0.5369467 0.5111551 0.5324564 0.5224173 0.5004112
#>  [4,] 0.3894724 0.5777962 0.6037699 0.5668600 0.5387211 0.4737213 0.4224071
#>  [5,] 0.3156231 0.5215800 0.5933195 0.5680220 0.5283610 0.4796133 0.4630742
#>  [6,] 0.3305403 0.5087163 0.5664509 0.5372329 0.4936943 0.4932925 0.5520248
#>  [7,] 0.4085173 0.5552035 0.5818466 0.5278974 0.4626357 0.4858113 0.5920249
#>  [8,] 0.4237087 0.5616782 0.6046209 0.5720157 0.4952329 0.4970842 0.5870366
#>  [9,] 0.3418478 0.4747626 0.5676651 0.6108962 0.5508288 0.4931862 0.5224297
#> [10,] 0.2020100 0.3156112 0.4214435 0.4895406 0.4463694 0.3600471 0.3480543
#>            [,8]      [,9]     [,10]
#>  [1,] 0.3795832 0.3341879 0.2508429
#>  [2,] 0.5037838 0.4185337 0.2852923
#>  [3,] 0.4876830 0.4391522 0.3134638
#>  [4,] 0.4328987 0.4422823 0.3568115
#>  [5,] 0.4843826 0.4802058 0.3888933
#>  [6,] 0.5961307 0.5408954 0.3906086
#>  [7,] 0.6476683 0.5617584 0.3668842
#>  [8,] 0.6172306 0.5295066 0.3434961
#>  [9,] 0.5259719 0.4657361 0.3271065
#> [10,] 0.3474414 0.3302025 0.2645650
pfocal_fast_binomial(data, vertical_radius = 2)
#>            [,1]      [,2]      [,3]      [,4]     [,5]      [,6]      [,7]
#>  [1,]  90.64194 104.75508  96.84562  97.61073 103.6128 105.26057 103.93494
#>  [2,] 117.87278 136.08286 122.16652 121.61056 133.7530 139.88265 139.10198
#>  [3,] 117.37550 148.67724 139.70269 132.39078 136.7955 134.82113 128.60752
#>  [4,]  99.38672 147.16687 155.00713 145.47522 136.6934 121.66523 108.99952
#>  [5,]  81.14406 133.70721 153.28405 147.04690 134.8189 122.38467 117.94110
#>  [6,]  84.73868 130.78648 146.14726 137.58216 125.4421 125.82185 141.16095
#>  [7,] 103.74437 143.11322 149.77344 134.40305 117.8514 124.70860 152.48202
#>  [8,] 108.34423 145.76546 156.42490 146.84194 128.1123 128.81991 150.97587
#>  [9,]  86.96173 123.41241 147.81027 156.47901 141.1488 127.32925 132.67854
#> [10,]  50.87323  80.52665 108.40883 124.76874 113.8051  93.28985  87.71188
#>            [,8]      [,9]    [,10]
#>  [1,]  97.63805  85.45409 62.39456
#>  [2,] 130.18874 109.01932 73.52101
#>  [3,] 124.91092 112.90638 79.75381
#>  [4,] 111.32170 114.12457 90.74518
#>  [5,] 123.67808 124.47509 98.97485
#>  [6,] 152.54187 139.98480 99.44262
#>  [7,] 167.10864 145.17444 93.56156
#>  [8,] 160.56592 137.55827 87.92727
#>  [9,] 135.83245 120.47871 83.67272
#> [10,]  88.95048  84.87053 66.21976
pfocal_fast_abs_rectangle(data, height = 2)
#>            [,1]     [,2]     [,3]     [,4]     [,5]     [,6]     [,7]     [,8]
#>  [1,] 0.9020415 1.672707 1.461137 1.299627 1.361947 1.107932 1.092848 1.160473
#>  [2,] 1.8154369 2.902370 1.783105 1.846757 2.587482 2.403897 2.419085 2.522214
#>  [3,] 1.6206939 2.886041 1.756008 1.242781 1.874063 2.373660 2.500599 2.653527
#>  [4,] 1.0993240 2.916593 3.001435 2.101536 2.216340 2.191952 1.444865 1.444062
#>  [5,] 0.5654067 1.941069 2.943355 2.506245 2.438433 2.242022 1.234453 1.210963
#>  [6,] 0.2077931 1.499542 2.452592 2.273450 2.062542 2.035168 1.980241 2.457064
#>  [7,] 1.0229799 2.234332 2.210794 2.206200 1.951350 1.670994 1.987706 2.952391
#>  [8,] 1.5938337 2.703918 2.710660 2.517375 1.500307 1.248509 2.072928 2.941371
#>  [9,] 1.3698506 2.343519 2.269725 2.807483 2.483104 1.665859 2.357177 2.733054
#> [10,] 0.8754471 1.617931 1.840122 2.835265 3.504406 2.169679 1.848923 2.306088
#>           [,9]    [,10]
#>  [1,] 0.982631 1.468834
#>  [2,] 1.882684 1.757361
#>  [3,] 2.215892 1.382792
#>  [4,] 1.905489 2.055860
#>  [5,] 1.470318 2.118408
#>  [6,] 2.162080 2.154136
#>  [7,] 3.012331 2.081293
#>  [8,] 2.782615 2.035366
#>  [9,] 2.303394 1.920062
#> [10,] 1.895639 2.148136

pfocal_fast_separated(data, 
                      kernel_list = list(binomial_kernel(vertical_radius = 2, 
                                                         horizontal_radius = 2), 
                                         distance_kernel(vertical_radius = 2,
                                                         horizontal_radius = 2)))
#>           [,1]     [,2]     [,3]     [,4]     [,5]     [,6]     [,7]     [,8]
#>  [1,] 1844.284 2376.845 3266.924 3406.008 3322.198 3277.704 3189.186 2879.743
#>  [2,] 2442.054 3154.344 4316.000 4519.486 4355.610 4216.370 4071.928 3683.910
#>  [3,] 3249.960 4213.918 5676.238 6003.360 5864.429 5714.690 5538.187 5039.243
#>  [4,] 3471.058 4467.533 5922.701 6318.107 6275.840 6200.454 6058.891 5545.373
#>  [5,] 3604.353 4605.615 5978.750 6364.530 6357.091 6286.711 6138.023 5654.746
#>  [6,] 3628.451 4675.818 6027.770 6433.838 6480.044 6383.499 6157.122 5671.000
#>  [7,] 3466.190 4557.733 5927.191 6392.903 6558.841 6505.265 6213.783 5691.382
#>  [8,] 3155.391 4203.066 5485.490 5917.566 6142.745 6153.386 5864.275 5349.677
#>  [9,] 2441.585 3214.320 4148.912 4406.112 4580.839 4632.219 4431.831 4035.289
#> [10,] 1874.418 2513.052 3224.299 3402.224 3522.090 3548.344 3369.922 3036.635
#>           [,9]    [,10]
#>  [1,] 2123.266 1564.814
#>  [2,] 2704.396 2041.250
#>  [3,] 3716.780 2800.733
#>  [4,] 4122.740 3109.467
#>  [5,] 4265.869 3261.074
#>  [6,] 4338.333 3358.305
#>  [7,] 4391.660 3406.432
#>  [8,] 4140.293 3206.172
#>  [9,] 3116.389 2403.186
#> [10,] 2315.967 1758.151