Fast methods for common kernel computations

Methods wrapping pfocal to implement common kernel computations with default argument values.

pfocal_fast_gaussian_radius(
  data,
  vertical_radius,
  vertical_sd = 1,
  horizontal_radius = vertical_radius,
  horizontal_sd = vertical_sd,
  tail_included = TRUE,
  na.rm = NA,
  mp = TRUE,
  debug_use_r_implementation = FALSE,
  ...,
  transform_function = "MULTIPLY",
  reduce_function = "SUM",
  mean_divider = "ONE",
  variance = FALSE
)

pfocal_fast_gaussian_confidence(
  data,
  vertical_r0 = 0.05,
  vertical_sd = 1,
  horizontal_r0 = vertical_r0,
  horizontal_sd = vertical_sd,
  tail_included = TRUE,
  na.rm = NA,
  mp = TRUE,
  debug_use_r_implementation = FALSE,
  ...,
  transform_function = "MULTIPLY",
  reduce_function = "SUM",
  mean_divider = "ONE",
  variance = FALSE
)

pfocal_fast_binomial(
  data,
  vertical_radius,
  horizontal_radius = vertical_radius,
  na.rm = NA,
  mp = TRUE,
  debug_use_r_implementation = FALSE,
  ...,
  transform_function = "MULTIPLY",
  reduce_function = "SUM",
  mean_divider = "ONE",
  variance = FALSE
)

pfocal_fast_abs_rectangle(
  data,
  height,
  width = height,
  value = 1,
  na.rm = NA,
  mp = TRUE,
  debug_use_r_implementation = FALSE,
  ...,
  transform_function = "MULTIPLY",
  reduce_function = "SUM",
  mean_divider = "ONE",
  variance = FALSE
)

pfocal_fast_separated(
  data,
  kernel_list,
  na.rm = NA,
  mp = TRUE,
  debug_use_r_implementation = FALSE,
  ...,
  transform_function = "MULTIPLY",
  reduce_function = "SUM",
  mean_divider = "ONE",
  variance = FALSE
)

Arguments

data

[matrix-type] Grid to compute onto.

vertical_radius

[numeric] The kernel's radius in the vertical dimension.

vertical_sd

[numeric] The kernel's standard deviation in the vertical dimension.

horizontal_radius

[numeric] The kernel's radius in the horizontal dimension.

horizontal_sd

[numeric] The kernel's standard deviation in the horizontal dimension.

tail_included

[logical] Whether or not to include the kernel tail.

na.rm

[NA OR logical] The behavior to adopt for dealing with missing values, default to NA (faster option). For possible values see pfocal_info_nan_policy().

mp

[logical] Whether to use the open_mp implementation, default to TRUE.

debug_use_r_implementation

[logical] Used for debugging purposes whether to use the slow R implementation instead of the fast C++ code. Default to FALSE.

...

None used at the moment .

transform_function

[character] The function to apply to the cell values covered by the kernel. For possible values, see pfocal_info_transform(). Default to "MULTIPLY".

reduce_function

[character] The function to apply to the kernel values after the function passed in transform_function has been applied (the function that summarize the data). For possible values, see pfocal_info_reduce(). Default to "SUM".

mean_divider

[character] Optional, allows to specify how the final value at each cell is divided by a value that can be function of the intermediate data resulting of applying transform_function. For possible values, see pfocal_info_mean_divisor(). Default to "ONE" (for no division).

variance

[logical] Whether to return the "variance" of the intermediate values at each point (for more details please see pfocal_info_variance()). Default to FALSE (just returns the value at each point).

vertical_r0

[numeric] The kernel's r0 (exponential) in the vertical dimension.

horizontal_r0

[numeric] The kernel's r0 (exponential) in the horizontal dimension.

height

[numeric] For rectangular kernels, the height of the rectangle.

width

[numeric] For rectangular kernels, the width of the rectangle.

value

[numeric] For single value matrices, the value.

kernel_list

[list] A list of kernels computed from functions in kernel-gaussian, kernel-binomial, kernel-circular, kernel-distance, kernel-exponential.

Value

The updated, convoluted grid.

Examples

data <- matrix(nrow = 10, ncol = 10, data = runif(10 * 10))

pfocal_fast_gaussian_radius(data, vertical_radius = 2)
#>            [,1]      [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
#>  [1,] 0.2134553 0.2788026 0.3054974 0.3218133 0.3130485 0.3313554 0.3845955
#>  [2,] 0.2627849 0.3484208 0.3895275 0.4308621 0.4205016 0.4171930 0.4647470
#>  [3,] 0.3086676 0.4184833 0.4686267 0.5228133 0.5195335 0.5086284 0.5530251
#>  [4,] 0.3392167 0.4809344 0.5341755 0.5808346 0.5845265 0.5696383 0.5966557
#>  [5,] 0.3338703 0.4650717 0.5108895 0.5756712 0.6032686 0.5787493 0.5826413
#>  [6,] 0.3341527 0.4832235 0.5287968 0.5778333 0.6040808 0.5742870 0.5687416
#>  [7,] 0.3357094 0.4687069 0.5046418 0.5520390 0.5939769 0.5733382 0.5677548
#>  [8,] 0.2954107 0.4058729 0.4342791 0.4789219 0.5168871 0.4937271 0.4926081
#>  [9,] 0.2689185 0.3538840 0.3773268 0.4193356 0.4540481 0.4407779 0.4560060
#> [10,] 0.2073702 0.2639796 0.2848680 0.3249752 0.3492130 0.3327592 0.3446954
#>            [,8]      [,9]     [,10]
#>  [1,] 0.3746468 0.2878709 0.1749158
#>  [2,] 0.4591419 0.3630830 0.2162687
#>  [3,] 0.5395255 0.4270478 0.2597962
#>  [4,] 0.5691886 0.4532766 0.2842012
#>  [5,] 0.5597917 0.4638871 0.3027343
#>  [6,] 0.5443467 0.4601814 0.3220403
#>  [7,] 0.5480482 0.4694005 0.3324309
#>  [8,] 0.4920008 0.4329555 0.3106756
#>  [9,] 0.4561162 0.3882053 0.2729838
#> [10,] 0.3446258 0.2857787 0.1903344
pfocal_fast_gaussian_confidence(data)
#>            [,1]      [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
#>  [1,] 0.2353417 0.3169199 0.3352389 0.3129465 0.2944583 0.3404072 0.4242207
#>  [2,] 0.3191522 0.4128880 0.4575047 0.4994409 0.4884618 0.5189785 0.6258480
#>  [3,] 0.3696346 0.4712689 0.5269987 0.6149241 0.6118161 0.5997646 0.6601532
#>  [4,] 0.3457976 0.4654841 0.5403636 0.6458655 0.6449489 0.5814559 0.5686326
#>  [5,] 0.3027244 0.4570216 0.5301416 0.6063520 0.6138929 0.5442814 0.5150990
#>  [6,] 0.3188898 0.5056584 0.5422327 0.5512430 0.5861711 0.5777585 0.5793751
#>  [7,] 0.3506175 0.5124631 0.5141672 0.5065237 0.5720552 0.6003075 0.5999441
#>  [8,] 0.3447440 0.4460059 0.4603458 0.5118322 0.5733350 0.5624664 0.5481465
#>  [9,] 0.3047337 0.3793327 0.4327824 0.5273209 0.5549532 0.5016972 0.5126059
#> [10,] 0.2414076 0.2925200 0.3289449 0.3946708 0.4097410 0.3735311 0.4067313
#>            [,8]      [,9]     [,10]
#>  [1,] 0.4396413 0.3663019 0.2442287
#>  [2,] 0.6303226 0.4790448 0.2714223
#>  [3,] 0.6222372 0.4449882 0.2319075
#>  [4,] 0.5193309 0.4014747 0.2355426
#>  [5,] 0.5034966 0.4511313 0.3082991
#>  [6,] 0.5688562 0.5021922 0.3620367
#>  [7,] 0.5637624 0.4764951 0.3529304
#>  [8,] 0.5282074 0.4611568 0.3498730
#>  [9,] 0.5257845 0.4547913 0.3210997
#> [10,] 0.4288974 0.3551791 0.2179726
pfocal_fast_binomial(data, vertical_radius = 2)
#>           [,1]      [,2]      [,3]     [,4]      [,5]      [,6]     [,7]
#>  [1,] 59.12104  80.74693  85.33681  79.2633  74.92255  87.10381 108.4142
#>  [2,] 82.23311 108.05332 119.90461 127.0182 125.28099 135.00212 160.8714
#>  [3,] 93.13185 120.43457 137.50161 156.7074 157.58218 156.11500 168.8563
#>  [4,] 87.43131 118.69796 140.57707 164.4372 165.29353 150.23987 144.4224
#>  [5,] 77.80619 117.29968 138.92051 155.9188 157.22656 141.60224 132.4334
#>  [6,] 81.92894 128.08913 139.66926 141.5548 148.25238 148.57124 147.2435
#>  [7,] 90.27630 130.30741 132.11214 130.5015 144.72579 154.39850 153.6039
#>  [8,] 88.04905 115.04186 120.48907 132.8986 146.89953 146.56382 142.4004
#>  [9,] 78.11604  98.02841 111.95091 134.6222 141.33050 130.75180 131.3603
#> [10,] 60.09300  74.16947  84.57554 100.8779 103.54684  96.29445 102.7599
#>           [,8]      [,9]    [,10]
#>  [1,] 113.4522  94.74300 61.57473
#>  [2,] 162.7635 123.95030 69.36751
#>  [3,] 160.5395 114.90822 59.00679
#>  [4,] 134.0295 103.57951 60.04613
#>  [5,] 130.0886 115.73499 78.49750
#>  [6,] 144.6721 128.79521 91.32695
#>  [7,] 144.3584 124.41569 90.35864
#>  [8,] 136.5192 120.34618 88.74500
#>  [9,] 134.7474 118.22042 81.45895
#> [10,] 109.0546  91.32076 55.59045
pfocal_fast_abs_rectangle(data, height = 2)
#>            [,1]     [,2]     [,3]      [,4]      [,5]      [,6]     [,7]
#>  [1,] 0.5625439 1.156198 1.431545 0.9315497 0.2848476 0.6731661 1.035549
#>  [2,] 0.9332431 1.623665 1.708434 2.0189392 1.7371492 1.4475618 2.232626
#>  [3,] 1.1081187 2.148898 1.449199 2.2271097 2.9391962 1.7806917 2.595264
#>  [4,] 1.3559912 2.614717 1.934990 2.5568865 3.3040701 2.2872710 2.745735
#>  [5,] 0.7047856 1.658963 1.459354 2.3517987 3.3319439 2.1707322 1.740608
#>  [6,] 0.3740170 1.884841 2.241337 1.8952843 2.7089980 2.1948774 1.673340
#>  [7,] 0.4514815 2.227140 3.217676 1.7736404 2.0407983 2.6615123 2.627289
#>  [8,] 1.0532098 2.409555 2.175686 1.2464401 1.8321911 2.7471721 2.303084
#>  [9,] 1.1186626 1.759460 1.171453 1.9033382 2.8745679 2.3034858 1.621616
#> [10,] 1.0818421 1.629480 1.552138 2.2452582 3.1259322 2.0081593 1.703947
#>           [,8]      [,9]     [,10]
#>  [1,] 1.141348 0.9410667 1.1367764
#>  [2,] 3.107106 2.6329876 1.9591180
#>  [3,] 3.725261 2.9378943 1.3726395
#>  [4,] 3.015261 1.8165035 0.9947082
#>  [5,] 1.532194 1.5666911 1.5108887
#>  [6,] 2.084374 2.4464189 2.4547729
#>  [7,] 3.288148 2.2280085 1.8712101
#>  [8,] 2.209997 1.7504611 1.6371492
#>  [9,] 2.185679 1.9459807 2.0155666
#> [10,] 3.020691 2.4578275 1.7714145

pfocal_fast_separated(data, 
                      kernel_list = list(binomial_kernel(vertical_radius = 2, 
                                                         horizontal_radius = 2), 
                                         distance_kernel(vertical_radius = 2,
                                                         horizontal_radius = 2)))
#>           [,1]     [,2]     [,3]     [,4]     [,5]     [,6]     [,7]     [,8]
#>  [1,] 1599.056 2180.742 2948.297 3268.655 3558.705 3695.900 3495.654 3087.113
#>  [2,] 2120.802 2927.798 3980.586 4378.422 4692.358 4807.393 4478.614 3887.033
#>  [3,] 2819.227 3899.754 5262.181 5761.760 6113.897 6273.320 5938.177 5230.864
#>  [4,] 3115.636 4305.760 5815.212 6388.993 6741.814 6921.529 6601.326 5810.226
#>  [5,] 3230.236 4422.231 6001.024 6620.430 6903.732 7016.581 6699.710 5881.535
#>  [6,] 3143.752 4298.392 5896.300 6530.652 6737.804 6788.898 6510.947 5757.374
#>  [7,] 2992.985 4110.572 5681.370 6298.645 6497.877 6585.676 6393.763 5738.709
#>  [8,] 2763.637 3788.023 5232.680 5767.606 5987.051 6152.912 6032.380 5447.223
#>  [9,] 2041.794 2797.777 3878.171 4240.246 4417.874 4580.345 4490.321 4034.084
#> [10,] 1484.534 2086.556 2893.918 3137.063 3294.789 3446.946 3355.969 2986.187
#>           [,9]    [,10]
#>  [1,] 2402.142 1759.288
#>  [2,] 2991.145 2210.010
#>  [3,] 4030.154 2990.451
#>  [4,] 4416.481 3241.290
#>  [5,] 4372.697 3175.365
#>  [6,] 4233.942 3103.825
#>  [7,] 4254.951 3166.429
#>  [8,] 4070.818 3036.117
#>  [9,] 2998.336 2247.739
#> [10,] 2235.045 1692.109