Fast methods for common kernel computations

Methods wrapping pfocal to implement common kernel computations with default argument values.

pfocal_fast_gaussian_radius(
  data,
  vertical_radius,
  vertical_sd = 1,
  horizontal_radius = vertical_radius,
  horizontal_sd = vertical_sd,
  tail_included = TRUE,
  na.rm = NA,
  mp = TRUE,
  debug_use_r_implementation = FALSE,
  ...,
  transform_function = "MULTIPLY",
  reduce_function = "SUM",
  mean_divider = "ONE",
  variance = FALSE
)

pfocal_fast_gaussian_confidence(
  data,
  vertical_r0 = 0.05,
  vertical_sd = 1,
  horizontal_r0 = vertical_r0,
  horizontal_sd = vertical_sd,
  tail_included = TRUE,
  na.rm = NA,
  mp = TRUE,
  debug_use_r_implementation = FALSE,
  ...,
  transform_function = "MULTIPLY",
  reduce_function = "SUM",
  mean_divider = "ONE",
  variance = FALSE
)

pfocal_fast_binomial(
  data,
  vertical_radius,
  horizontal_radius = vertical_radius,
  na.rm = NA,
  mp = TRUE,
  debug_use_r_implementation = FALSE,
  ...,
  transform_function = "MULTIPLY",
  reduce_function = "SUM",
  mean_divider = "ONE",
  variance = FALSE
)

pfocal_fast_abs_rectangle(
  data,
  height,
  width = height,
  value = 1,
  na.rm = NA,
  mp = TRUE,
  debug_use_r_implementation = FALSE,
  ...,
  transform_function = "MULTIPLY",
  reduce_function = "SUM",
  mean_divider = "ONE",
  variance = FALSE
)

pfocal_fast_separated(
  data,
  kernel_list,
  na.rm = NA,
  mp = TRUE,
  debug_use_r_implementation = FALSE,
  ...,
  transform_function = "MULTIPLY",
  reduce_function = "SUM",
  mean_divider = "ONE",
  variance = FALSE
)

Arguments

data

[matrix-type] Grid to compute onto.

vertical_radius

[numeric] The kernel's radius in the vertical dimension.

vertical_sd

[numeric] The kernel's standard deviation in the vertical dimension.

horizontal_radius

[numeric] The kernel's radius in the horizontal dimension.

horizontal_sd

[numeric] The kernel's standard deviation in the horizontal dimension.

tail_included

[logical] Whether or not to include the kernel tail.

na.rm

[NA OR logical] The behavior to adopt for dealing with missing values, default to NA (faster option). For possible values see pfocal_info_nan_policy().

mp

[logical] Whether to use the open_mp implementation, default to TRUE.

debug_use_r_implementation

[logical] Used for debugging purposes whether to use the slow R implementation instead of the fast C++ code. Default to FALSE.

...

None used at the moment .

transform_function

[character] The function to apply to the cell values covered by the kernel. For possible values, see pfocal_info_transform(). Default to "MULTIPLY".

reduce_function

[character] The function to apply to the kernel values after the function passed in transform_function has been applied (the function that summarize the data). For possible values, see pfocal_info_reduce(). Default to "SUM".

mean_divider

[character] Optional, allows to specify how the final value at each cell is divided by a value that can be function of the intermediate data resulting of applying transform_function. For possible values, see pfocal_info_mean_divisor(). Default to "ONE" (for no division).

variance

[logical] Whether to return the "variance" of the intermediate values at each point (for more details please see pfocal_info_variance()). Default to FALSE (just returns the value at each point).

vertical_r0

[numeric] The kernel's r0 (exponential) in the vertical dimension.

horizontal_r0

[numeric] The kernel's r0 (exponential) in the horizontal dimension.

height

[numeric] For rectangular kernels, the height of the rectangle.

width

[numeric] For rectangular kernels, the width of the rectangle.

value

[numeric] For single value matrices, the value.

kernel_list

[list] A list of kernels computed from functions in kernel-gaussian, kernel-binomial, kernel-circular, kernel-distance, kernel-exponential.

Value

The updated, convoluted grid.

Examples

data <- matrix(nrow = 10, ncol = 10, data = runif(10 * 10))

pfocal_fast_gaussian_radius(data, vertical_radius = 2)
#>            [,1]      [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
#>  [1,] 0.2239344 0.3112392 0.3089392 0.2647087 0.2621423 0.2947582 0.3131279
#>  [2,] 0.2837891 0.4100388 0.4134397 0.3521829 0.3535128 0.4008318 0.4140824
#>  [3,] 0.3141100 0.4765336 0.4908553 0.4086469 0.3874454 0.4325687 0.4574832
#>  [4,] 0.3111009 0.4864187 0.5336082 0.4901482 0.4937281 0.5423759 0.5491005
#>  [5,] 0.2985349 0.4892051 0.5416020 0.4850839 0.4755207 0.5286008 0.5410957
#>  [6,] 0.2594862 0.4217786 0.4861865 0.4720482 0.4714312 0.4960161 0.4927716
#>  [7,] 0.2529410 0.3907820 0.4637919 0.4765088 0.4745828 0.4910718 0.4870333
#>  [8,] 0.2237022 0.3254248 0.3818285 0.4043550 0.4057494 0.4090698 0.3977797
#>  [9,] 0.2003942 0.2681335 0.3081779 0.3278459 0.3059341 0.2871190 0.2894349
#> [10,] 0.1565867 0.1894603 0.2225700 0.2595327 0.2520418 0.2320237 0.2252149
#>            [,8]      [,9]     [,10]
#>  [1,] 0.2799996 0.2266624 0.1687835
#>  [2,] 0.3534364 0.2643690 0.1816665
#>  [3,] 0.4213554 0.3367617 0.2245015
#>  [4,] 0.4892048 0.3806682 0.2410358
#>  [5,] 0.4969049 0.4076427 0.2646569
#>  [6,] 0.4789022 0.4372904 0.3009418
#>  [7,] 0.4869400 0.4747818 0.3477626
#>  [8,] 0.4055091 0.4207913 0.3260696
#>  [9,] 0.3402361 0.3958822 0.3217635
#> [10,] 0.2553098 0.2938639 0.2398155
pfocal_fast_gaussian_confidence(data)
#>            [,1]      [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
#>  [1,] 0.2618318 0.3408152 0.3341798 0.3053850 0.3286601 0.3717465 0.3781890
#>  [2,] 0.3695388 0.5125230 0.4923027 0.3963244 0.3824498 0.4450103 0.4852461
#>  [3,] 0.3682480 0.5483262 0.5461214 0.4369586 0.3957110 0.4362784 0.4787538
#>  [4,] 0.3304889 0.5328203 0.5712941 0.4851709 0.4649391 0.5162174 0.5329268
#>  [5,] 0.2918402 0.5028844 0.5734584 0.5142511 0.5172394 0.5810404 0.5743698
#>  [6,] 0.2367865 0.4229516 0.5207884 0.5216606 0.5312036 0.5696785 0.5593979
#>  [7,] 0.2040082 0.3454863 0.4450627 0.5020199 0.5274907 0.5454900 0.5308779
#>  [8,] 0.2346356 0.3359757 0.3863250 0.4148872 0.4179275 0.4189642 0.4139975
#>  [9,] 0.2617086 0.3155317 0.3349035 0.3449232 0.3127700 0.2848056 0.2898502
#> [10,] 0.2110589 0.2298825 0.2561437 0.2872667 0.2490130 0.2026917 0.2009522
#>            [,8]      [,9]     [,10]
#>  [1,] 0.3110921 0.2545537 0.2172729
#>  [2,] 0.4242306 0.3212361 0.2324940
#>  [3,] 0.4539507 0.3450013 0.2098554
#>  [4,] 0.4881369 0.3736290 0.2171010
#>  [5,] 0.5027930 0.3915238 0.2373872
#>  [6,] 0.5179210 0.4424193 0.2870727
#>  [7,] 0.5128157 0.4818096 0.3427176
#>  [8,] 0.4476328 0.5060314 0.4150460
#>  [9,] 0.3724731 0.4839056 0.4201850
#> [10,] 0.2674613 0.3402899 0.2884508
pfocal_fast_binomial(data, vertical_radius = 2)
#>           [,1]      [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
#>  [1,] 66.97393  87.95129  85.36957  77.78634  82.68912  95.17597  95.89180
#>  [2,] 94.84307 131.63048 126.10949 102.41745  97.64611 113.79663 123.37488
#>  [3,] 95.72226 142.14567 141.41321 113.43969 101.86291 113.62823 125.54845
#>  [4,] 85.66103 137.67256 146.67127 124.90559 118.08081 130.69645 136.56141
#>  [5,] 74.85718 128.75724 146.40814 132.98277 132.40936 147.10471 147.48389
#>  [6,] 60.41034 108.16730 133.67335 135.33133 138.84232 147.92652 146.17098
#>  [7,] 52.26203  89.16593 114.73735 127.54505 134.37329 137.66302 134.85187
#>  [8,] 59.53217  85.84575  99.60084 106.53142 108.52943 107.29846 106.29519
#>  [9,] 66.39757  82.11009  87.19437  88.76147  81.93013  73.09055  74.34939
#> [10,] 52.73285  59.82721  65.52066  71.18230  62.70367  49.69819  49.76853
#>            [,8]      [,9]     [,10]
#>  [1,]  79.71178  65.35665  53.50131
#>  [2,] 108.97770  83.98148  58.94156
#>  [3,] 118.02611  89.93816  53.81175
#>  [4,] 124.52894  94.73993  54.39083
#>  [5,] 130.04290 101.14540  61.08009
#>  [6,] 134.18652 112.35329  72.38984
#>  [7,] 131.94509 123.59680  88.38860
#>  [8,] 117.30192 130.81993 106.75086
#>  [9,]  97.22089 124.11217 107.50131
#> [10,]  68.46413  87.29141  74.21103
pfocal_fast_abs_rectangle(data, height = 2)
#>            [,1]      [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
#>  [1,] 0.6109397 0.7919585 0.6750990 0.7206674 1.2128576 1.2870328 1.0727978
#>  [2,] 1.2408247 2.2541085 2.4666886 1.7620718 1.3736585 2.0419259 2.7344695
#>  [3,] 0.9599846 2.5475524 2.9454719 1.8180336 1.0130165 1.3121847 1.7861920
#>  [4,] 0.6486540 2.0502098 2.6039681 2.0193745 1.5017399 1.4876006 1.3962058
#>  [5,] 0.6180212 1.8100338 2.9806811 2.3272584 1.5461256 2.7058979 3.1459923
#>  [6,] 0.4153884 1.7106913 2.9184870 2.2685855 1.4424150 2.2183716 2.4112628
#>  [7,] 0.2575046 1.0306656 1.6073903 2.2882042 2.2858060 2.2686655 2.3995526
#>  [8,] 0.3483960 1.1453489 1.3583833 1.8961264 2.3317573 2.3787239 2.5452876
#>  [9,] 0.7808070 1.6961728 1.5998072 1.2788930 0.9590801 0.9049274 1.0680831
#> [10,] 1.3171114 1.5505959 0.8460518 1.7687409 1.6066445 1.0428907 0.9253462
#>            [,8]      [,9]    [,10]
#>  [1,] 0.9638610 0.4529168 1.079926
#>  [2,] 2.5837365 1.1676558 1.630375
#>  [3,] 2.2394442 1.4603343 0.760018
#>  [4,] 1.8847420 2.1536330 1.269610
#>  [5,] 2.4833737 1.7897411 1.162296
#>  [6,] 1.9483251 1.8114071 1.467217
#>  [7,] 2.1819136 2.5099127 1.931006
#>  [8,] 1.9815067 1.9335810 2.077199
#>  [9,] 0.8322478 1.8904148 3.229979
#> [10,] 0.9277573 2.0029466 2.677023

pfocal_fast_separated(data, 
                      kernel_list = list(binomial_kernel(vertical_radius = 2, 
                                                         horizontal_radius = 2), 
                                         distance_kernel(vertical_radius = 2,
                                                         horizontal_radius = 2)))
#>           [,1]     [,2]     [,3]     [,4]     [,5]     [,6]     [,7]     [,8]
#>  [1,] 1730.945 2116.431 2765.774 2996.260 2963.315 2783.086 2627.123 2438.568
#>  [2,] 2273.070 2782.428 3674.874 4069.841 4076.528 3832.823 3588.100 3279.983
#>  [3,] 3020.088 3747.646 4940.845 5540.885 5622.342 5337.134 5010.062 4549.733
#>  [4,] 3209.785 4032.440 5290.173 5978.748 6114.670 5852.599 5516.841 4985.622
#>  [5,] 3079.885 3965.925 5228.569 5971.311 6180.198 6009.897 5738.275 5195.599
#>  [6,] 2881.513 3791.299 5022.047 5741.513 5982.522 5918.122 5776.719 5309.245
#>  [7,] 2649.327 3543.576 4701.343 5293.240 5505.948 5539.164 5539.091 5188.131
#>  [8,] 2262.800 3099.038 4153.446 4599.680 4789.580 4912.451 4993.864 4713.302
#>  [9,] 1572.431 2178.957 2965.847 3263.394 3415.431 3572.155 3696.015 3511.443
#> [10,] 1161.095 1587.257 2127.117 2247.005 2301.064 2429.793 2585.878 2536.814
#>           [,9]    [,10]
#>  [1,] 1841.549 1288.731
#>  [2,] 2472.337 1743.873
#>  [3,] 3420.008 2408.958
#>  [4,] 3759.407 2673.075
#>  [5,] 3911.082 2828.296
#>  [6,] 3975.966 2944.503
#>  [7,] 3869.835 2956.525
#>  [8,] 3487.700 2736.861
#>  [9,] 2560.426 2049.702
#> [10,] 1885.381 1562.439