Fast methods for common kernel computations

Methods wrapping pfocal to implement common kernel computations with default argument values.

pfocal_fast_gaussian_radius(
  data,
  vertical_radius,
  vertical_sd = 1,
  horizontal_radius = vertical_radius,
  horizontal_sd = vertical_sd,
  tail_included = TRUE,
  na.rm = NA,
  mp = TRUE,
  debug_use_r_implementation = FALSE,
  ...,
  transform_function = "MULTIPLY",
  reduce_function = "SUM",
  mean_divider = "ONE",
  variance = FALSE
)

pfocal_fast_gaussian_confidence(
  data,
  vertical_r0 = 0.05,
  vertical_sd = 1,
  horizontal_r0 = vertical_r0,
  horizontal_sd = vertical_sd,
  tail_included = TRUE,
  na.rm = NA,
  mp = TRUE,
  debug_use_r_implementation = FALSE,
  ...,
  transform_function = "MULTIPLY",
  reduce_function = "SUM",
  mean_divider = "ONE",
  variance = FALSE
)

pfocal_fast_binomial(
  data,
  vertical_radius,
  horizontal_radius = vertical_radius,
  na.rm = NA,
  mp = TRUE,
  debug_use_r_implementation = FALSE,
  ...,
  transform_function = "MULTIPLY",
  reduce_function = "SUM",
  mean_divider = "ONE",
  variance = FALSE
)

pfocal_fast_abs_rectangle(
  data,
  height,
  width = height,
  value = 1,
  na.rm = NA,
  mp = TRUE,
  debug_use_r_implementation = FALSE,
  ...,
  transform_function = "MULTIPLY",
  reduce_function = "SUM",
  mean_divider = "ONE",
  variance = FALSE
)

pfocal_fast_separated(
  data,
  kernel_list,
  na.rm = NA,
  mp = TRUE,
  debug_use_r_implementation = FALSE,
  ...,
  transform_function = "MULTIPLY",
  reduce_function = "SUM",
  mean_divider = "ONE",
  variance = FALSE
)

Arguments

data

[matrix-type] Grid to compute onto.

vertical_radius

[numeric] The kernel's radius in the vertical dimension.

vertical_sd

[numeric] The kernel's standard deviation in the vertical dimension.

horizontal_radius

[numeric] The kernel's radius in the horizontal dimension.

horizontal_sd

[numeric] The kernel's standard deviation in the horizontal dimension.

tail_included

[logical] Whether or not to include the kernel tail.

na.rm

[NA OR logical] The behavior to adopt for dealing with missing values, default to NA (faster option). For possible values see pfocal_info_nan_policy().

mp

[logical] Whether to use the open_mp implementation, default to TRUE.

debug_use_r_implementation

[logical] Used for debugging purposes whether to use the slow R implementation instead of the fass C++ code. Default to FALSE.

...

None used at the moment .

transform_function

[character] The function to apply to the cell values covered by the kernel. For possible values, see pfocal_info_transform(). Default to "MULTIPLY".

reduce_function

[character] The function to apply to the kernel values after the function passed in transform_function has been applied (the function that summarize the data). For possible values, see pfocal_info_reduce(). Default to "SUM".

mean_divider

[character] Optional, allows to specify how the final value at each cell is divided by a value that can be function of the intermediate data resulting of applying transform_function. For possible values, see pfocal_info_mean_divisor(). Default to "ONE" (for no division).

variance

[logical] Whether to return the "variance" of the intermediate values at each point (for more details please see pfocal_info_variance()). Default to FALSE (just returns the value at each point).

vertical_r0

[numeric] The kernel's r0 (exponential) in the vertical dimension.

horizontal_r0

[numeric] The kernel's r0 (exponential) in the horizontal dimension.

height

[numeric] For rectangular kernels, the height of the rectangle.

width

[numeric] For rectangular kernels, the width of the rectangle.

value

[numeric] For single value matrices, the value.

kernel_list

[list] A list of kernels computed from functions in kernel-gaussian, kernel-binomial, kernel-circular, kernel-distance, kernel-exponential.

Value

The updated, convoluted grid.

Examples

data <- matrix(nrow = 10, ncol = 10, data = runif(10 * 10))

pfocal_fast_gaussian_radius(data, vertical_radius = 2)
#>            [,1]      [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
#>  [1,] 0.1774017 0.2659878 0.2890475 0.2915179 0.3109371 0.2995465 0.2670823
#>  [2,] 0.2282219 0.3437515 0.3634595 0.3613956 0.4036578 0.4161852 0.3838537
#>  [3,] 0.2573470 0.3930709 0.4244327 0.4209535 0.4614886 0.4755812 0.4391094
#>  [4,] 0.2836624 0.4514878 0.4962020 0.4846916 0.5382952 0.5820862 0.5450657
#>  [5,] 0.2927643 0.4766332 0.5278709 0.5026136 0.5393788 0.5801281 0.5509933
#>  [6,] 0.2766402 0.4659176 0.5353901 0.5117699 0.5325383 0.5618340 0.5265762
#>  [7,] 0.2757704 0.4742522 0.5726889 0.5618975 0.5559609 0.5544836 0.5260891
#>  [8,] 0.2412455 0.4150715 0.5034190 0.4952295 0.4794573 0.4616007 0.4398495
#>  [9,] 0.1994389 0.3356788 0.4265262 0.4402639 0.4070770 0.3535738 0.3292928
#> [10,] 0.1516823 0.2483745 0.3213681 0.3473269 0.3299100 0.2880788 0.2696293
#>            [,8]      [,9]     [,10]
#>  [1,] 0.2508150 0.2411159 0.1972633
#>  [2,] 0.3598554 0.3499351 0.2858533
#>  [3,] 0.3996525 0.3838392 0.3175560
#>  [4,] 0.4724440 0.4293816 0.3435473
#>  [5,] 0.4858523 0.4343572 0.3300194
#>  [6,] 0.4624385 0.4176983 0.3134177
#>  [7,] 0.4786271 0.4137739 0.2845332
#>  [8,] 0.4284370 0.3861942 0.2573805
#>  [9,] 0.3389503 0.3129197 0.2053204
#> [10,] 0.2808489 0.2610365 0.1705741
pfocal_fast_gaussian_confidence(data)
#>            [,1]      [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
#>  [1,] 0.1995643 0.3196200 0.3568646 0.3479556 0.3483680 0.3153464 0.2901759
#>  [2,] 0.2849809 0.3972352 0.3924047 0.3825716 0.4244034 0.4167595 0.3817506
#>  [3,] 0.2884345 0.4013603 0.4055471 0.4209140 0.4987143 0.5210367 0.4603125
#>  [4,] 0.2944661 0.4552301 0.5032857 0.5158292 0.5922770 0.6489915 0.5928978
#>  [5,] 0.2871580 0.4887511 0.5549500 0.5288088 0.5758771 0.6484538 0.6328981
#>  [6,] 0.2648297 0.4767821 0.5507401 0.4989493 0.5125364 0.5672066 0.5640350
#>  [7,] 0.2720384 0.4915813 0.5783681 0.5295258 0.5244626 0.5482001 0.5213807
#>  [8,] 0.2884828 0.4948294 0.5978438 0.5790288 0.5419283 0.4989106 0.4498539
#>  [9,] 0.2561184 0.4033380 0.5128088 0.5465766 0.4968294 0.4060321 0.3664448
#> [10,] 0.1527149 0.2400846 0.3424188 0.4114836 0.3805320 0.2946803 0.2831258
#>            [,8]      [,9]     [,10]
#>  [1,] 0.3018911 0.2815195 0.2108576
#>  [2,] 0.3866706 0.3892211 0.3151550
#>  [3,] 0.4121807 0.4253996 0.3706158
#>  [4,] 0.4894879 0.4632444 0.4015625
#>  [5,] 0.5414377 0.4768523 0.3797495
#>  [6,] 0.4897918 0.4000601 0.2778641
#>  [7,] 0.4562188 0.3645636 0.2180925
#>  [8,] 0.4383227 0.4011348 0.2544498
#>  [9,] 0.4038026 0.4094362 0.2881035
#> [10,] 0.3317348 0.3243360 0.2206833
pfocal_fast_binomial(data, vertical_radius = 2)
#>           [,1]      [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
#>  [1,] 51.41189  81.09540  90.11259  88.42083  86.69507  79.41564  73.93215
#>  [2,] 72.64056 101.23827 101.45461 100.09420 108.63152 107.10455  98.02471
#>  [3,] 75.39026 104.14771 105.79229 109.92506 128.73143 135.32894 119.63522
#>  [4,] 75.37321 115.79848 127.83488 131.76140 150.85024 165.63444 150.97950
#>  [5,] 73.77435 124.62157 141.71762 136.81222 147.32788 165.99383 161.30838
#>  [6,] 68.95304 123.07482 141.41650 129.93331 131.92594 147.34431 146.92440
#>  [7,] 70.37281 126.48980 148.18911 137.30924 133.40472 138.64773 132.81068
#>  [8,] 74.48956 127.27057 153.17053 149.56464 138.83883 127.30231 115.95205
#>  [9,] 64.98046 104.43769 133.07753 142.29662 128.79938 105.01811  95.27680
#> [10,] 38.62010  61.85351  87.83755 104.97062  96.15417  74.63544  71.04075
#>            [,8]      [,9]     [,10]
#>  [1,]  76.26057  72.67317  53.46582
#>  [2,]  98.62448 100.35953  80.10475
#>  [3,] 107.71467 110.30343  94.65829
#>  [4,] 127.06376 120.16443 102.22523
#>  [5,] 138.71541 121.40244  95.09598
#>  [6,] 128.10436 103.57834  70.57122
#>  [7,] 117.21892  93.02559  55.73929
#>  [8,] 112.66778 101.89824  65.42360
#>  [9,] 105.29759 105.99796  74.13009
#> [10,]  83.67523  82.97079  56.72061
pfocal_fast_abs_rectangle(data, height = 2)
#>            [,1]      [,2]     [,3]      [,4]     [,5]     [,6]      [,7]
#>  [1,] 0.0220953 0.5663456 1.436646 1.4360785 1.353602 1.289554 0.7031918
#>  [2,] 0.7897827 1.9722722 2.414886 1.7925603 1.835567 2.111963 1.3397218
#>  [3,] 0.8455139 1.9078823 1.442640 0.5654242 1.275620 1.935643 1.3650277
#>  [4,] 0.6315980 1.5243197 1.759202 1.6699928 2.190967 2.718949 2.5457026
#>  [5,] 0.6775565 2.0089857 2.840231 2.3347018 2.196627 3.192281 3.7260864
#>  [6,] 0.3093409 1.5296590 2.772945 1.9826307 1.630956 2.183695 2.3405451
#>  [7,] 0.4374283 1.3601262 2.585320 1.9165728 1.375689 2.017385 2.3027090
#>  [8,] 0.3381822 1.7042000 3.122254 2.5270005 2.027615 2.690031 2.5215816
#>  [9,] 0.7734332 1.8436377 2.549500 2.7325138 2.308868 1.971320 1.2644353
#> [10,] 0.8460260 1.2046760 1.409125 2.3030776 2.553367 1.743391 1.1298615
#>            [,8]      [,9]     [,10]
#>  [1,] 0.9405658 0.9874282 0.6852845
#>  [2,] 1.8750169 1.8862686 1.5481855
#>  [3,] 1.2238718 1.6776943 2.1581812
#>  [4,] 1.4457991 1.0555517 2.2032495
#>  [5,] 2.8149023 1.7904010 2.5581778
#>  [6,] 2.4372013 2.0185665 2.0684640
#>  [7,] 2.0046062 1.1965986 0.9266181
#>  [8,] 1.9708591 2.1371053 1.3309510
#>  [9,] 1.0690461 1.9822529 1.9961784
#> [10,] 1.6331257 2.0083388 2.0077733

pfocal_fast_separated(data, 
                      kernel_list = list(binomial_kernel(vertical_radius = 2, 
                                                         horizontal_radius = 2), 
                                         distance_kernel(vertical_radius = 2,
                                                         horizontal_radius = 2)))
#>           [,1]     [,2]     [,3]     [,4]     [,5]     [,6]     [,7]     [,8]
#>  [1,] 1386.881 1813.977 2504.858 2782.503 2787.103 2775.300 2789.526 2590.253
#>  [2,] 1902.231 2518.080 3516.592 3983.922 4031.972 3990.388 3995.007 3718.736
#>  [3,] 2649.474 3516.904 4840.914 5499.855 5580.203 5483.815 5459.105 5101.431
#>  [4,] 2869.320 3767.135 5153.379 5909.226 6032.017 5897.705 5848.775 5479.436
#>  [5,] 3017.881 3952.459 5366.060 6222.442 6394.671 6183.703 6013.189 5553.327
#>  [6,] 3191.479 4193.446 5615.358 6492.826 6672.376 6402.059 6112.961 5555.560
#>  [7,] 3146.009 4160.253 5500.801 6269.345 6424.259 6180.777 5854.416 5251.388
#>  [8,] 2801.564 3759.558 4942.379 5553.916 5698.995 5543.837 5218.744 4581.065
#>  [9,] 2132.802 2871.800 3743.733 4137.135 4220.261 4118.660 3847.195 3295.708
#> [10,] 1614.573 2225.755 2880.437 3110.096 3144.498 3086.073 2883.692 2464.829
#>           [,9]    [,10]
#>  [1,] 1861.408 1408.386
#>  [2,] 2640.314 1948.419
#>  [3,] 3662.390 2663.719
#>  [4,] 3952.620 2841.419
#>  [5,] 4011.486 2850.284
#>  [6,] 4049.934 2896.064
#>  [7,] 3869.259 2829.537
#>  [8,] 3365.171 2506.850
#>  [9,] 2371.831 1790.253
#> [10,] 1785.253 1387.179