Fast methods for common kernel computations

Methods wrapping pfocal to implement common kernel computations with default argument values.

pfocal_fast_gaussian_radius(
  data,
  vertical_radius,
  vertical_sd = 1,
  horizontal_radius = vertical_radius,
  horizontal_sd = vertical_sd,
  tail_included = TRUE,
  na.rm = NA,
  mp = TRUE,
  debug_use_r_implementation = FALSE,
  ...,
  transform_function = "MULTIPLY",
  reduce_function = "SUM",
  mean_divider = "ONE",
  variance = FALSE
)

pfocal_fast_gaussian_confidence(
  data,
  vertical_r0 = 0.05,
  vertical_sd = 1,
  horizontal_r0 = vertical_r0,
  horizontal_sd = vertical_sd,
  tail_included = TRUE,
  na.rm = NA,
  mp = TRUE,
  debug_use_r_implementation = FALSE,
  ...,
  transform_function = "MULTIPLY",
  reduce_function = "SUM",
  mean_divider = "ONE",
  variance = FALSE
)

pfocal_fast_binomial(
  data,
  vertical_radius,
  horizontal_radius = vertical_radius,
  na.rm = NA,
  mp = TRUE,
  debug_use_r_implementation = FALSE,
  ...,
  transform_function = "MULTIPLY",
  reduce_function = "SUM",
  mean_divider = "ONE",
  variance = FALSE
)

pfocal_fast_abs_rectangle(
  data,
  height,
  width = height,
  value = 1,
  na.rm = NA,
  mp = TRUE,
  debug_use_r_implementation = FALSE,
  ...,
  transform_function = "MULTIPLY",
  reduce_function = "SUM",
  mean_divider = "ONE",
  variance = FALSE
)

pfocal_fast_separated(
  data,
  kernel_list,
  na.rm = NA,
  mp = TRUE,
  debug_use_r_implementation = FALSE,
  ...,
  transform_function = "MULTIPLY",
  reduce_function = "SUM",
  mean_divider = "ONE",
  variance = FALSE
)

Arguments

data

[matrix-type] Grid to compute onto.

vertical_radius

[numeric] The kernel's radius in the vertical dimension.

vertical_sd

[numeric] The kernel's standard deviation in the vertical dimension.

horizontal_radius

[numeric] The kernel's radius in the horizontal dimension.

horizontal_sd

[numeric] The kernel's standard deviation in the horizontal dimension.

tail_included

[logical] Whether or not to include the kernel tail.

na.rm

[NA OR logical] The behavior to adopt for dealing with missing values, default to NA (faster option). For possible values see pfocal_info_nan_policy().

mp

[logical] Whether to use the open_mp implementation, default to TRUE.

debug_use_r_implementation

[logical] Used for debugging purposes whether to use the slow R implementation instead of the fass C++ code. Default to FALSE.

...

None used at the moment .

transform_function

[character] The function to apply to the cell values covered by the kernel. For possible values, see pfocal_info_transform(). Default to "MULTIPLY".

reduce_function

[character] The function to apply to the kernel values after the function passed in transform_function has been applied (the function that summarize the data). For possible values, see pfocal_info_reduce(). Default to "SUM".

mean_divider

[character] Optional, allows to specify how the final value at each cell is divided by a value that can be function of the intermediate data resulting of applying transform_function. For possible values, see pfocal_info_mean_divisor(). Default to "ONE" (for no division).

variance

[logical] Whether to return the "variance" of the intermediate values at each point (for more details please see pfocal_info_variance()). Default to FALSE (just returns the value at each point).

vertical_r0

[numeric] The kernel's r0 (exponential) in the vertical dimension.

horizontal_r0

[numeric] The kernel's r0 (exponential) in the horizontal dimension.

height

[numeric] For rectangular kernels, the height of the rectangle.

width

[numeric] For rectangular kernels, the width of the rectangle.

value

[numeric] For single value matrices, the value.

kernel_list

[list] A list of kernels computed from functions in kernel-gaussian, kernel-binomial, kernel-circular, kernel-distance, kernel-exponential.

Value

The updated, convoluted grid.

Examples

data <- matrix(nrow = 10, ncol = 10, data = runif(10 * 10))

pfocal_fast_gaussian_radius(data, vertical_radius = 2)
#>            [,1]      [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
#>  [1,] 0.1563763 0.2359805 0.2805013 0.3120259 0.3211696 0.2934751 0.2379406
#>  [2,] 0.2170537 0.3052831 0.3336558 0.3706940 0.4096681 0.4016119 0.3384094
#>  [3,] 0.2398882 0.3254354 0.3455859 0.3855303 0.4527666 0.4810415 0.4296530
#>  [4,] 0.2798648 0.3689962 0.3705644 0.3850862 0.4537933 0.5211938 0.5118654
#>  [5,] 0.2674627 0.3648170 0.3776450 0.3781027 0.4285029 0.5046203 0.5131439
#>  [6,] 0.2647005 0.3575944 0.3655833 0.3506650 0.3992628 0.4885125 0.5218878
#>  [7,] 0.2949819 0.4048118 0.4271835 0.3890419 0.3851389 0.4570829 0.5051350
#>  [8,] 0.2676135 0.3815397 0.4165956 0.3817943 0.3591542 0.3932839 0.4217874
#>  [9,] 0.2413000 0.3406480 0.3848738 0.3659147 0.3369947 0.3484928 0.3608942
#> [10,] 0.2224720 0.2944988 0.3176658 0.3033613 0.2777503 0.2816184 0.2902485
#>            [,8]      [,9]     [,10]
#>  [1,] 0.1889753 0.1786523 0.1608238
#>  [2,] 0.2637562 0.2208229 0.1785840
#>  [3,] 0.3553970 0.3018614 0.2324459
#>  [4,] 0.4415023 0.3642377 0.2709561
#>  [5,] 0.4609928 0.3936427 0.2878009
#>  [6,] 0.5142024 0.4703256 0.3431686
#>  [7,] 0.5041365 0.4731668 0.3553055
#>  [8,] 0.4382797 0.4313157 0.3245839
#>  [9,] 0.3819395 0.3911925 0.2988529
#> [10,] 0.2912600 0.2794020 0.2063380
pfocal_fast_gaussian_confidence(data)
#>            [,1]      [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
#>  [1,] 0.1709907 0.3071928 0.3891779 0.3990579 0.3496061 0.2655952 0.1869846
#>  [2,] 0.2286066 0.3772298 0.4690101 0.5366511 0.5473613 0.4632036 0.3420472
#>  [3,] 0.2628670 0.3406485 0.3756601 0.4751833 0.5842788 0.5983075 0.5091357
#>  [4,] 0.3187238 0.3423793 0.2868522 0.3316511 0.4815558 0.6144973 0.6063697
#>  [5,] 0.3133689 0.3669015 0.2946739 0.2658618 0.3764178 0.5353260 0.5753975
#>  [6,] 0.2441066 0.3610012 0.3639037 0.3022992 0.3245270 0.4375083 0.5093274
#>  [7,] 0.2175418 0.3725631 0.4533695 0.3933375 0.3305468 0.3795498 0.4615834
#>  [8,] 0.2559026 0.4015898 0.4956316 0.4663079 0.3997732 0.4015733 0.4326901
#>  [9,] 0.3338716 0.4370329 0.4725966 0.4666553 0.4527952 0.4524587 0.4383590
#> [10,] 0.3154430 0.3821611 0.3821579 0.3770528 0.3671012 0.3687520 0.3513815
#>            [,8]      [,9]     [,10]
#>  [1,] 0.1507168 0.1876802 0.2047787
#>  [2,] 0.2629038 0.2521012 0.2290270
#>  [3,] 0.3981119 0.3033171 0.2105841
#>  [4,] 0.4877380 0.3474223 0.2224803
#>  [5,] 0.5022476 0.3967015 0.2751317
#>  [6,] 0.5166543 0.4799820 0.3641876
#>  [7,] 0.5118266 0.5137763 0.3992253
#>  [8,] 0.4879519 0.5222088 0.4036352
#>  [9,] 0.4626361 0.4837698 0.3651383
#> [10,] 0.3232324 0.2934222 0.2091509
pfocal_fast_binomial(data, vertical_radius = 2)
#>           [,1]      [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
#>  [1,] 43.39199  78.41464 100.43077 103.13729  90.04086  67.89323  46.25765
#>  [2,] 58.15137  95.86625 121.38726 138.62501 140.94594 119.52111  87.63695
#>  [3,] 67.98513  88.29908  98.03325 122.34655 150.60382 154.49650 131.90038
#>  [4,] 80.73913  87.50247  73.65563  84.42595 124.60780 158.25708 156.12041
#>  [5,] 79.81177  93.22777  74.44889  66.56488  96.19536 137.28475 149.34047
#>  [6,] 63.39980  93.68650  93.28218  77.10247  82.26383 110.65529 131.10861
#>  [7,] 54.53175  95.79382 115.68014 101.10703  85.05945  95.58343 116.84945
#>  [8,] 65.75054 104.98914 127.12676 119.92866 103.81713 102.54808 112.14703
#>  [9,] 85.27760 114.20998 123.13574 121.09884 117.23757 114.81232 112.58147
#> [10,] 79.28298  97.71438  97.39951  95.10653  94.38660  93.81706  89.35911
#>            [,8]      [,9]     [,10]
#>  [1,]  38.06814  47.81589  50.93638
#>  [2,]  67.87654  65.65960  58.22607
#>  [3,] 102.24070  77.96808  53.52446
#>  [4,] 124.94830  88.55076  55.13611
#>  [5,] 130.84153 103.42427  70.55437
#>  [6,] 133.08375 122.66115  91.50191
#>  [7,] 131.89229 133.03580 102.39325
#>  [8,] 128.25987 135.96518 104.81053
#>  [9,] 119.03534 122.59355  92.40992
#> [10,]  83.46399  75.72555  54.01710
pfocal_fast_abs_rectangle(data, height = 2)
#>            [,1]      [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
#>  [1,] 0.2357306 0.6369939 1.1980288 1.3183180 0.8515494 0.5440510 0.4629209
#>  [2,] 0.3391664 1.6067452 2.5105454 2.5488409 2.5156656 1.7775511 1.1987390
#>  [3,] 0.3357987 1.4118732 1.6555216 1.9796667 3.0988745 3.0462505 1.8400103
#>  [4,] 1.1197626 1.5503283 0.5844465 1.0104600 1.8639809 2.9085280 3.0030716
#>  [5,] 1.4624176 2.5399372 1.1500133 0.3703424 0.7856544 2.0858562 3.2333862
#>  [6,] 0.5796341 1.7145796 1.7025566 0.8422998 0.6331097 1.8992943 2.2652809
#>  [7,] 0.3130275 1.0281356 1.9667360 1.6677160 0.7116159 1.1480354 1.8484692
#>  [8,] 0.4665518 0.9641454 2.2243575 2.8357301 1.3750191 0.6994677 1.5865792
#>  [9,] 0.8337999 1.3707600 1.8128523 2.2363916 1.6917872 2.0550422 1.6550480
#> [10,] 1.5959300 2.7256650 1.9665865 1.8846438 2.0151770 2.3843369 2.3057114
#>            [,8]      [,9]     [,10]
#>  [1,] 0.2848872 0.2905867 1.0000223
#>  [2,] 0.7037551 0.4195345 1.6907043
#>  [3,] 1.2186453 1.0875414 0.9635426
#>  [4,] 2.3912024 1.8245539 1.0925556
#>  [5,] 2.7059736 1.4148938 0.9903770
#>  [6,] 1.8359302 1.9035237 1.7547610
#>  [7,] 2.3572743 2.3834598 2.5536552
#>  [8,] 1.9092683 1.8372364 2.0830965
#>  [9,] 1.1961144 2.6155541 2.9632838
#> [10,] 1.9225518 2.1841894 2.1220260

pfocal_fast_separated(data, 
                      kernel_list = list(binomial_kernel(vertical_radius = 2, 
                                                         horizontal_radius = 2), 
                                         distance_kernel(vertical_radius = 2,
                                                         horizontal_radius = 2)))
#>           [,1]     [,2]     [,3]     [,4]     [,5]     [,6]     [,7]     [,8]
#>  [1,] 1355.126 1903.719 2631.968 2961.720 3008.417 2899.167 2680.698 2275.992
#>  [2,] 1695.035 2324.916 3285.509 3716.942 3834.157 3810.631 3659.470 3210.364
#>  [3,] 2243.554 3054.900 4310.253 4852.510 4950.330 4881.561 4744.002 4293.495
#>  [4,] 2299.142 3121.222 4472.031 5127.752 5312.766 5308.792 5256.024 4875.374
#>  [5,] 2276.939 3021.739 4300.119 5008.071 5320.806 5423.781 5478.268 5205.217
#>  [6,] 2398.355 3073.217 4211.744 4871.055 5228.256 5388.243 5521.719 5337.483
#>  [7,] 2583.001 3262.660 4310.250 4870.530 5174.657 5317.913 5461.675 5297.447
#>  [8,] 2582.731 3290.371 4276.709 4717.590 4920.636 4987.749 5059.298 4865.396
#>  [9,] 1944.695 2514.951 3289.607 3596.006 3707.620 3727.660 3767.910 3622.435
#> [10,] 1540.481 2027.106 2685.705 2891.264 2922.263 2907.659 2915.679 2773.553
#>           [,9]     [,10]
#>  [1,] 1498.786  997.3827
#>  [2,] 2193.048 1459.1839
#>  [3,] 3056.919 2103.9810
#>  [4,] 3587.203 2529.1969
#>  [5,] 3923.601 2833.3038
#>  [6,] 4066.277 3021.7790
#>  [7,] 4033.774 3081.3461
#>  [8,] 3686.332 2857.0275
#>  [9,] 2743.369 2167.6080
#> [10,] 2099.853 1651.0302