Fast methods for common kernel computations

Methods wrapping pfocal to implement common kernel computations with default argument values.

pfocal_fast_gaussian_radius(
  data,
  vertical_radius,
  vertical_sd = 1,
  horizontal_radius = vertical_radius,
  horizontal_sd = vertical_sd,
  tail_included = TRUE,
  na.rm = NA,
  mp = TRUE,
  debug_use_r_implementation = FALSE,
  ...,
  transform_function = "MULTIPLY",
  reduce_function = "SUM",
  mean_divider = "ONE",
  variance = FALSE
)

pfocal_fast_gaussian_confidence(
  data,
  vertical_r0 = 0.05,
  vertical_sd = 1,
  horizontal_r0 = vertical_r0,
  horizontal_sd = vertical_sd,
  tail_included = TRUE,
  na.rm = NA,
  mp = TRUE,
  debug_use_r_implementation = FALSE,
  ...,
  transform_function = "MULTIPLY",
  reduce_function = "SUM",
  mean_divider = "ONE",
  variance = FALSE
)

pfocal_fast_binomial(
  data,
  vertical_radius,
  horizontal_radius = vertical_radius,
  na.rm = NA,
  mp = TRUE,
  debug_use_r_implementation = FALSE,
  ...,
  transform_function = "MULTIPLY",
  reduce_function = "SUM",
  mean_divider = "ONE",
  variance = FALSE
)

pfocal_fast_abs_rectangle(
  data,
  height,
  width = height,
  value = 1,
  na.rm = NA,
  mp = TRUE,
  debug_use_r_implementation = FALSE,
  ...,
  transform_function = "MULTIPLY",
  reduce_function = "SUM",
  mean_divider = "ONE",
  variance = FALSE
)

pfocal_fast_separated(
  data,
  kernel_list,
  na.rm = NA,
  mp = TRUE,
  debug_use_r_implementation = FALSE,
  ...,
  transform_function = "MULTIPLY",
  reduce_function = "SUM",
  mean_divider = "ONE",
  variance = FALSE
)

Arguments

data

[matrix-type] Grid to compute onto.

vertical_radius

[numeric] The kernel's radius in the vertical dimension.

vertical_sd

[numeric] The kernel's standard deviation in the vertical dimension.

horizontal_radius

[numeric] The kernel's radius in the horizontal dimension.

horizontal_sd

[numeric] The kernel's standard deviation in the horizontal dimension.

tail_included

[logical] Whether or not to include the kernel tail.

na.rm

[NA OR logical] The behavior to adopt for dealing with missing values, default to NA (faster option). For possible values see pfocal_info_nan_policy().

mp

[logical] Whether to use the open_mp implementation, default to TRUE.

debug_use_r_implementation

[logical] Used for debugging purposes whether to use the slow R implementation instead of the fast C++ code. Default to FALSE.

...

None used at the moment .

transform_function

[character] The function to apply to the cell values covered by the kernel. For possible values, see pfocal_info_transform(). Default to "MULTIPLY".

reduce_function

[character] The function to apply to the kernel values after the function passed in transform_function has been applied (the function that summarize the data). For possible values, see pfocal_info_reduce(). Default to "SUM".

mean_divider

[character] Optional, allows to specify how the final value at each cell is divided by a value that can be function of the intermediate data resulting of applying transform_function. For possible values, see pfocal_info_mean_divisor(). Default to "ONE" (for no division).

variance

[logical] Whether to return the "variance" of the intermediate values at each point (for more details please see pfocal_info_variance()). Default to FALSE (just returns the value at each point).

vertical_r0

[numeric] The kernel's r0 (exponential) in the vertical dimension.

horizontal_r0

[numeric] The kernel's r0 (exponential) in the horizontal dimension.

height

[numeric] For rectangular kernels, the height of the rectangle.

width

[numeric] For rectangular kernels, the width of the rectangle.

value

[numeric] For single value matrices, the value.

kernel_list

[list] A list of kernels computed from functions in kernel-gaussian, kernel-binomial, kernel-circular, kernel-distance, kernel-exponential.

Value

The updated, convoluted grid.

Examples

data <- matrix(nrow = 10, ncol = 10, data = runif(10 * 10))

pfocal_fast_gaussian_radius(data, vertical_radius = 2)
#>            [,1]      [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
#>  [1,] 0.2823537 0.3497556 0.3558794 0.3064959 0.2261171 0.1876028 0.2043062
#>  [2,] 0.3597434 0.4469663 0.4511789 0.3964480 0.2988670 0.2504659 0.2843589
#>  [3,] 0.4197448 0.5073973 0.4840909 0.4177389 0.3292317 0.2821029 0.3132842
#>  [4,] 0.4638896 0.5703354 0.5412636 0.4664599 0.3826397 0.3428016 0.3748651
#>  [5,] 0.4605674 0.5731372 0.5325097 0.4583754 0.3961497 0.3796631 0.4224269
#>  [6,] 0.4258651 0.5200700 0.4855465 0.4551646 0.4485667 0.4584461 0.4785817
#>  [7,] 0.4016272 0.5168712 0.4983941 0.4688294 0.4728381 0.4928345 0.5098730
#>  [8,] 0.3267285 0.4263191 0.4300934 0.4398669 0.4717302 0.5062814 0.5219891
#>  [9,] 0.2563479 0.3381028 0.3491679 0.3703353 0.4183897 0.4552692 0.4558599
#> [10,] 0.1752465 0.2515398 0.2818274 0.3052446 0.3401087 0.3668777 0.3687736
#>            [,8]      [,9]     [,10]
#>  [1,] 0.2535046 0.2910551 0.2356447
#>  [2,] 0.3485265 0.3785563 0.3002202
#>  [3,] 0.3785631 0.4060644 0.3222902
#>  [4,] 0.4376244 0.4681017 0.3787725
#>  [5,] 0.4614369 0.4574452 0.3676981
#>  [6,] 0.4743442 0.4589366 0.3847673
#>  [7,] 0.4956629 0.4743866 0.3966944
#>  [8,] 0.4865775 0.4487513 0.3771545
#>  [9,] 0.4165372 0.3851650 0.3223375
#> [10,] 0.3431905 0.3207381 0.2621968
pfocal_fast_gaussian_confidence(data)
#>            [,1]      [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
#>  [1,] 0.3308068 0.4078024 0.4307352 0.3898342 0.2922271 0.2408049 0.2567692
#>  [2,] 0.4294760 0.5350699 0.5462980 0.4816553 0.3691379 0.3200971 0.3627313
#>  [3,] 0.4464391 0.5488756 0.5362554 0.4539978 0.3473904 0.3046034 0.3521244
#>  [4,] 0.4716721 0.5825681 0.5510466 0.4561109 0.3388186 0.2766929 0.3149882
#>  [5,] 0.4837956 0.5851129 0.5264339 0.4560545 0.3800649 0.3264658 0.3595997
#>  [6,] 0.4618978 0.5508067 0.4722891 0.4207166 0.4231886 0.4287106 0.4492863
#>  [7,] 0.4155772 0.5305096 0.4916400 0.4493379 0.4810745 0.5446357 0.5697658
#>  [8,] 0.3525802 0.4728929 0.4959349 0.5099087 0.5552342 0.6295421 0.6494778
#>  [9,] 0.2692284 0.3692139 0.4239389 0.4942216 0.5624846 0.6047373 0.5948434
#> [10,] 0.1693969 0.2506472 0.2968319 0.3453728 0.3945214 0.4039564 0.3957460
#>            [,8]      [,9]     [,10]
#>  [1,] 0.3114621 0.3668103 0.3127745
#>  [2,] 0.4423129 0.4852248 0.3829513
#>  [3,] 0.4398418 0.4728890 0.3570377
#>  [4,] 0.3911034 0.4085121 0.3104373
#>  [5,] 0.4097442 0.4065859 0.3296094
#>  [6,] 0.4636332 0.4651574 0.4013042
#>  [7,] 0.5398868 0.5253104 0.4555373
#>  [8,] 0.5666838 0.5003456 0.4259169
#>  [9,] 0.5130572 0.4444086 0.3668682
#> [10,] 0.3846372 0.3651160 0.2878012
pfocal_fast_binomial(data, vertical_radius = 2)
#>            [,1]      [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
#>  [1,]  83.01324 105.46976 110.28190  98.82844  75.14318  61.16515  65.28937
#>  [2,] 109.03871 138.66814 140.54371 122.70345  94.82353  81.58627  92.19126
#>  [3,] 114.51912 144.38580 140.61907 117.59940  89.42504  77.37245  89.88062
#>  [4,] 119.69849 149.86969 141.65819 116.18089  86.42802  70.78545  80.91061
#>  [5,] 123.39852 150.20424 135.77723 115.00683  95.57738  83.21740  90.10010
#>  [6,] 118.24031 142.39269 123.95641 108.19252 107.36936 110.17886 114.88237
#>  [7,] 106.44268 135.79886 126.15702 114.66052 123.10421 140.31913 145.86176
#>  [8,]  89.92300 121.57682 127.90371 130.62281 143.87567 163.47694 166.34804
#>  [9,]  68.57573  95.57818 110.48406 126.91594 144.33402 156.03290 152.61213
#> [10,]  42.80831  63.22445  75.58967  88.70860 100.87438 104.12723 101.09105
#>            [,8]      [,9]     [,10]
#>  [1,]  80.99665  94.29705  79.34519
#>  [2,] 114.58802 125.67056  98.69908
#>  [3,] 113.80357 121.78037  91.32244
#>  [4,] 101.39820 105.81504  79.32253
#>  [5,] 103.54218 104.83636  83.10779
#>  [6,] 119.79555 121.49073 101.96749
#>  [7,] 139.31889 135.34040 114.54538
#>  [8,] 146.79761 130.43125 107.87038
#>  [9,] 133.79037 116.77908  93.44768
#> [10,]  97.88208  92.60307  71.76658
pfocal_fast_abs_rectangle(data, height = 2)
#>            [,1]     [,2]     [,3]     [,4]      [,5]      [,6]      [,7]
#>  [1,] 0.9990285 1.185446 0.977859 1.532437 0.9892015 0.6252390 0.6581351
#>  [2,] 1.6667323 2.682006 2.589163 2.977016 2.0469831 1.1482914 1.3396793
#>  [3,] 1.4291856 2.372707 2.107584 2.163797 1.6958661 1.2094703 1.4541346
#>  [4,] 1.5839007 2.387607 2.146990 1.985745 1.1216422 0.9498045 0.8801682
#>  [5,] 1.4681367 2.993735 2.773813 2.535906 1.8981529 0.7689659 0.7426762
#>  [6,] 1.5509940 3.022694 1.801353 1.372341 2.0462061 1.4361177 1.1941862
#>  [7,] 1.3363922 2.717779 2.128962 1.047698 1.3889377 2.1347960 1.9979184
#>  [8,] 0.9907857 2.340682 2.644802 2.179396 1.6420469 2.0381030 2.9786287
#>  [9,] 1.0110885 1.861072 1.511736 2.121032 2.5315968 2.5350917 3.3237357
#> [10,] 0.6417836 1.331510 1.409377 1.656529 2.4988959 2.7176485 2.4227685
#>           [,8]      [,9]    [,10]
#>  [1,] 0.524332 0.8934082 1.575403
#>  [2,] 1.605192 1.9978386 2.564376
#>  [3,] 2.039242 2.4593876 2.446912
#>  [4,] 1.085023 1.8906349 1.963873
#>  [5,] 1.555764 1.5805860 1.434848
#>  [6,] 2.066133 1.6018964 1.730225
#>  [7,] 1.757457 1.7716390 2.640739
#>  [8,] 2.854409 2.1256823 2.826738
#>  [9,] 2.934234 1.2274045 1.666977
#> [10,] 2.245357 1.8555405 2.182386

pfocal_fast_separated(data, 
                      kernel_list = list(binomial_kernel(vertical_radius = 2, 
                                                         horizontal_radius = 2), 
                                         distance_kernel(vertical_radius = 2,
                                                         horizontal_radius = 2)))
#>           [,1]     [,2]     [,3]     [,4]     [,5]     [,6]     [,7]     [,8]
#>  [1,] 1895.069 2348.856 2930.536 2843.582 2682.052 2586.939 2548.329 2444.465
#>  [2,] 2506.854 3075.733 3850.340 3714.825 3464.772 3303.005 3205.474 3034.557
#>  [3,] 3373.926 4163.535 5223.390 5044.920 4686.541 4457.478 4322.908 4105.669
#>  [4,] 3536.906 4350.627 5534.447 5415.035 5076.956 4900.061 4829.551 4622.862
#>  [5,] 3489.944 4303.213 5570.313 5574.931 5301.267 5161.640 5146.361 4963.485
#>  [6,] 3398.971 4266.058 5638.197 5826.897 5667.080 5537.790 5502.210 5288.295
#>  [7,] 3162.829 4066.219 5504.348 5896.818 5905.817 5846.961 5823.392 5581.650
#>  [8,] 2755.037 3607.274 4958.325 5469.748 5618.582 5662.406 5716.911 5513.548
#>  [9,] 2039.018 2699.642 3729.026 4190.559 4381.305 4448.295 4497.970 4335.018
#> [10,] 1478.664 2028.101 2821.543 3236.287 3456.630 3539.381 3537.663 3342.804
#>           [,9]    [,10]
#>  [1,] 1880.828 1565.780
#>  [2,] 2322.151 1948.732
#>  [3,] 3174.324 2629.876
#>  [4,] 3578.480 2906.465
#>  [5,] 3808.516 3011.888
#>  [6,] 3987.717 3046.102
#>  [7,] 4155.342 3096.144
#>  [8,] 4101.549 3041.143
#>  [9,] 3204.481 2352.310
#> [10,] 2452.045 1768.730