Fast methods for common kernel computations

Methods wrapping pfocal to implement common kernel computations with default argument values.

pfocal_fast_gaussian_radius(
  data,
  vertical_radius,
  vertical_sd = 1,
  horizontal_radius = vertical_radius,
  horizontal_sd = vertical_sd,
  tail_included = TRUE,
  na.rm = NA,
  mp = TRUE,
  debug_use_r_implementation = FALSE,
  ...,
  transform_function = "MULTIPLY",
  reduce_function = "SUM",
  mean_divider = "ONE",
  variance = FALSE
)

pfocal_fast_gaussian_confidence(
  data,
  vertical_r0 = 0.05,
  vertical_sd = 1,
  horizontal_r0 = vertical_r0,
  horizontal_sd = vertical_sd,
  tail_included = TRUE,
  na.rm = NA,
  mp = TRUE,
  debug_use_r_implementation = FALSE,
  ...,
  transform_function = "MULTIPLY",
  reduce_function = "SUM",
  mean_divider = "ONE",
  variance = FALSE
)

pfocal_fast_binomial(
  data,
  vertical_radius,
  horizontal_radius = vertical_radius,
  na.rm = NA,
  mp = TRUE,
  debug_use_r_implementation = FALSE,
  ...,
  transform_function = "MULTIPLY",
  reduce_function = "SUM",
  mean_divider = "ONE",
  variance = FALSE
)

pfocal_fast_abs_rectangle(
  data,
  height,
  width = height,
  value = 1,
  na.rm = NA,
  mp = TRUE,
  debug_use_r_implementation = FALSE,
  ...,
  transform_function = "MULTIPLY",
  reduce_function = "SUM",
  mean_divider = "ONE",
  variance = FALSE
)

pfocal_fast_separated(
  data,
  kernel_list,
  na.rm = NA,
  mp = TRUE,
  debug_use_r_implementation = FALSE,
  ...,
  transform_function = "MULTIPLY",
  reduce_function = "SUM",
  mean_divider = "ONE",
  variance = FALSE
)

Arguments

data

[matrix-type] Grid to compute onto.

vertical_radius

[numeric] The kernel's radius in the vertical dimension.

vertical_sd

[numeric] The kernel's standard deviation in the vertical dimension.

horizontal_radius

[numeric] The kernel's radius in the horizontal dimension.

horizontal_sd

[numeric] The kernel's standard deviation in the horizontal dimension.

tail_included

[logical] Whether or not to include the kernel tail.

na.rm

[NA OR logical] The behavior to adopt for dealing with missing values, default to NA (faster option). For possible values see pfocal_info_nan_policy().

mp

[logical] Whether to use the open_mp implementation, default to TRUE.

debug_use_r_implementation

[logical] Used for debugging purposes whether to use the slow R implementation instead of the fast C++ code. Default to FALSE.

...

None used at the moment .

transform_function

[character] The function to apply to the cell values covered by the kernel. For possible values, see pfocal_info_transform(). Default to "MULTIPLY".

reduce_function

[character] The function to apply to the kernel values after the function passed in transform_function has been applied (the function that summarize the data). For possible values, see pfocal_info_reduce(). Default to "SUM".

mean_divider

[character] Optional, allows to specify how the final value at each cell is divided by a value that can be function of the intermediate data resulting of applying transform_function. For possible values, see pfocal_info_mean_divisor(). Default to "ONE" (for no division).

variance

[logical] Whether to return the "variance" of the intermediate values at each point (for more details please see pfocal_info_variance()). Default to FALSE (just returns the value at each point).

vertical_r0

[numeric] The kernel's r0 (exponential) in the vertical dimension.

horizontal_r0

[numeric] The kernel's r0 (exponential) in the horizontal dimension.

height

[numeric] For rectangular kernels, the height of the rectangle.

width

[numeric] For rectangular kernels, the width of the rectangle.

value

[numeric] For single value matrices, the value.

kernel_list

[list] A list of kernels computed from functions in kernel-gaussian, kernel-binomial, kernel-circular, kernel-distance, kernel-exponential.

Value

The updated, convoluted grid.

Examples

data <- matrix(nrow = 10, ncol = 10, data = runif(10 * 10))

pfocal_fast_gaussian_radius(data, vertical_radius = 2)
#>            [,1]      [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
#>  [1,] 0.1709732 0.2301307 0.2725320 0.3129039 0.3455893 0.3546369 0.3290118
#>  [2,] 0.2316487 0.3187797 0.3656632 0.4079228 0.4453538 0.4447916 0.4107196
#>  [3,] 0.2567740 0.3635839 0.4149730 0.4577070 0.5040395 0.5055137 0.4563211
#>  [4,] 0.2775961 0.4184333 0.4751572 0.5046812 0.5472610 0.5505397 0.5125150
#>  [5,] 0.2804580 0.4381513 0.4983050 0.5236120 0.5648210 0.5553105 0.4991230
#>  [6,] 0.2786010 0.4384737 0.4905476 0.5042279 0.5238277 0.5085580 0.4780201
#>  [7,] 0.2577783 0.4387308 0.5193777 0.5329806 0.5461537 0.5422166 0.5111694
#>  [8,] 0.2424864 0.4017288 0.4722877 0.4844405 0.4806438 0.4641714 0.4455710
#>  [9,] 0.1955412 0.3311992 0.4144784 0.4379967 0.4275216 0.4177345 0.4005157
#> [10,] 0.1453726 0.2604431 0.3463762 0.3698248 0.3505067 0.3425263 0.3392454
#>            [,8]      [,9]     [,10]
#>  [1,] 0.3140656 0.3098233 0.2295248
#>  [2,] 0.3921202 0.3826800 0.2886585
#>  [3,] 0.4220423 0.4017572 0.2938419
#>  [4,] 0.4762790 0.4327627 0.3118992
#>  [5,] 0.4562503 0.4144203 0.3045998
#>  [6,] 0.4577269 0.3957491 0.2685479
#>  [7,] 0.4704354 0.3991885 0.2839345
#>  [8,] 0.4294974 0.3726640 0.2678653
#>  [9,] 0.3730757 0.3134129 0.2191924
#> [10,] 0.3075612 0.2437110 0.1739717
pfocal_fast_gaussian_confidence(data)
#>            [,1]      [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
#>  [1,] 0.2075059 0.2612073 0.2775074 0.3054024 0.3487447 0.3747053 0.3598678
#>  [2,] 0.2793739 0.3575511 0.4171127 0.4862837 0.5337311 0.5311876 0.4844920
#>  [3,] 0.2914323 0.4051012 0.5044258 0.5920165 0.6303092 0.6043703 0.5413345
#>  [4,] 0.2775593 0.4347707 0.5253114 0.5771188 0.6220385 0.6144228 0.5560351
#>  [5,] 0.2766879 0.4447179 0.4706896 0.4612147 0.5141997 0.5263585 0.4796168
#>  [6,] 0.2641786 0.4285621 0.4340103 0.4037193 0.4448485 0.4580172 0.4261550
#>  [7,] 0.2654074 0.4459144 0.5036447 0.5032252 0.5136492 0.4972501 0.4733118
#>  [8,] 0.2824704 0.4609073 0.5638748 0.6031063 0.5863458 0.5411494 0.5143405
#>  [9,] 0.2436066 0.4028670 0.5252170 0.5727471 0.5403458 0.5149935 0.5053130
#> [10,] 0.1514369 0.2846956 0.4007478 0.4308953 0.4055388 0.4136133 0.4088460
#>            [,8]      [,9]     [,10]
#>  [1,] 0.3817335 0.4228037 0.3322938
#>  [2,] 0.4826253 0.5108131 0.3930753
#>  [3,] 0.4841583 0.4328279 0.3014716
#>  [4,] 0.4640396 0.3635934 0.2451580
#>  [5,] 0.4283846 0.3771642 0.2799376
#>  [6,] 0.4228230 0.4086565 0.3012773
#>  [7,] 0.4750825 0.4299442 0.3028068
#>  [8,] 0.5031871 0.4212757 0.2847015
#>  [9,] 0.4701234 0.3704280 0.2480370
#> [10,] 0.3452714 0.2671136 0.2011167
pfocal_fast_binomial(data, vertical_radius = 2)
#>           [,1]      [,2]      [,3]     [,4]      [,5]      [,6]      [,7]
#>  [1,] 52.41195  66.79308  70.69879  77.6858  89.24113  94.38369  91.42431
#>  [2,] 70.97730  92.78485 107.63574 125.2566 137.98845 136.11414 124.80992
#>  [3,] 73.69558 104.91191 130.50245 153.1247 163.38139 156.86731 140.28952
#>  [4,] 71.57325 111.83116 134.77477 148.5744 158.86929 157.30531 141.53029
#>  [5,] 70.90426 112.45345 120.45400 118.8821 130.35668 134.98056 123.34051
#>  [6,] 68.54317 109.45256 112.64892 105.1502 113.52524 117.57428 110.63175
#>  [7,] 69.09068 113.87383 129.27710 129.1922 130.52634 126.35360 120.82967
#>  [8,] 71.85885 117.87322 145.52149 154.8657 149.96811 138.52484 133.22732
#>  [9,] 62.20358 104.84994 137.33812 148.6793 141.40434 133.82082 131.42060
#> [10,] 38.43699  72.87841 101.82533 109.2830 103.91806 104.62578 103.43943
#>            [,8]      [,9]     [,10]
#>  [1,]  98.28808 107.83598  85.24002
#>  [2,] 125.80244 130.69712 100.31922
#>  [3,] 126.16063 111.55566  77.75997
#>  [4,] 118.63976  93.51596  62.35606
#>  [5,] 109.58789  96.04853  69.73953
#>  [6,] 109.13806 105.23781  77.55386
#>  [7,] 120.90034 110.92953  77.47411
#>  [8,] 129.23509 109.34571  72.29655
#>  [9,] 121.02588  96.68286  63.91739
#> [10,]  88.03698  68.63284  49.72392
pfocal_fast_abs_rectangle(data, height = 2)
#>            [,1]      [,2]      [,3]      [,4]      [,5]     [,6]     [,7]
#>  [1,] 0.4377419 0.9911051 0.7894864 0.6394339 0.6736271 1.005630 1.097434
#>  [2,] 0.9610629 1.5799447 1.4272021 1.6309432 1.7406509 2.385555 2.347660
#>  [3,] 1.2237407 1.6443577 1.4289010 2.2523681 2.6710533 2.501116 2.076676
#>  [4,] 0.8457801 1.4069785 1.8927310 2.6929967 2.9573380 2.795094 2.655198
#>  [5,] 0.4127549 1.5958756 2.5668551 1.9961468 2.2803842 2.783424 2.483469
#>  [6,] 0.4448807 2.0889031 2.2074454 0.7963540 1.2549500 2.056627 1.463062
#>  [7,] 0.1953110 1.4835299 2.0120996 1.2512184 1.2974607 1.791586 1.637271
#>  [8,] 0.4659438 1.7749375 2.6555274 2.6070311 2.7424489 2.663286 1.980634
#>  [9,] 0.8073377 1.8917678 1.9041894 2.6549766 2.7245627 2.280434 1.780720
#> [10,] 0.4261050 1.1767270 1.8627012 2.7517463 2.1338007 1.825182 2.417113
#>            [,8]     [,9]     [,10]
#>  [1,] 0.9408039 1.422511 1.4625152
#>  [2,] 1.6052729 2.494310 3.2251894
#>  [3,] 1.5716872 2.133542 2.5312451
#>  [4,] 2.3172488 1.599398 0.9617833
#>  [5,] 2.4359369 1.334470 1.2227483
#>  [6,] 1.0877836 1.677538 2.0040359
#>  [7,] 1.5146553 1.992868 1.7196169
#>  [8,] 2.3815972 2.124853 1.6930760
#>  [9,] 2.0802857 2.234315 1.1951996
#> [10,] 2.4331124 1.693431 1.2205389

pfocal_fast_separated(data, 
                      kernel_list = list(binomial_kernel(vertical_radius = 2, 
                                                         horizontal_radius = 2), 
                                         distance_kernel(vertical_radius = 2,
                                                         horizontal_radius = 2)))
#>           [,1]     [,2]     [,3]     [,4]     [,5]     [,6]     [,7]     [,8]
#>  [1,] 1402.164 2003.473 2819.587 3170.566 3309.544 3384.467 3373.929 3053.423
#>  [2,] 1901.300 2685.126 3756.907 4236.784 4411.865 4461.233 4373.144 3901.230
#>  [3,] 2527.986 3489.430 4826.334 5434.883 5627.965 5684.970 5630.061 5118.117
#>  [4,] 2748.854 3749.202 5158.480 5818.213 5974.775 5962.732 5853.719 5303.505
#>  [5,] 2869.236 3902.065 5350.663 6066.598 6204.234 6108.850 5882.490 5237.428
#>  [6,] 2946.532 3988.839 5427.717 6151.523 6282.591 6154.392 5862.700 5169.503
#>  [7,] 2892.871 3894.621 5237.730 5913.141 6064.307 5963.571 5682.358 5034.029
#>  [8,] 2673.992 3619.868 4820.686 5457.475 5667.287 5588.678 5293.703 4704.702
#>  [9,] 2071.132 2820.811 3732.422 4218.106 4413.500 4347.747 4061.979 3580.338
#> [10,] 1612.997 2238.180 2946.362 3332.986 3510.719 3438.582 3160.606 2763.714
#>           [,9]    [,10]
#>  [1,] 2171.832 1632.670
#>  [2,] 2753.255 2026.835
#>  [3,] 3706.604 2744.048
#>  [4,] 3864.028 2846.866
#>  [5,] 3791.689 2752.295
#>  [6,] 3733.278 2702.840
#>  [7,] 3669.504 2689.615
#>  [8,] 3460.465 2544.817
#>  [9,] 2641.994 1943.242
#> [10,] 2042.891 1470.606