Fast methods for common kernel computations

Methods wrapping pfocal to implement common kernel computations with default argument values.

pfocal_fast_gaussian_radius(
  data,
  vertical_radius,
  vertical_sd = 1,
  horizontal_radius = vertical_radius,
  horizontal_sd = vertical_sd,
  tail_included = TRUE,
  na.rm = NA,
  mp = TRUE,
  debug_use_r_implementation = FALSE,
  ...,
  transform_function = "MULTIPLY",
  reduce_function = "SUM",
  mean_divider = "ONE",
  variance = FALSE
)

pfocal_fast_gaussian_confidence(
  data,
  vertical_r0 = 0.05,
  vertical_sd = 1,
  horizontal_r0 = vertical_r0,
  horizontal_sd = vertical_sd,
  tail_included = TRUE,
  na.rm = NA,
  mp = TRUE,
  debug_use_r_implementation = FALSE,
  ...,
  transform_function = "MULTIPLY",
  reduce_function = "SUM",
  mean_divider = "ONE",
  variance = FALSE
)

pfocal_fast_binomial(
  data,
  vertical_radius,
  horizontal_radius = vertical_radius,
  na.rm = NA,
  mp = TRUE,
  debug_use_r_implementation = FALSE,
  ...,
  transform_function = "MULTIPLY",
  reduce_function = "SUM",
  mean_divider = "ONE",
  variance = FALSE
)

pfocal_fast_abs_rectangle(
  data,
  height,
  width = height,
  value = 1,
  na.rm = NA,
  mp = TRUE,
  debug_use_r_implementation = FALSE,
  ...,
  transform_function = "MULTIPLY",
  reduce_function = "SUM",
  mean_divider = "ONE",
  variance = FALSE
)

pfocal_fast_separated(
  data,
  kernel_list,
  na.rm = NA,
  mp = TRUE,
  debug_use_r_implementation = FALSE,
  ...,
  transform_function = "MULTIPLY",
  reduce_function = "SUM",
  mean_divider = "ONE",
  variance = FALSE
)

Arguments

data

[matrix-type] Grid to compute onto.

vertical_radius

[numeric] The kernel's radius in the vertical dimension.

vertical_sd

[numeric] The kernel's standard deviation in the vertical dimension.

horizontal_radius

[numeric] The kernel's radius in the horizontal dimension.

horizontal_sd

[numeric] The kernel's standard deviation in the horizontal dimension.

tail_included

[logical] Whether or not to include the kernel tail.

na.rm

[NA OR logical] The behavior to adopt for dealing with missing values, default to NA (faster option). For possible values see pfocal_info_nan_policy().

mp

[logical] Whether to use the open_mp implementation, default to TRUE.

debug_use_r_implementation

[logical] Used for debugging purposes whether to use the slow R implementation instead of the fass C++ code. Default to FALSE.

...

None used at the moment .

transform_function

[character] The function to apply to the cell values covered by the kernel. For possible values, see pfocal_info_transform(). Default to "MULTIPLY".

reduce_function

[character] The function to apply to the kernel values after the function passed in transform_function has been applied (the function that summarize the data). For possible values, see pfocal_info_reduce(). Default to "SUM".

mean_divider

[character] Optional, allows to specify how the final value at each cell is divided by a value that can be function of the intermediate data resulting of applying transform_function. For possible values, see pfocal_info_mean_divisor(). Default to "ONE" (for no division).

variance

[logical] Whether to return the "variance" of the intermediate values at each point (for more details please see pfocal_info_variance()). Default to FALSE (just returns the value at each point).

vertical_r0

[numeric] The kernel's r0 (exponential) in the vertical dimension.

horizontal_r0

[numeric] The kernel's r0 (exponential) in the horizontal dimension.

height

[numeric] For rectangular kernels, the height of the rectangle.

width

[numeric] For rectangular kernels, the width of the rectangle.

value

[numeric] For single value matrices, the value.

kernel_list

[list] A list of kernels computed from functions in kernel-gaussian, kernel-binomial, kernel-circular, kernel-distance, kernel-exponential.

Value

The updated, convoluted grid.

Examples

data <- matrix(nrow = 10, ncol = 10, data = runif(10 * 10))

pfocal_fast_gaussian_radius(data, vertical_radius = 2)
#>            [,1]      [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
#>  [1,] 0.2157772 0.2576607 0.2758016 0.2877784 0.2826570 0.3153037 0.3491962
#>  [2,] 0.2739120 0.3503491 0.3918486 0.4064490 0.3839098 0.4027693 0.4379059
#>  [3,] 0.3069422 0.4115794 0.4708619 0.4877071 0.4470343 0.4403682 0.4667624
#>  [4,] 0.3091923 0.4472581 0.5384275 0.5526873 0.4854108 0.4558892 0.4744148
#>  [5,] 0.3244631 0.4983848 0.6002832 0.5949754 0.5025896 0.4551345 0.4638962
#>  [6,] 0.3041929 0.4788062 0.5656587 0.5519180 0.4623912 0.4078460 0.4161103
#>  [7,] 0.2839656 0.4713575 0.5699700 0.5554160 0.4766544 0.4390897 0.4443598
#>  [8,] 0.2643624 0.4319086 0.5038499 0.4864292 0.4280980 0.4052460 0.4152979
#>  [9,] 0.2270262 0.3598779 0.4080102 0.4014762 0.3774366 0.3726109 0.3825958
#> [10,] 0.1617869 0.2624708 0.3049147 0.3079269 0.3046991 0.3120505 0.3209847
#>            [,8]      [,9]     [,10]
#>  [1,] 0.3460077 0.3153649 0.2164690
#>  [2,] 0.4471698 0.4093345 0.2753906
#>  [3,] 0.4836576 0.4417241 0.2979818
#>  [4,] 0.5030306 0.4700466 0.3211840
#>  [5,] 0.4874287 0.4482536 0.3132505
#>  [6,] 0.4456744 0.4121496 0.2911220
#>  [7,] 0.4606896 0.4220245 0.3034210
#>  [8,] 0.4257848 0.3869931 0.2786754
#>  [9,] 0.3831961 0.3419689 0.2471365
#> [10,] 0.3225108 0.2894805 0.2057452
pfocal_fast_gaussian_confidence(data)
#>            [,1]      [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
#>  [1,] 0.2417516 0.2751230 0.2854109 0.2981658 0.3020527 0.3683479 0.4205045
#>  [2,] 0.3734004 0.4462536 0.4702719 0.4932306 0.4846830 0.5289898 0.5762967
#>  [3,] 0.3832656 0.4791412 0.5252065 0.5564964 0.5316893 0.5279403 0.5609940
#>  [4,] 0.3122604 0.4561670 0.5536006 0.5641929 0.4973249 0.4610144 0.4809379
#>  [5,] 0.2642885 0.4580200 0.5955660 0.5926637 0.4815068 0.4092418 0.4090718
#>  [6,] 0.2622437 0.4657486 0.6011978 0.5994304 0.4724477 0.3767343 0.3597807
#>  [7,] 0.3026274 0.5003495 0.6015512 0.5762301 0.4597739 0.3942736 0.3905454
#>  [8,] 0.3475744 0.5298032 0.5690043 0.5131424 0.4447574 0.4430959 0.4714870
#>  [9,] 0.2888983 0.4409644 0.4640745 0.4381310 0.4394641 0.4703604 0.5006944
#> [10,] 0.1547330 0.2597463 0.3151555 0.3503583 0.3888859 0.4114143 0.4183837
#>            [,8]      [,9]     [,10]
#>  [1,] 0.4060313 0.3756170 0.2586799
#>  [2,] 0.5652456 0.5195467 0.3619376
#>  [3,] 0.5779069 0.5230260 0.3552199
#>  [4,] 0.5253097 0.4732007 0.3078760
#>  [5,] 0.4618334 0.4297109 0.2802954
#>  [6,] 0.3954063 0.3869883 0.2789520
#>  [7,] 0.3980376 0.3814955 0.2996302
#>  [8,] 0.4623760 0.4050532 0.3081377
#>  [9,] 0.4990384 0.4297452 0.3016210
#> [10,] 0.4240953 0.3763302 0.2511452
pfocal_fast_binomial(data, vertical_radius = 2)
#>           [,1]      [,2]      [,3]      [,4]      [,5]      [,6]     [,7]
#>  [1,] 61.20541  72.03752  74.11354  75.38081  78.92200  94.29723 106.9832
#>  [2,] 94.40552 114.96094 120.46898 124.06256 125.18021 135.42221 147.1980
#>  [3,] 97.25156 124.87927 136.92819 142.32455 137.48665 136.23147 144.9267
#>  [4,] 79.63059 118.71744 143.08836 145.11145 128.92689 117.70410 124.3549
#>  [5,] 66.53141 117.22157 152.37122 151.13540 123.69360 102.55845 104.4197
#>  [6,] 66.27769 119.87502 156.21578 154.01600 122.16994  95.51860  92.1102
#>  [7,] 78.21628 129.64631 155.90428 147.05403 118.28798 100.14609  99.4671
#>  [8,] 88.87704 135.83622 146.95401 131.20092 114.30148 113.26841 120.4901
#>  [9,] 74.55712 113.33763 120.91740 113.83271 113.88370 122.12449 130.5481
#> [10,] 40.05806  66.36137  80.30926  88.58984  98.03692 104.09121 107.0164
#>           [,8]      [,9]    [,10]
#>  [1,] 105.9519  95.66445 66.14668
#>  [2,] 147.2708 133.21385 92.47461
#>  [3,] 150.1828 135.36887 91.48528
#>  [4,] 135.2168 122.29718 79.12239
#>  [5,] 117.1497 109.65810 71.57629
#>  [6,] 101.4934  99.80722 71.65012
#>  [7,] 102.0926  98.01424 75.69275
#>  [8,] 118.6863 105.29781 78.38693
#>  [9,] 129.6212 112.39363 77.59289
#> [10,] 107.9660  95.68166 63.42741
pfocal_fast_abs_rectangle(data, height = 2)
#>            [,1]      [,2]      [,3]      [,4]      [,5]      [,6]     [,7]
#>  [1,] 0.5850563 0.6196274 0.1891715 0.5437439 0.4781283 0.4680693 1.283681
#>  [2,] 1.3164488 2.0328339 1.5971515 2.2237187 1.6382181 1.6169807 3.055330
#>  [3,] 1.5928337 2.7126518 1.9442887 2.5179018 2.7238244 2.1429441 2.671803
#>  [4,] 1.2547524 1.9173749 1.5686840 2.1577657 2.2056268 1.8437760 1.704248
#>  [5,] 0.5134188 1.2804215 2.3876061 3.0748930 1.7504122 1.7159359 1.567371
#>  [6,] 0.3509019 1.5968144 2.3641208 3.0129115 2.3753534 1.5238413 1.271793
#>  [7,] 0.3443218 1.4610806 2.1009427 2.8034374 2.3298645 1.2352743 1.113727
#>  [8,] 0.7794800 2.1139467 2.6550221 2.9548982 2.0316921 1.4551105 1.928710
#>  [9,] 0.8948451 2.6950103 2.6820852 1.7482220 1.3766287 1.5265542 1.978126
#> [10,] 0.3191192 1.4082679 1.8168687 1.5609441 1.8943351 2.3228024 2.136803
#>           [,8]     [,9]    [,10]
#>  [1,] 1.064605 1.058149 1.122517
#>  [2,] 2.379535 2.299287 2.532359
#>  [3,] 2.631156 2.555527 2.524248
#>  [4,] 2.272806 2.367365 1.730799
#>  [5,] 1.831270 2.545009 1.527411
#>  [6,] 1.480299 2.169915 1.500719
#>  [7,] 1.036594 1.370454 1.618973
#>  [8,] 1.935781 1.584574 1.748986
#>  [9,] 2.415675 1.739561 1.347770
#> [10,] 2.305965 2.540282 1.965858

pfocal_fast_separated(data, 
                      kernel_list = list(binomial_kernel(vertical_radius = 2, 
                                                         horizontal_radius = 2), 
                                         distance_kernel(vertical_radius = 2,
                                                         horizontal_radius = 2)))
#>           [,1]     [,2]     [,3]     [,4]     [,5]     [,6]     [,7]     [,8]
#>  [1,] 1607.659 2158.287 2912.463 3153.834 3317.590 3407.208 3376.242 3148.167
#>  [2,] 2094.178 2815.157 3779.832 4091.593 4282.288 4362.066 4264.975 3921.952
#>  [3,] 2811.841 3784.516 4985.785 5391.054 5622.013 5688.699 5516.374 5037.669
#>  [4,] 3187.011 4286.176 5532.261 5937.469 6108.397 6068.911 5762.359 5176.457
#>  [5,] 3322.885 4432.949 5617.902 5997.390 6102.921 5960.502 5562.284 4947.526
#>  [6,] 3342.081 4400.822 5526.345 5909.148 5975.439 5768.555 5352.627 4773.101
#>  [7,] 3206.848 4207.010 5339.208 5780.698 5853.952 5643.445 5265.655 4735.300
#>  [8,] 2843.243 3757.442 4879.132 5378.351 5489.585 5327.467 5021.562 4551.378
#>  [9,] 2165.529 2819.279 3677.535 4042.537 4097.978 3960.911 3746.890 3421.251
#> [10,] 1629.342 2111.964 2760.389 3042.813 3108.903 3068.682 2972.804 2741.843
#>           [,9]    [,10]
#>  [1,] 2408.513 1828.189
#>  [2,] 2989.217 2299.013
#>  [3,] 3851.827 2965.928
#>  [4,] 3940.456 3052.998
#>  [5,] 3747.626 2907.892
#>  [6,] 3606.929 2776.080
#>  [7,] 3574.821 2727.470
#>  [8,] 3431.863 2611.368
#>  [9,] 2574.004 1963.802
#> [10,] 2074.842 1565.920