Fast methods for common kernel computations

Methods wrapping pfocal to implement common kernel computations with default argument values.

pfocal_fast_gaussian_radius(
  data,
  vertical_radius,
  vertical_sd = 1,
  horizontal_radius = vertical_radius,
  horizontal_sd = vertical_sd,
  tail_included = TRUE,
  na.rm = NA,
  mp = TRUE,
  debug_use_r_implementation = FALSE,
  ...,
  transform_function = "MULTIPLY",
  reduce_function = "SUM",
  mean_divider = "ONE",
  variance = FALSE
)

pfocal_fast_gaussian_confidence(
  data,
  vertical_r0 = 0.05,
  vertical_sd = 1,
  horizontal_r0 = vertical_r0,
  horizontal_sd = vertical_sd,
  tail_included = TRUE,
  na.rm = NA,
  mp = TRUE,
  debug_use_r_implementation = FALSE,
  ...,
  transform_function = "MULTIPLY",
  reduce_function = "SUM",
  mean_divider = "ONE",
  variance = FALSE
)

pfocal_fast_binomial(
  data,
  vertical_radius,
  horizontal_radius = vertical_radius,
  na.rm = NA,
  mp = TRUE,
  debug_use_r_implementation = FALSE,
  ...,
  transform_function = "MULTIPLY",
  reduce_function = "SUM",
  mean_divider = "ONE",
  variance = FALSE
)

pfocal_fast_abs_rectangle(
  data,
  height,
  width = height,
  value = 1,
  na.rm = NA,
  mp = TRUE,
  debug_use_r_implementation = FALSE,
  ...,
  transform_function = "MULTIPLY",
  reduce_function = "SUM",
  mean_divider = "ONE",
  variance = FALSE
)

pfocal_fast_separated(
  data,
  kernel_list,
  na.rm = NA,
  mp = TRUE,
  debug_use_r_implementation = FALSE,
  ...,
  transform_function = "MULTIPLY",
  reduce_function = "SUM",
  mean_divider = "ONE",
  variance = FALSE
)

Arguments

data

[matrix-type] Grid to compute onto.

vertical_radius

[numeric] The kernel's radius in the vertical dimension.

vertical_sd

[numeric] The kernel's standard deviation in the vertical dimension.

horizontal_radius

[numeric] The kernel's radius in the horizontal dimension.

horizontal_sd

[numeric] The kernel's standard deviation in the horizontal dimension.

tail_included

[logical] Whether or not to include the kernel tail.

na.rm

[NA OR logical] The behavior to adopt for dealing with missing values, default to NA (faster option). For possible values see pfocal_info_nan_policy().

mp

[logical] Whether to use the open_mp implementation, default to TRUE.

debug_use_r_implementation

[logical] Used for debugging purposes whether to use the slow R implementation instead of the fast C++ code. Default to FALSE.

...

None used at the moment .

transform_function

[character] The function to apply to the cell values covered by the kernel. For possible values, see pfocal_info_transform(). Default to "MULTIPLY".

reduce_function

[character] The function to apply to the kernel values after the function passed in transform_function has been applied (the function that summarize the data). For possible values, see pfocal_info_reduce(). Default to "SUM".

mean_divider

[character] Optional, allows to specify how the final value at each cell is divided by a value that can be function of the intermediate data resulting of applying transform_function. For possible values, see pfocal_info_mean_divisor(). Default to "ONE" (for no division).

variance

[logical] Whether to return the "variance" of the intermediate values at each point (for more details please see pfocal_info_variance()). Default to FALSE (just returns the value at each point).

vertical_r0

[numeric] The kernel's r0 (exponential) in the vertical dimension.

horizontal_r0

[numeric] The kernel's r0 (exponential) in the horizontal dimension.

height

[numeric] For rectangular kernels, the height of the rectangle.

width

[numeric] For rectangular kernels, the width of the rectangle.

value

[numeric] For single value matrices, the value.

kernel_list

[list] A list of kernels computed from functions in kernel-gaussian, kernel-binomial, kernel-circular, kernel-distance, kernel-exponential.

Value

The updated, convoluted grid.

Examples

data <- matrix(nrow = 10, ncol = 10, data = runif(10 * 10))

pfocal_fast_gaussian_radius(data, vertical_radius = 2)
#>            [,1]      [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
#>  [1,] 0.1731277 0.2024456 0.1997931 0.2209819 0.2394558 0.2686606 0.3171553
#>  [2,] 0.2500151 0.2768376 0.2572098 0.2863179 0.3302371 0.3828722 0.4411662
#>  [3,] 0.2775013 0.3198914 0.3016764 0.3310575 0.3991239 0.4821537 0.5496857
#>  [4,] 0.3138352 0.3578227 0.3300180 0.3614232 0.4530352 0.5493929 0.6002153
#>  [5,] 0.3520904 0.4222842 0.3936887 0.4147878 0.5097293 0.5982542 0.6171494
#>  [6,] 0.3663098 0.4446928 0.4181743 0.4202241 0.5114148 0.5933302 0.5828621
#>  [7,] 0.3835868 0.4781289 0.4646712 0.4670487 0.5388899 0.5722926 0.5206039
#>  [8,] 0.3861611 0.4761926 0.4627448 0.4580608 0.5073881 0.5069389 0.4301415
#>  [9,] 0.3296629 0.4113348 0.4142147 0.4137515 0.4418269 0.4160584 0.3426287
#> [10,] 0.2638915 0.3080223 0.3067465 0.3182921 0.3365461 0.2948354 0.2312212
#>            [,8]      [,9]     [,10]
#>  [1,] 0.3379018 0.2975732 0.1852242
#>  [2,] 0.4596850 0.4083543 0.2620380
#>  [3,] 0.5595310 0.5008184 0.3336244
#>  [4,] 0.5954459 0.5382720 0.3652605
#>  [5,] 0.5925031 0.5375851 0.3642534
#>  [6,] 0.5526948 0.5258481 0.3779072
#>  [7,] 0.4979263 0.4863488 0.3387693
#>  [8,] 0.4128838 0.4218346 0.2979490
#>  [9,] 0.3460119 0.3667989 0.2599727
#> [10,] 0.2509257 0.2756654 0.1922058
pfocal_fast_gaussian_confidence(data)
#>            [,1]      [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
#>  [1,] 0.2164723 0.2762295 0.2819881 0.2877057 0.2788485 0.2826814 0.3071542
#>  [2,] 0.3140658 0.3589555 0.3492355 0.3994991 0.4285335 0.4478090 0.5060479
#>  [3,] 0.3041553 0.3090648 0.2801504 0.3478091 0.4322562 0.5216709 0.6284974
#>  [4,] 0.2842794 0.2923268 0.2502488 0.2884531 0.4037764 0.5512117 0.6689598
#>  [5,] 0.3285104 0.3903248 0.3493016 0.3517372 0.4737325 0.6187849 0.6658596
#>  [6,] 0.3589178 0.4701979 0.4502531 0.4476221 0.5595466 0.6543834 0.6078894
#>  [7,] 0.3999971 0.5207140 0.5078780 0.5089300 0.5874035 0.6123996 0.5092176
#>  [8,] 0.4513213 0.5696082 0.5582825 0.5386470 0.5675000 0.5442384 0.4448806
#>  [9,] 0.4265720 0.5124745 0.5119546 0.5003541 0.5088613 0.4531663 0.3766994
#> [10,] 0.3011082 0.3391013 0.3477524 0.3718398 0.3866192 0.3184955 0.2492944
#>            [,8]      [,9]     [,10]
#>  [1,] 0.3219673 0.2855630 0.1640850
#>  [2,] 0.5408469 0.4800524 0.2950904
#>  [3,] 0.6634202 0.5900131 0.3961031
#>  [4,] 0.6864343 0.6076875 0.4169789
#>  [5,] 0.6367248 0.5770109 0.4006601
#>  [6,] 0.5363806 0.5078353 0.3634600
#>  [7,] 0.4362728 0.4270972 0.3039144
#>  [8,] 0.4208402 0.4310603 0.3036084
#>  [9,] 0.4222013 0.4619696 0.3283181
#> [10,] 0.3053246 0.3499603 0.2426101
pfocal_fast_binomial(data, vertical_radius = 2)
#>            [,1]      [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
#>  [1,]  55.41968  71.32250  73.48133  73.67119  72.18240  72.46580  78.40514
#>  [2,]  79.07289  91.57154  90.66480 100.55802 109.16969 115.36096 129.73391
#>  [3,]  76.97707  79.76279  73.18789  87.93788 110.60143 134.58664 161.21093
#>  [4,]  72.94144  76.14863  64.94896  73.44083 103.98892 142.92219 172.43801
#>  [5,]  82.35245  98.76445  88.23064  88.82798 120.45045 158.17928 171.34737
#>  [6,]  92.15541 120.43947 114.98263 114.88114 143.54723 166.78984 155.76532
#>  [7,] 102.91849 135.10023 132.16568 131.88810 151.41478 156.85361 131.90192
#>  [8,] 115.06692 146.66023 143.38553 139.10472 145.88441 138.37203 114.94320
#>  [9,] 109.10976 133.50835 131.82752 130.38019 131.11452 115.46058  97.66416
#> [10,]  75.84235  87.99207  88.84805  95.46337  97.60096  80.01695  64.78749
#>            [,8]      [,9]     [,10]
#>  [1,]  83.28828  72.64007  42.19239
#>  [2,] 139.80993 123.17806  76.24087
#>  [3,] 172.39295 152.45972 100.88362
#>  [4,] 178.31134 157.24682 106.96633
#>  [5,] 165.49533 148.04505 102.80119
#>  [6,] 138.62919 128.05409  91.94570
#>  [7,] 113.55399 108.30154  78.28250
#>  [8,] 109.18826 110.29950  79.03383
#>  [9,] 108.04991 117.28530  84.12454
#> [10,]  78.34781  89.26938  63.13442
pfocal_fast_abs_rectangle(data, height = 2)
#>            [,1]      [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
#>  [1,] 0.3283139 0.6634403 0.8355967 0.8140377 0.3797587 0.4806929 0.8567592
#>  [2,] 0.9114312 2.0612936 1.6674945 1.7751863 2.0770283 1.4819789 1.6074078
#>  [3,] 1.4215151 2.3274285 0.9832693 1.4315631 2.5340071 1.9798177 2.2740634
#>  [4,] 1.0202542 1.1526838 0.2959147 0.5915731 1.1406631 1.4699714 2.5739128
#>  [5,] 1.0987746 1.6609877 0.9776277 0.6468019 1.0020217 1.8404863 2.6340166
#>  [6,] 1.0813142 2.3064159 2.3332138 1.5061146 1.5422987 2.9210392 3.4284924
#>  [7,] 0.7941636 1.9473403 2.0532498 1.5076742 1.9357191 2.9897196 2.7701810
#>  [8,] 1.3412012 2.7632568 2.2626237 1.9294291 2.3093263 2.8197374 2.2858869
#>  [9,] 1.6088438 2.7402375 2.8135632 2.3680224 1.8702633 2.4448618 1.8067242
#> [10,] 1.7952420 2.1078062 1.7696007 1.9695393 2.0824759 2.4901979 0.9747890
#>            [,8]      [,9]     [,10]
#>  [1,] 0.5323169 0.6273073 0.6326324
#>  [2,] 1.8567524 2.3064299 1.4912518
#>  [3,] 2.7910358 2.8616716 2.3866821
#>  [4,] 3.2983317 2.7189462 2.5946992
#>  [5,] 2.9871703 2.7986822 2.2952262
#>  [6,] 2.2945645 2.3858707 2.7009390
#>  [7,] 1.4987360 1.8088196 2.2094281
#>  [8,] 0.9520927 1.4848042 1.5267762
#>  [9,] 1.4657844 2.2921376 2.2258762
#> [10,] 1.1731429 2.6848965 2.4479152

pfocal_fast_separated(data, 
                      kernel_list = list(binomial_kernel(vertical_radius = 2, 
                                                         horizontal_radius = 2), 
                                         distance_kernel(vertical_radius = 2,
                                                         horizontal_radius = 2)))
#>           [,1]     [,2]     [,3]     [,4]     [,5]     [,6]     [,7]     [,8]
#>  [1,] 1168.908 1576.745 2235.416 2472.235 2739.024 3068.498 3208.443 3018.463
#>  [2,] 1446.787 1946.318 2827.091 3186.310 3587.839 4073.162 4328.004 4132.527
#>  [3,] 2076.569 2787.344 3999.922 4511.361 4967.620 5493.736 5794.699 5525.815
#>  [4,] 2376.981 3188.262 4606.449 5246.239 5698.024 6185.252 6506.639 6203.527
#>  [5,] 2541.746 3366.803 4872.767 5564.307 5931.384 6308.905 6622.490 6304.708
#>  [6,] 2822.723 3671.038 5220.851 5852.734 6029.704 6230.478 6483.255 6118.641
#>  [7,] 3086.839 3997.082 5584.230 6099.730 6068.584 6098.628 6246.310 5787.492
#>  [8,] 2994.219 3908.157 5429.611 5807.343 5641.165 5570.567 5599.093 5047.191
#>  [9,] 2320.962 3028.175 4192.183 4404.899 4205.439 4107.837 4067.847 3573.581
#> [10,] 1822.014 2386.087 3236.365 3312.645 3117.586 3044.410 3009.346 2635.631
#>           [,9]    [,10]
#>  [1,] 2373.063 1832.560
#>  [2,] 3240.466 2505.822
#>  [3,] 4289.075 3275.495
#>  [4,] 4748.845 3583.763
#>  [5,] 4750.298 3540.317
#>  [6,] 4533.972 3356.322
#>  [7,] 4212.770 3138.460
#>  [8,] 3590.504 2715.446
#>  [9,] 2473.871 1897.044
#> [10,] 1845.087 1451.099