Fast methods for common kernel computations

Methods wrapping pfocal to implement common kernel computations with default argument values.

pfocal_fast_gaussian_radius(
  data,
  vertical_radius,
  vertical_sd = 1,
  horizontal_radius = vertical_radius,
  horizontal_sd = vertical_sd,
  tail_included = TRUE,
  na.rm = NA,
  mp = TRUE,
  debug_use_r_implementation = FALSE,
  ...,
  transform_function = "MULTIPLY",
  reduce_function = "SUM",
  mean_divider = "ONE",
  variance = FALSE
)

pfocal_fast_gaussian_confidence(
  data,
  vertical_r0 = 0.05,
  vertical_sd = 1,
  horizontal_r0 = vertical_r0,
  horizontal_sd = vertical_sd,
  tail_included = TRUE,
  na.rm = NA,
  mp = TRUE,
  debug_use_r_implementation = FALSE,
  ...,
  transform_function = "MULTIPLY",
  reduce_function = "SUM",
  mean_divider = "ONE",
  variance = FALSE
)

pfocal_fast_binomial(
  data,
  vertical_radius,
  horizontal_radius = vertical_radius,
  na.rm = NA,
  mp = TRUE,
  debug_use_r_implementation = FALSE,
  ...,
  transform_function = "MULTIPLY",
  reduce_function = "SUM",
  mean_divider = "ONE",
  variance = FALSE
)

pfocal_fast_abs_rectangle(
  data,
  height,
  width = height,
  value = 1,
  na.rm = NA,
  mp = TRUE,
  debug_use_r_implementation = FALSE,
  ...,
  transform_function = "MULTIPLY",
  reduce_function = "SUM",
  mean_divider = "ONE",
  variance = FALSE
)

pfocal_fast_separated(
  data,
  kernel_list,
  na.rm = NA,
  mp = TRUE,
  debug_use_r_implementation = FALSE,
  ...,
  transform_function = "MULTIPLY",
  reduce_function = "SUM",
  mean_divider = "ONE",
  variance = FALSE
)

Arguments

data

[matrix-type] Grid to compute onto.

vertical_radius

[numeric] The kernel's radius in the vertical dimension.

vertical_sd

[numeric] The kernel's standard deviation in the vertical dimension.

horizontal_radius

[numeric] The kernel's radius in the horizontal dimension.

horizontal_sd

[numeric] The kernel's standard deviation in the horizontal dimension.

tail_included

[logical] Whether or not to include the kernel tail.

na.rm

[NA OR logical] The behavior to adopt for dealing with missing values, default to NA (faster option). For possible values see pfocal_info_nan_policy().

mp

[logical] Whether to use the open_mp implementation, default to TRUE.

debug_use_r_implementation

[logical] Used for debugging purposes whether to use the slow R implementation instead of the fast C++ code. Default to FALSE.

...

None used at the moment .

transform_function

[character] The function to apply to the cell values covered by the kernel. For possible values, see pfocal_info_transform(). Default to "MULTIPLY".

reduce_function

[character] The function to apply to the kernel values after the function passed in transform_function has been applied (the function that summarize the data). For possible values, see pfocal_info_reduce(). Default to "SUM".

mean_divider

[character] Optional, allows to specify how the final value at each cell is divided by a value that can be function of the intermediate data resulting of applying transform_function. For possible values, see pfocal_info_mean_divisor(). Default to "ONE" (for no division).

variance

[logical] Whether to return the "variance" of the intermediate values at each point (for more details please see pfocal_info_variance()). Default to FALSE (just returns the value at each point).

vertical_r0

[numeric] The kernel's r0 (exponential) in the vertical dimension.

horizontal_r0

[numeric] The kernel's r0 (exponential) in the horizontal dimension.

height

[numeric] For rectangular kernels, the height of the rectangle.

width

[numeric] For rectangular kernels, the width of the rectangle.

value

[numeric] For single value matrices, the value.

kernel_list

[list] A list of kernels computed from functions in kernel-gaussian, kernel-binomial, kernel-circular, kernel-distance, kernel-exponential.

Value

The updated, convoluted grid.

Examples

data <- matrix(nrow = 10, ncol = 10, data = runif(10 * 10))

pfocal_fast_gaussian_radius(data, vertical_radius = 2)
#>            [,1]      [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
#>  [1,] 0.2830305 0.3564921 0.3408578 0.3259481 0.3222032 0.3035866 0.2962917
#>  [2,] 0.3325950 0.4289823 0.4384392 0.4417393 0.4360511 0.3986543 0.3750587
#>  [3,] 0.3752513 0.4881955 0.5146084 0.5160212 0.4998064 0.4612955 0.4359248
#>  [4,] 0.4175809 0.5456135 0.5792113 0.5867062 0.5616501 0.4971508 0.4630561
#>  [5,] 0.3749925 0.4997839 0.5575886 0.5845959 0.5737357 0.5161714 0.4679628
#>  [6,] 0.3830634 0.5048049 0.5507981 0.5760334 0.5672494 0.5066553 0.4665458
#>  [7,] 0.3722066 0.4974701 0.5366323 0.5633897 0.5691651 0.5107413 0.4610967
#>  [8,] 0.3240828 0.4291809 0.4601034 0.4979177 0.5240972 0.4760897 0.4250587
#>  [9,] 0.2933370 0.3871176 0.3974634 0.4124065 0.4406652 0.4188529 0.3862937
#> [10,] 0.2279868 0.3074376 0.3093696 0.3171378 0.3410101 0.3203725 0.2961359
#>            [,8]      [,9]     [,10]
#>  [1,] 0.3081050 0.3043337 0.2188925
#>  [2,] 0.3800499 0.3648993 0.2522887
#>  [3,] 0.4309301 0.4134842 0.2946543
#>  [4,] 0.4767233 0.4731422 0.3476182
#>  [5,] 0.4627504 0.4541688 0.3346118
#>  [6,] 0.4745532 0.4783096 0.3659562
#>  [7,] 0.4714131 0.4828513 0.3703392
#>  [8,] 0.4282309 0.4328547 0.3266719
#>  [9,] 0.3883098 0.3929411 0.2940551
#> [10,] 0.3101014 0.3156529 0.2241443
pfocal_fast_gaussian_confidence(data)
#>            [,1]      [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
#>  [1,] 0.3391918 0.4080249 0.3744669 0.3813887 0.3970256 0.3585916 0.3280565
#>  [2,] 0.4431841 0.5425455 0.5231566 0.5200600 0.5379305 0.5194067 0.4943683
#>  [3,] 0.4416645 0.5746424 0.5849502 0.5501566 0.5316087 0.5234293 0.5273928
#>  [4,] 0.3742877 0.5327090 0.5960687 0.5807571 0.5233649 0.4670856 0.4559266
#>  [5,] 0.3333446 0.4711273 0.5751657 0.6291648 0.5851919 0.4810180 0.4129428
#>  [6,] 0.3857132 0.4895984 0.5589010 0.6209399 0.6114945 0.5177385 0.4379752
#>  [7,] 0.4100732 0.5096402 0.5262626 0.5579605 0.5823347 0.5259161 0.4714885
#>  [8,] 0.3605253 0.4760321 0.4824606 0.5022986 0.5488491 0.5285680 0.4969866
#>  [9,] 0.3298406 0.4559274 0.4644590 0.4778968 0.5167909 0.5002450 0.4752956
#> [10,] 0.2546476 0.3559088 0.3664894 0.3761874 0.4031502 0.3808459 0.3460853
#>            [,8]      [,9]     [,10]
#>  [1,] 0.3348742 0.3326059 0.2547826
#>  [2,] 0.4912520 0.4698497 0.3273237
#>  [3,] 0.5370220 0.5105676 0.3354386
#>  [4,] 0.4750453 0.4609344 0.3099806
#>  [5,] 0.3993016 0.3872596 0.2922507
#>  [6,] 0.4218682 0.4296470 0.3632864
#>  [7,] 0.4925302 0.5217936 0.4264188
#>  [8,] 0.5197248 0.5361128 0.4020403
#>  [9,] 0.4857831 0.4866191 0.3429776
#> [10,] 0.3430588 0.3338488 0.2217933
pfocal_fast_binomial(data, vertical_radius = 2)
#>            [,1]      [,2]      [,3]      [,4]     [,5]      [,6]      [,7]
#>  [1,]  85.90097 103.57174  96.57368  97.23217 101.0219  92.57386  84.06030
#>  [2,] 113.98447 140.82752 135.74899 133.32406 137.0905 133.46618 127.23403
#>  [3,] 112.90168 149.01269 151.32118 142.11051 136.3598 134.29180 135.05426
#>  [4,]  94.61895 136.39056 153.11615 149.58226 134.6620 120.45670 116.81141
#>  [5,]  85.27089 122.63011 148.81656 161.23849 149.1918 122.85768 105.28395
#>  [6,]  96.97069 126.08832 143.79178 159.68505 156.7579 132.52455 110.97281
#>  [7,] 103.32737 130.07185 134.77786 143.36036 148.6838 135.51697 121.13917
#>  [8,]  93.91605 123.87139 125.78311 130.29573 140.2267 136.12300 128.14852
#>  [9,]  84.41941 116.48161 120.15984 123.36903 132.1505 129.37740 122.06475
#> [10,]  64.56326  90.07144  93.80455  96.11021 101.6621  97.13778  88.30818
#>            [,8]      [,9]     [,10]
#>  [1,]  85.71089  85.04564  64.34104
#>  [2,] 127.90794 121.16441  84.58422
#>  [3,] 139.91282 131.03321  86.70759
#>  [4,] 122.58430 116.57972  78.37503
#>  [5,] 102.98117 100.21906  75.37189
#>  [6,] 107.59536 110.80733  91.91035
#>  [7,] 125.92059 133.46639 107.96493
#>  [8,] 135.63485 139.42692 104.65554
#>  [9,] 126.14918 124.31474  87.63926
#> [10,]  88.45091  84.54565  56.76815
pfocal_fast_abs_rectangle(data, height = 2)
#>            [,1]     [,2]     [,3]      [,4]     [,5]     [,6]      [,7]
#>  [1,] 0.7124518 1.674245 1.131312 0.7593575 1.313131 1.017938 0.5567656
#>  [2,] 1.4622444 2.929693 2.127446 1.8342458 2.640661 2.454112 1.8262239
#>  [3,] 1.7018963 2.564628 2.317420 2.1545347 2.248753 2.585908 2.6010521
#>  [4,] 1.3106323 2.519248 2.942671 2.3311981 1.694720 1.578508 2.0172080
#>  [5,] 0.5909398 1.640401 2.256489 2.6726797 2.643370 1.763207 1.2395955
#>  [6,] 1.1483516 1.452999 1.549433 2.7867027 3.104935 2.544383 1.7675051
#>  [7,] 1.8508166 2.733025 2.175061 2.3375255 2.597339 2.450179 1.6411020
#>  [8,] 1.1654617 2.210336 2.052566 1.6150077 2.204046 2.611793 1.6932999
#>  [9,] 0.7124929 1.844654 1.982799 1.4893142 2.094844 2.202301 2.1698675
#> [10,] 0.9645162 2.395859 2.405735 1.8294572 2.494724 2.310343 1.9415138
#>            [,8]      [,9]    [,10]
#>  [1,] 0.8719176 0.8982772 1.035914
#>  [2,] 1.7876030 2.0302359 2.150845
#>  [3,] 2.2289368 2.4001487 2.221034
#>  [4,] 2.1516229 2.5412108 2.489183
#>  [5,] 1.3545631 1.8760450 1.668112
#>  [6,] 1.4580281 0.9705895 1.268333
#>  [7,] 1.8381021 1.9300442 2.687070
#>  [8,] 1.7751473 2.5057244 2.597008
#>  [9,] 1.9957280 2.2264304 2.465953
#> [10,] 1.8587715 2.2743595 2.355674

pfocal_fast_separated(data, 
                      kernel_list = list(binomial_kernel(vertical_radius = 2, 
                                                         horizontal_radius = 2), 
                                         distance_kernel(vertical_radius = 2,
                                                         horizontal_radius = 2)))
#>           [,1]     [,2]     [,3]     [,4]     [,5]     [,6]     [,7]     [,8]
#>  [1,] 1900.415 2459.859 3318.160 3451.205 3354.277 3315.545 3264.319 2949.604
#>  [2,] 2449.941 3185.694 4289.416 4488.267 4370.341 4271.962 4138.582 3700.050
#>  [3,] 3240.441 4276.717 5711.155 5991.970 5861.441 5690.059 5419.711 4791.644
#>  [4,] 3505.498 4660.937 6178.385 6490.970 6398.721 6234.665 5933.672 5252.198
#>  [5,] 3512.681 4690.268 6190.711 6483.436 6397.669 6285.586 6060.287 5432.027
#>  [6,] 3341.661 4519.073 6029.104 6347.411 6267.244 6192.625 6022.037 5417.980
#>  [7,] 3181.629 4338.497 5886.271 6255.001 6182.116 6122.505 5974.279 5360.296
#>  [8,] 2960.361 3997.620 5469.386 5835.748 5758.856 5727.079 5643.687 5084.039
#>  [9,] 2187.568 2904.234 4004.340 4294.797 4248.975 4278.059 4298.309 3919.728
#> [10,] 1616.101 2139.787 2944.414 3190.615 3183.103 3230.441 3269.858 2991.123
#>           [,9]    [,10]
#>  [1,] 2175.146 1660.562
#>  [2,] 2718.600 2102.272
#>  [3,] 3505.296 2697.549
#>  [4,] 3834.915 2931.793
#>  [5,] 3975.338 3045.413
#>  [6,] 3935.957 3020.519
#>  [7,] 3852.017 2955.811
#>  [8,] 3656.481 2823.186
#>  [9,] 2843.645 2226.380
#> [10,] 2200.056 1711.685