Fast methods for common kernel computations

Methods wrapping pfocal to implement common kernel computations with default argument values.

pfocal_fast_gaussian_radius(
  data,
  vertical_radius,
  vertical_sd = 1,
  horizontal_radius = vertical_radius,
  horizontal_sd = vertical_sd,
  tail_included = TRUE,
  na.rm = NA,
  mp = TRUE,
  debug_use_r_implementation = FALSE,
  ...,
  transform_function = "MULTIPLY",
  reduce_function = "SUM",
  mean_divider = "ONE",
  variance = FALSE
)

pfocal_fast_gaussian_confidence(
  data,
  vertical_r0 = 0.05,
  vertical_sd = 1,
  horizontal_r0 = vertical_r0,
  horizontal_sd = vertical_sd,
  tail_included = TRUE,
  na.rm = NA,
  mp = TRUE,
  debug_use_r_implementation = FALSE,
  ...,
  transform_function = "MULTIPLY",
  reduce_function = "SUM",
  mean_divider = "ONE",
  variance = FALSE
)

pfocal_fast_binomial(
  data,
  vertical_radius,
  horizontal_radius = vertical_radius,
  na.rm = NA,
  mp = TRUE,
  debug_use_r_implementation = FALSE,
  ...,
  transform_function = "MULTIPLY",
  reduce_function = "SUM",
  mean_divider = "ONE",
  variance = FALSE
)

pfocal_fast_abs_rectangle(
  data,
  height,
  width = height,
  value = 1,
  na.rm = NA,
  mp = TRUE,
  debug_use_r_implementation = FALSE,
  ...,
  transform_function = "MULTIPLY",
  reduce_function = "SUM",
  mean_divider = "ONE",
  variance = FALSE
)

pfocal_fast_separated(
  data,
  kernel_list,
  na.rm = NA,
  mp = TRUE,
  debug_use_r_implementation = FALSE,
  ...,
  transform_function = "MULTIPLY",
  reduce_function = "SUM",
  mean_divider = "ONE",
  variance = FALSE
)

Arguments

data

[matrix-type] Grid to compute onto.

vertical_radius

[numeric] The kernel's radius in the vertical dimension.

vertical_sd

[numeric] The kernel's standard deviation in the vertical dimension.

horizontal_radius

[numeric] The kernel's radius in the horizontal dimension.

horizontal_sd

[numeric] The kernel's standard deviation in the horizontal dimension.

tail_included

[logical] Whether or not to include the kernel tail.

na.rm

[NA OR logical] The behavior to adopt for dealing with missing values, default to NA (faster option). For possible values see pfocal_info_nan_policy().

mp

[logical] Whether to use the open_mp implementation, default to TRUE.

debug_use_r_implementation

[logical] Used for debugging purposes whether to use the slow R implementation instead of the fass C++ code. Default to FALSE.

...

None used at the moment .

transform_function

[character] The function to apply to the cell values covered by the kernel. For possible values, see pfocal_info_transform(). Default to "MULTIPLY".

reduce_function

[character] The function to apply to the kernel values after the function passed in transform_function has been applied (the function that summarize the data). For possible values, see pfocal_info_reduce(). Default to "SUM".

mean_divider

[character] Optional, allows to specify how the final value at each cell is divided by a value that can be function of the intermediate data resulting of applying transform_function. For possible values, see pfocal_info_mean_divisor(). Default to "ONE" (for no division).

variance

[logical] Whether to return the "variance" of the intermediate values at each point (for more details please see pfocal_info_variance()). Default to FALSE (just returns the value at each point).

vertical_r0

[numeric] The kernel's r0 (exponential) in the vertical dimension.

horizontal_r0

[numeric] The kernel's r0 (exponential) in the horizontal dimension.

height

[numeric] For rectangular kernels, the height of the rectangle.

width

[numeric] For rectangular kernels, the width of the rectangle.

value

[numeric] For single value matrices, the value.

kernel_list

[list] A list of kernels computed from functions in kernel-gaussian, kernel-binomial, kernel-circular, kernel-distance, kernel-exponential.

Value

The updated, convoluted grid.

Examples

data <- matrix(nrow = 10, ncol = 10, data = runif(10 * 10))

pfocal_fast_gaussian_radius(data, vertical_radius = 2)
#>            [,1]      [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
#>  [1,] 0.2474072 0.3293507 0.3279583 0.3110587 0.3151395 0.3079236 0.2997816
#>  [2,] 0.2860435 0.3803978 0.3881491 0.3841804 0.3959956 0.3917980 0.3797673
#>  [3,] 0.3356397 0.4258833 0.4185547 0.4076685 0.4164904 0.4233197 0.4334013
#>  [4,] 0.3737855 0.4796038 0.4681514 0.4515071 0.4654013 0.4709843 0.4758184
#>  [5,] 0.3799912 0.4810117 0.4703095 0.4589074 0.4596366 0.4555532 0.4684480
#>  [6,] 0.3954794 0.5003391 0.4979029 0.4889283 0.4708497 0.4320677 0.4343231
#>  [7,] 0.4020587 0.5213966 0.5105218 0.4932666 0.4841564 0.4471873 0.4538642
#>  [8,] 0.3577083 0.4730113 0.4837955 0.4827114 0.4623911 0.4003989 0.3904499
#>  [9,] 0.3203937 0.4239485 0.4296089 0.4254694 0.4079196 0.3538773 0.3560695
#> [10,] 0.2293788 0.3207348 0.3337706 0.3390511 0.3421083 0.3047142 0.2992388
#>            [,8]      [,9]     [,10]
#>  [1,] 0.2819027 0.2345194 0.1579381
#>  [2,] 0.3594101 0.3026041 0.1965484
#>  [3,] 0.4228532 0.3527743 0.2290793
#>  [4,] 0.4747121 0.4167291 0.2855239
#>  [5,] 0.4816151 0.4330382 0.3049791
#>  [6,] 0.4804882 0.4752265 0.3563032
#>  [7,] 0.5091482 0.5054146 0.3866615
#>  [8,] 0.4563482 0.4783525 0.3702169
#>  [9,] 0.4247179 0.4393548 0.3357915
#> [10,] 0.3523181 0.3630648 0.2706266
pfocal_fast_gaussian_confidence(data)
#>            [,1]      [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
#>  [1,] 0.3018414 0.4076637 0.4053858 0.3726199 0.3648425 0.3331045 0.3011856
#>  [2,] 0.3671983 0.4923090 0.5121445 0.5049816 0.5018730 0.4829847 0.4706673
#>  [3,] 0.3738098 0.4878123 0.5036374 0.5098400 0.5251326 0.5361763 0.5434717
#>  [4,] 0.3354048 0.4264819 0.4150512 0.4191444 0.4625947 0.5017868 0.5230926
#>  [5,] 0.3202228 0.3864012 0.3559349 0.3489147 0.3767871 0.4058997 0.4329832
#>  [6,] 0.3970381 0.4778617 0.4384588 0.3991876 0.3787995 0.3661584 0.3895215
#>  [7,] 0.4701463 0.6076924 0.5975046 0.5563903 0.5068448 0.4343351 0.4213562
#>  [8,] 0.4554636 0.6151722 0.6499317 0.6519658 0.5964911 0.4755310 0.4411783
#>  [9,] 0.3528931 0.4863563 0.5259760 0.5565676 0.5400711 0.4484024 0.4398160
#> [10,] 0.2236646 0.3145116 0.3190606 0.3312252 0.3632188 0.3522471 0.3673941
#>            [,8]      [,9]     [,10]
#>  [1,] 0.2766552 0.2431197 0.1818306
#>  [2,] 0.4523877 0.3746625 0.2320686
#>  [3,] 0.5248188 0.4301727 0.2569655
#>  [4,] 0.5024323 0.4108233 0.2611174
#>  [5,] 0.4385904 0.3809459 0.2648838
#>  [6,] 0.4384033 0.4292942 0.3305313
#>  [7,] 0.4900013 0.5256704 0.4292091
#>  [8,] 0.5242336 0.5785635 0.4695015
#>  [9,] 0.5374450 0.5654767 0.4245089
#> [10,] 0.4302089 0.4168478 0.2887726
pfocal_fast_binomial(data, vertical_radius = 2)
#>            [,1]      [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
#>  [1,]  76.33959 103.57237 103.52025  95.94232  92.15064  85.31065  76.90870
#>  [2,]  95.19062 128.49141 133.62639 130.92163 128.95731 124.62922 121.21835
#>  [3,]  95.02906 125.60296 130.08641 130.90621 135.18709 138.80592 141.18006
#>  [4,]  84.82887 108.30277 106.44682 107.06359 117.76942 128.49274 134.00514
#>  [5,]  82.14420  99.55621  91.70863  88.15980  95.16356 103.73631 111.55502
#>  [6,] 100.61367 122.90120 113.29271 102.01404  96.89039  94.32828 100.46534
#>  [7,] 120.35415 156.44520 154.68699 143.18675 128.47788 109.92656 106.79298
#>  [8,] 116.48508 158.56612 168.86077 168.25371 152.40384 122.57560 113.66269
#>  [9,]  90.84895 125.92809 137.47343 144.22901 138.37778 116.79154 114.21024
#> [10,]  56.81367  79.29419  82.08068  85.48862  91.50428  89.30193  93.80931
#>            [,8]      [,9]     [,10]
#>  [1,]  70.83446  62.60305  45.48641
#>  [2,] 116.25686  96.71735  60.03966
#>  [3,] 136.20345 110.86503  65.91867
#>  [4,] 128.66727 105.29682  65.90638
#>  [5,] 112.79129  98.53810  67.47382
#>  [6,] 112.04503 111.10271  84.24250
#>  [7,] 124.34303 135.27784 109.14372
#>  [8,] 135.87055 151.14861 120.93403
#>  [9,] 138.25746 146.47734 109.21948
#> [10,] 109.19203 106.96390  73.70333
pfocal_fast_abs_rectangle(data, height = 2)
#>            [,1]     [,2]     [,3]      [,4]     [,5]     [,6]      [,7]
#>  [1,] 0.7927982 1.545394 1.729390 1.0093014 1.024912 1.172284 0.6480907
#>  [2,] 1.0902588 2.342830 2.546920 1.9242596 2.110231 2.173373 1.6428615
#>  [3,] 1.1282982 2.075855 1.956657 2.4158563 2.367902 2.150154 2.1798808
#>  [4,] 1.2630620 2.412960 2.172670 1.8686161 2.019489 2.234788 2.3733844
#>  [5,] 0.5983147 1.566172 1.379848 0.8920305 1.485184 1.981388 2.0132119
#>  [6,] 1.1490534 1.845663 1.053464 1.1637267 1.118389 1.011738 1.1360439
#>  [7,] 1.5085818 2.785189 2.329392 1.7996259 1.532239 1.549753 1.1277935
#>  [8,] 1.4081346 3.163126 3.281631 2.8440638 2.925194 2.572889 1.6564751
#>  [9,] 1.3005818 2.471071 2.383225 2.8563172 3.372007 2.095900 1.0231433
#> [10,] 0.6877071 1.769149 1.823011 1.5312049 2.243160 2.253685 1.3748024
#>            [,8]      [,9]     [,10]
#>  [1,] 0.5608721 0.2761554 0.8703588
#>  [2,] 1.9832119 1.5984331 1.2964713
#>  [3,] 2.1854860 2.4863865 1.4975600
#>  [4,] 2.4813921 2.1959085 1.6900674
#>  [5,] 2.1881294 1.7616221 1.2914756
#>  [6,] 1.4399869 1.5064943 1.2508414
#>  [7,] 1.8695304 2.2381932 2.1041584
#>  [8,] 1.7494920 2.2487942 2.8954025
#>  [9,] 1.8041433 2.5103874 2.8443849
#> [10,] 2.3243991 3.1482912 2.4489897

pfocal_fast_separated(data, 
                      kernel_list = list(binomial_kernel(vertical_radius = 2, 
                                                         horizontal_radius = 2), 
                                         distance_kernel(vertical_radius = 2,
                                                         horizontal_radius = 2)))
#>           [,1]     [,2]     [,3]     [,4]     [,5]     [,6]     [,7]     [,8]
#>  [1,] 1725.170 2241.420 3023.091 3244.546 3243.048 3184.840 3025.399 2653.548
#>  [2,] 2126.483 2746.938 3745.267 4046.388 4082.720 4055.495 3890.262 3440.579
#>  [3,] 2760.860 3567.876 4817.084 5148.838 5155.281 5111.842 4925.925 4403.869
#>  [4,] 2964.874 3836.221 5151.280 5455.477 5448.273 5438.783 5304.553 4815.673
#>  [5,] 3153.646 4079.635 5447.448 5707.905 5642.762 5626.080 5534.497 5089.481
#>  [6,] 3243.035 4204.285 5604.958 5826.237 5712.615 5698.765 5638.905 5212.894
#>  [7,] 3210.296 4160.823 5534.660 5713.120 5590.672 5636.770 5636.684 5223.936
#>  [8,] 3079.694 3982.386 5271.804 5424.311 5306.054 5399.869 5448.793 5051.291
#>  [9,] 2499.732 3254.886 4290.861 4393.463 4254.337 4285.771 4284.437 3934.800
#> [10,] 1884.939 2512.280 3322.564 3437.922 3351.088 3380.626 3371.094 3091.712
#>           [,9]    [,10]
#>  [1,] 1960.235 1426.029
#>  [2,] 2554.906 1880.098
#>  [3,] 3300.208 2448.191
#>  [4,] 3653.677 2747.933
#>  [5,] 3864.538 2956.007
#>  [6,] 3927.122 3068.061
#>  [7,] 3913.950 3125.707
#>  [8,] 3773.752 3056.164
#>  [9,] 2914.465 2403.106
#> [10,] 2320.375 1915.017