Fast methods for common kernel computations

Methods wrapping pfocal to implement common kernel computations with default argument values.

pfocal_fast_gaussian_radius(
  data,
  vertical_radius,
  vertical_sd = 1,
  horizontal_radius = vertical_radius,
  horizontal_sd = vertical_sd,
  tail_included = TRUE,
  na.rm = NA,
  mp = TRUE,
  debug_use_r_implementation = FALSE,
  ...,
  transform_function = "MULTIPLY",
  reduce_function = "SUM",
  mean_divider = "ONE",
  variance = FALSE
)

pfocal_fast_gaussian_confidence(
  data,
  vertical_r0 = 0.05,
  vertical_sd = 1,
  horizontal_r0 = vertical_r0,
  horizontal_sd = vertical_sd,
  tail_included = TRUE,
  na.rm = NA,
  mp = TRUE,
  debug_use_r_implementation = FALSE,
  ...,
  transform_function = "MULTIPLY",
  reduce_function = "SUM",
  mean_divider = "ONE",
  variance = FALSE
)

pfocal_fast_binomial(
  data,
  vertical_radius,
  horizontal_radius = vertical_radius,
  na.rm = NA,
  mp = TRUE,
  debug_use_r_implementation = FALSE,
  ...,
  transform_function = "MULTIPLY",
  reduce_function = "SUM",
  mean_divider = "ONE",
  variance = FALSE
)

pfocal_fast_abs_rectangle(
  data,
  height,
  width = height,
  value = 1,
  na.rm = NA,
  mp = TRUE,
  debug_use_r_implementation = FALSE,
  ...,
  transform_function = "MULTIPLY",
  reduce_function = "SUM",
  mean_divider = "ONE",
  variance = FALSE
)

pfocal_fast_separated(
  data,
  kernel_list,
  na.rm = NA,
  mp = TRUE,
  debug_use_r_implementation = FALSE,
  ...,
  transform_function = "MULTIPLY",
  reduce_function = "SUM",
  mean_divider = "ONE",
  variance = FALSE
)

Arguments

data

[matrix-type] Grid to compute onto.

vertical_radius

[numeric] The kernel's radius in the vertical dimension.

vertical_sd

[numeric] The kernel's standard deviation in the vertical dimension.

horizontal_radius

[numeric] The kernel's radius in the horizontal dimension.

horizontal_sd

[numeric] The kernel's standard deviation in the horizontal dimension.

tail_included

[logical] Whether or not to include the kernel tail.

na.rm

[NA OR logical] The behavior to adopt for dealing with missing values, default to NA (faster option). For possible values see pfocal_info_nan_policy().

mp

[logical] Whether to use the open_mp implementation, default to TRUE.

debug_use_r_implementation

[logical] Used for debugging purposes whether to use the slow R implementation instead of the fast C++ code. Default to FALSE.

...

None used at the moment .

transform_function

[character] The function to apply to the cell values covered by the kernel. For possible values, see pfocal_info_transform(). Default to "MULTIPLY".

reduce_function

[character] The function to apply to the kernel values after the function passed in transform_function has been applied (the function that summarize the data). For possible values, see pfocal_info_reduce(). Default to "SUM".

mean_divider

[character] Optional, allows to specify how the final value at each cell is divided by a value that can be function of the intermediate data resulting of applying transform_function. For possible values, see pfocal_info_mean_divisor(). Default to "ONE" (for no division).

variance

[logical] Whether to return the "variance" of the intermediate values at each point (for more details please see pfocal_info_variance()). Default to FALSE (just returns the value at each point).

vertical_r0

[numeric] The kernel's r0 (exponential) in the vertical dimension.

horizontal_r0

[numeric] The kernel's r0 (exponential) in the horizontal dimension.

height

[numeric] For rectangular kernels, the height of the rectangle.

width

[numeric] For rectangular kernels, the width of the rectangle.

value

[numeric] For single value matrices, the value.

kernel_list

[list] A list of kernels computed from functions in kernel-gaussian, kernel-binomial, kernel-circular, kernel-distance, kernel-exponential.

Value

The updated, convoluted grid.

Examples

data <- matrix(nrow = 10, ncol = 10, data = runif(10 * 10))

pfocal_fast_gaussian_radius(data, vertical_radius = 2)
#>            [,1]      [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
#>  [1,] 0.2073847 0.2710120 0.2536773 0.2325945 0.2606583 0.3052734 0.3292630
#>  [2,] 0.2721798 0.3484755 0.3242323 0.3050201 0.3554642 0.4139620 0.4296491
#>  [3,] 0.3206804 0.4232747 0.4167910 0.3979120 0.4342006 0.4777389 0.4851412
#>  [4,] 0.3818749 0.4987112 0.4757651 0.4447081 0.4735506 0.5104887 0.5230100
#>  [5,] 0.3866774 0.5065191 0.4977676 0.4730971 0.4927891 0.5075280 0.5068862
#>  [6,] 0.3652599 0.4968891 0.5040367 0.4872609 0.5024734 0.5054210 0.5070158
#>  [7,] 0.3875473 0.5127978 0.5066682 0.4797117 0.4621278 0.4391771 0.4566463
#>  [8,] 0.3257185 0.4315074 0.4364137 0.4281458 0.4257261 0.4054449 0.4166194
#>  [9,] 0.2663007 0.3630871 0.3870827 0.3855367 0.3619029 0.3310990 0.3489808
#> [10,] 0.2057759 0.2722851 0.2799473 0.2764779 0.2556581 0.2365726 0.2618250
#>            [,8]      [,9]     [,10]
#>  [1,] 0.3238888 0.3054926 0.2395752
#>  [2,] 0.4272639 0.4146316 0.3204056
#>  [3,] 0.4886737 0.4791590 0.3664899
#>  [4,] 0.5409091 0.5393972 0.4091061
#>  [5,] 0.5335043 0.5290525 0.3941369
#>  [6,] 0.5394880 0.5362775 0.4020245
#>  [7,] 0.5041542 0.4992769 0.3801218
#>  [8,] 0.4555393 0.4471674 0.3455496
#>  [9,] 0.3829141 0.3771428 0.3036046
#> [10,] 0.2932470 0.2997465 0.2505842
pfocal_fast_gaussian_confidence(data)
#>            [,1]      [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
#>  [1,] 0.2609792 0.3190035 0.2677533 0.2383136 0.2842002 0.3359161 0.3663504
#>  [2,] 0.3176833 0.4060974 0.4040932 0.3950557 0.4536225 0.5302336 0.5500845
#>  [3,] 0.3009351 0.4209698 0.4549673 0.4549334 0.5078472 0.5813792 0.5895235
#>  [4,] 0.3536177 0.4785656 0.4571273 0.4178397 0.4526545 0.5089611 0.5269522
#>  [5,] 0.4301428 0.5446879 0.4907347 0.4442308 0.4670605 0.4802856 0.4766576
#>  [6,] 0.4478766 0.5754474 0.5609590 0.5350474 0.5230450 0.4844617 0.4652699
#>  [7,] 0.3961165 0.5322508 0.5462263 0.5303624 0.5117778 0.4716108 0.4725384
#>  [8,] 0.3076458 0.4386325 0.4696762 0.4648386 0.4723121 0.4622176 0.4865595
#>  [9,] 0.2746446 0.3860087 0.4167181 0.4161226 0.4060538 0.3984393 0.4415053
#> [10,] 0.2448099 0.3115562 0.3193349 0.3209639 0.2774991 0.2392901 0.2705337
#>            [,8]      [,9]     [,10]
#>  [1,] 0.3537359 0.3050781 0.2336060
#>  [2,] 0.5239472 0.4910422 0.3900977
#>  [3,] 0.5778805 0.5894761 0.4719044
#>  [4,] 0.5587549 0.5862227 0.4454540
#>  [5,] 0.5240261 0.5285011 0.3747421
#>  [6,] 0.5022514 0.4849126 0.3451546
#>  [7,] 0.5078409 0.4882472 0.3663113
#>  [8,] 0.5265264 0.5070522 0.3945801
#>  [9,] 0.4872769 0.4843721 0.3969860
#> [10,] 0.3120656 0.3305237 0.2906032
pfocal_fast_binomial(data, vertical_radius = 2)
#>            [,1]      [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
#>  [1,]  66.49977  81.00566  68.86570  61.37254  72.18280  86.88745  94.10449
#>  [2,]  80.45417 104.54509 103.77438 102.06127 116.84362 136.43455 141.49475
#>  [3,]  78.09499 109.61814 116.62632 116.16319 129.50038 148.05915 151.41011
#>  [4,]  90.11974 121.91047 117.51024 107.84426 117.18261 131.04611 135.84180
#>  [5,] 110.04913 140.12082 127.82372 114.81170 119.01671 121.78859 122.74798
#>  [6,] 115.16659 148.33306 144.29476 136.14631 132.26051 122.89239 119.33943
#>  [7,] 100.42421 136.14998 140.90813 136.85466 131.21812 121.96765 121.60780
#>  [8,]  79.40979 113.65019 122.17082 120.99857 120.32537 119.40367 125.80105
#>  [9,]  70.62867  99.47387 107.61680 107.02441 103.37064 102.33931 112.23600
#> [10,]  60.94437  78.87697  82.23114  80.76435  70.44757  62.09628  68.82918
#>            [,8]      [,9]     [,10]
#>  [1,]  90.57188  79.00096  58.92286
#>  [2,] 135.54421 127.10167  99.38381
#>  [3,] 150.77651 151.98289 120.58312
#>  [4,] 145.17422 150.34048 114.24111
#>  [5,] 134.72072 135.36448  96.28901
#>  [6,] 128.39017 124.67259  88.13415
#>  [7,] 130.12135 125.72912  92.64667
#>  [8,] 136.05119 132.70496 101.41213
#>  [9,] 124.68883 125.62775 100.67728
#> [10,]  80.44552  86.01815  73.35481
pfocal_fast_abs_rectangle(data, height = 2)
#>            [,1]     [,2]      [,3]       [,4]      [,5]      [,6]      [,7]
#>  [1,] 0.3656397 1.284014 0.9892596 0.07177262 0.6622355 0.8217164 0.6713449
#>  [2,] 1.3061913 2.472699 1.9176937 1.11897303 1.2515097 1.8060576 2.0557748
#>  [3,] 1.0580267 1.341453 1.7235725 2.19182798 1.8873763 2.8667520 3.1013217
#>  [4,] 0.4722258 1.481883 2.1244915 1.87268664 1.7438984 2.0846076 2.5814900
#>  [5,] 1.1070979 2.865310 2.1969131 0.90112762 1.0556395 1.6004013 1.7979937
#>  [6,] 1.4419539 2.769779 2.2596624 1.85846782 2.1138794 2.6602137 1.6255135
#>  [7,] 1.3854535 2.750630 2.6270652 2.68732279 2.5532516 1.8904565 1.6380575
#>  [8,] 0.9558631 2.077849 2.1607874 1.68256757 2.0164887 1.9922467 1.4962000
#>  [9,] 0.4564595 1.117777 1.9149226 1.52606718 1.6149897 2.2113098 1.8224565
#> [10,] 0.9672261 2.009390 1.7459519 1.90618250 1.9175684 1.1433008 1.4828841
#>           [,8]      [,9]     [,10]
#>  [1,] 1.127405 0.7486429 0.5923658
#>  [2,] 2.468823 1.8390977 1.6845115
#>  [3,] 2.371887 2.2614392 2.9028686
#>  [4,] 2.015594 2.3519817 3.2634848
#>  [5,] 1.738691 2.6537343 2.5815806
#>  [6,] 1.669669 2.4721292 1.7054947
#>  [7,] 2.001275 2.0654482 1.9950283
#>  [8,] 2.078650 2.0830669 2.0829911
#>  [9,] 2.508985 2.3629755 2.2338131
#> [10,] 2.110216 1.8991210 2.4975598

pfocal_fast_separated(data, 
                      kernel_list = list(binomial_kernel(vertical_radius = 2, 
                                                         horizontal_radius = 2), 
                                         distance_kernel(vertical_radius = 2,
                                                         horizontal_radius = 2)))
#>           [,1]     [,2]     [,3]     [,4]     [,5]     [,6]     [,7]     [,8]
#>  [1,] 1410.260 1834.616 2582.158 2941.248 3084.708 3182.531 3318.628 3247.224
#>  [2,] 1926.253 2455.895 3416.758 3835.213 3962.944 4081.099 4292.440 4206.245
#>  [3,] 2732.382 3452.201 4706.020 5155.148 5237.044 5367.847 5609.287 5425.805
#>  [4,] 3103.206 3972.943 5397.157 5867.436 5912.150 6002.133 6177.445 5882.646
#>  [5,] 3203.660 4126.111 5591.652 6033.543 6039.010 6109.369 6240.027 5898.110
#>  [6,] 3182.714 4081.876 5493.932 5866.776 5849.179 5938.559 6056.631 5711.998
#>  [7,] 3141.910 4041.136 5379.510 5686.171 5659.284 5731.405 5786.466 5436.193
#>  [8,] 2918.977 3786.164 4985.404 5223.637 5188.683 5210.296 5185.764 4864.511
#>  [9,] 2088.230 2722.555 3604.389 3823.857 3857.809 3905.226 3891.829 3661.926
#> [10,] 1453.784 1915.853 2545.117 2734.710 2815.620 2903.285 2939.326 2815.488
#>           [,9]    [,10]
#>  [1,] 2452.432 1849.834
#>  [2,] 3158.037 2419.296
#>  [3,] 4063.789 3114.212
#>  [4,] 4392.717 3386.862
#>  [5,] 4427.740 3449.759
#>  [6,] 4321.615 3391.476
#>  [7,] 4120.654 3242.064
#>  [8,] 3678.744 2908.150
#>  [9,] 2750.639 2177.666
#> [10,] 2138.983 1688.436