Fast methods for common kernel computations

Methods wrapping pfocal to implement common kernel computations with default argument values.

pfocal_fast_gaussian_radius(
  data,
  vertical_radius,
  vertical_sd = 1,
  horizontal_radius = vertical_radius,
  horizontal_sd = vertical_sd,
  tail_included = TRUE,
  na.rm = NA,
  mp = TRUE,
  debug_use_r_implementation = FALSE,
  ...,
  transform_function = "MULTIPLY",
  reduce_function = "SUM",
  mean_divider = "ONE",
  variance = FALSE
)

pfocal_fast_gaussian_confidence(
  data,
  vertical_r0 = 0.05,
  vertical_sd = 1,
  horizontal_r0 = vertical_r0,
  horizontal_sd = vertical_sd,
  tail_included = TRUE,
  na.rm = NA,
  mp = TRUE,
  debug_use_r_implementation = FALSE,
  ...,
  transform_function = "MULTIPLY",
  reduce_function = "SUM",
  mean_divider = "ONE",
  variance = FALSE
)

pfocal_fast_binomial(
  data,
  vertical_radius,
  horizontal_radius = vertical_radius,
  na.rm = NA,
  mp = TRUE,
  debug_use_r_implementation = FALSE,
  ...,
  transform_function = "MULTIPLY",
  reduce_function = "SUM",
  mean_divider = "ONE",
  variance = FALSE
)

pfocal_fast_abs_rectangle(
  data,
  height,
  width = height,
  value = 1,
  na.rm = NA,
  mp = TRUE,
  debug_use_r_implementation = FALSE,
  ...,
  transform_function = "MULTIPLY",
  reduce_function = "SUM",
  mean_divider = "ONE",
  variance = FALSE
)

pfocal_fast_separated(
  data,
  kernel_list,
  na.rm = NA,
  mp = TRUE,
  debug_use_r_implementation = FALSE,
  ...,
  transform_function = "MULTIPLY",
  reduce_function = "SUM",
  mean_divider = "ONE",
  variance = FALSE
)

Arguments

data

[matrix-type] Grid to compute onto.

vertical_radius

[numeric] The kernel's radius in the vertical dimension.

vertical_sd

[numeric] The kernel's standard deviation in the vertical dimension.

horizontal_radius

[numeric] The kernel's radius in the horizontal dimension.

horizontal_sd

[numeric] The kernel's standard deviation in the horizontal dimension.

tail_included

[logical] Whether or not to include the kernel tail.

na.rm

[NA OR logical] The behavior to adopt for dealing with missing values, default to NA (faster option). For possible values see pfocal_info_nan_policy().

mp

[logical] Whether to use the open_mp implementation, default to TRUE.

debug_use_r_implementation

[logical] Used for debugging purposes whether to use the slow R implementation instead of the fast C++ code. Default to FALSE.

...

None used at the moment .

transform_function

[character] The function to apply to the cell values covered by the kernel. For possible values, see pfocal_info_transform(). Default to "MULTIPLY".

reduce_function

[character] The function to apply to the kernel values after the function passed in transform_function has been applied (the function that summarize the data). For possible values, see pfocal_info_reduce(). Default to "SUM".

mean_divider

[character] Optional, allows to specify how the final value at each cell is divided by a value that can be function of the intermediate data resulting of applying transform_function. For possible values, see pfocal_info_mean_divisor(). Default to "ONE" (for no division).

variance

[logical] Whether to return the "variance" of the intermediate values at each point (for more details please see pfocal_info_variance()). Default to FALSE (just returns the value at each point).

vertical_r0

[numeric] The kernel's r0 (exponential) in the vertical dimension.

horizontal_r0

[numeric] The kernel's r0 (exponential) in the horizontal dimension.

height

[numeric] For rectangular kernels, the height of the rectangle.

width

[numeric] For rectangular kernels, the width of the rectangle.

value

[numeric] For single value matrices, the value.

kernel_list

[list] A list of kernels computed from functions in kernel-gaussian, kernel-binomial, kernel-circular, kernel-distance, kernel-exponential.

Value

The updated, convoluted grid.

Examples

data <- matrix(nrow = 10, ncol = 10, data = runif(10 * 10))

pfocal_fast_gaussian_radius(data, vertical_radius = 2)
#>            [,1]      [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
#>  [1,] 0.2308718 0.2744517 0.2653565 0.2657975 0.2642718 0.2730405 0.3189429
#>  [2,] 0.2637514 0.3224855 0.3396707 0.3567389 0.3457632 0.3358861 0.3741603
#>  [3,] 0.3208018 0.3894988 0.3962382 0.4018229 0.3989115 0.4035441 0.4530349
#>  [4,] 0.3643354 0.4438369 0.4533710 0.4489657 0.4305403 0.4358886 0.5004174
#>  [5,] 0.3507506 0.4312129 0.4325024 0.4248492 0.4190531 0.4284470 0.4936344
#>  [6,] 0.3450493 0.4408663 0.4582584 0.4353694 0.4029535 0.4143008 0.4910288
#>  [7,] 0.3628610 0.4616319 0.4543039 0.4069594 0.3812100 0.4151542 0.4990367
#>  [8,] 0.2945834 0.3927338 0.4074884 0.3699623 0.3315671 0.3468795 0.4149359
#>  [9,] 0.2624464 0.3466681 0.3431030 0.2921855 0.2658395 0.3010648 0.3659722
#> [10,] 0.1844328 0.2463743 0.2519967 0.2125227 0.1844821 0.2109925 0.2535601
#>            [,8]      [,9]     [,10]
#>  [1,] 0.3869399 0.4138831 0.3292069
#>  [2,] 0.4528597 0.4917679 0.3906111
#>  [3,] 0.5403963 0.5726611 0.4407477
#>  [4,] 0.6052741 0.6397207 0.4846572
#>  [5,] 0.5980640 0.6281681 0.4665404
#>  [6,] 0.5809970 0.5835667 0.4144626
#>  [7,] 0.5752430 0.5653576 0.4048669
#>  [8,] 0.4692998 0.4501360 0.3180919
#>  [9,] 0.3993458 0.3698730 0.2625048
#> [10,] 0.2705014 0.2533138 0.1898240
pfocal_fast_gaussian_confidence(data)
#>            [,1]      [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
#>  [1,] 0.3018905 0.3752286 0.3707117 0.3703778 0.3501144 0.3419238 0.3982027
#>  [2,] 0.3165525 0.3808994 0.3815236 0.4027585 0.4032841 0.3884099 0.4452078
#>  [3,] 0.3024647 0.3471242 0.3514275 0.3696367 0.3790644 0.3879526 0.4435243
#>  [4,] 0.3218010 0.3815553 0.4040449 0.4161509 0.4161569 0.4378562 0.4794423
#>  [5,] 0.3627086 0.4642549 0.5055963 0.5094881 0.4857384 0.4810618 0.5143435
#>  [6,] 0.4199079 0.5339059 0.5309440 0.4903104 0.4568982 0.4609611 0.5444819
#>  [7,] 0.3949424 0.4904604 0.4519619 0.3885549 0.3569461 0.3848386 0.5131407
#>  [8,] 0.3002104 0.3963903 0.3932107 0.3482792 0.3160842 0.3496821 0.4555303
#>  [9,] 0.2436748 0.3499097 0.3876679 0.3458551 0.3032629 0.3395610 0.3939087
#> [10,] 0.2179750 0.2989033 0.3145061 0.2597600 0.2267362 0.2659386 0.2874093
#>            [,8]      [,9]     [,10]
#>  [1,] 0.4691816 0.4796769 0.3761719
#>  [2,] 0.5620709 0.6175999 0.4973909
#>  [3,] 0.5672158 0.6439249 0.5181671
#>  [4,] 0.5661539 0.6208945 0.4844656
#>  [5,] 0.5754763 0.5815285 0.4185594
#>  [6,] 0.6433341 0.6242810 0.4226414
#>  [7,] 0.6516553 0.6526266 0.4538605
#>  [8,] 0.5433117 0.5363254 0.3826294
#>  [9,] 0.3934378 0.3505409 0.2546602
#> [10,] 0.2565019 0.2127656 0.1666001
pfocal_fast_binomial(data, vertical_radius = 2)
#>            [,1]      [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
#>  [1,]  75.38928  94.83759  94.88231  93.25559  88.49982  86.64563 100.11020
#>  [2,]  81.56662  99.42751  99.90522 103.29896 102.92693 100.22770 115.19773
#>  [3,]  76.53348  89.94146  90.45991  94.56065  97.52168  99.52350 114.43372
#>  [4,]  80.20555  98.22965 103.69002 106.72937 107.49781 110.76551 121.81534
#>  [5,]  93.02256 121.23640 130.22224 129.77507 124.31051 123.25346 132.94879
#>  [6,] 106.70665 137.39529 136.55575 125.49848 115.97447 117.51890 139.05085
#>  [7,] 100.69133 126.53242 117.21712 100.07640  91.01134  99.23832 131.96749
#>  [8,]  77.05121 102.76526 102.07547  89.13734  80.26846  89.74063 116.89582
#>  [9,]  62.85121  91.24228  99.39439  88.03692  77.91489  86.83662 100.89031
#> [10,]  54.19761  76.39516  79.70752  66.37174  58.49059  67.22144  73.06699
#>            [,8]      [,9]     [,10]
#>  [1,] 119.20757 122.90023  95.01228
#>  [2,] 146.07906 160.92496 127.79138
#>  [3,] 147.00395 167.08078 133.73089
#>  [4,] 144.94170 158.01954 122.85485
#>  [5,] 149.97991 150.15761 107.73845
#>  [6,] 165.62217 160.82812 108.92682
#>  [7,] 168.24471 167.78829 115.73771
#>  [8,] 141.11700 138.58868  98.02870
#>  [9,] 102.37988  91.44283  65.77498
#> [10,]  65.04956  53.85213  40.85189
pfocal_fast_abs_rectangle(data, height = 2)
#>            [,1]     [,2]      [,3]     [,4]      [,5]      [,6]      [,7]
#>  [1,] 0.9123981 1.825356 1.4882854 1.555914 1.3357958 0.8951081 1.3444382
#>  [2,] 1.1937000 2.477146 1.9118970 1.870729 2.3159353 1.8284246 1.7585128
#>  [3,] 0.8868184 1.290504 0.9793677 1.218780 1.6932406 1.4636563 0.8441996
#>  [4,] 1.3577055 1.769887 1.1749902 1.385627 1.0022966 1.4634804 1.7629325
#>  [5,] 1.1008372 1.559697 1.4201313 2.190042 1.8320667 1.8298971 2.0447433
#>  [6,] 1.2464545 2.205581 2.4595985 2.647457 2.5561280 2.1241463 1.9761573
#>  [7,] 1.5773677 2.869574 2.4716536 1.697944 1.4743582 1.4517020 1.6331532
#>  [8,] 1.1056982 2.150900 1.6505719 1.089479 1.1556069 0.8920131 1.2003061
#>  [9,] 0.4559630 1.097315 1.5172508 1.647542 1.4043990 1.0545056 1.6968806
#> [10,] 0.8138401 1.365023 1.9097849 2.187618 0.9972596 1.2163947 1.8692600
#>           [,8]     [,9]     [,10]
#>  [1,] 1.743465 1.622316 1.5252301
#>  [2,] 2.585281 2.886813 3.0929680
#>  [3,] 1.649040 2.583561 2.9612354
#>  [4,] 1.964636 2.949743 3.2929667
#>  [5,] 1.669419 2.346112 2.7380224
#>  [6,] 2.327886 2.508848 2.0669133
#>  [7,] 2.947368 3.568731 2.8515245
#>  [8,] 2.377275 3.186225 3.1432623
#>  [9,] 1.936443 1.748952 1.7372344
#> [10,] 1.595591 0.979646 0.9159418

pfocal_fast_separated(data, 
                      kernel_list = list(binomial_kernel(vertical_radius = 2, 
                                                         horizontal_radius = 2), 
                                         distance_kernel(vertical_radius = 2,
                                                         horizontal_radius = 2)))
#>           [,1]     [,2]     [,3]     [,4]     [,5]     [,6]     [,7]     [,8]
#>  [1,] 1340.217 1764.583 2400.797 2528.287 2650.184 2936.107 3210.701 3216.004
#>  [2,] 1800.254 2379.322 3236.239 3421.428 3581.976 3930.896 4259.097 4216.042
#>  [3,] 2624.087 3483.020 4655.356 4906.325 5084.504 5468.226 5818.037 5694.780
#>  [4,] 2842.188 3732.559 4957.872 5211.209 5393.517 5806.958 6196.156 6077.442
#>  [5,] 2864.922 3685.921 4874.360 5118.560 5315.686 5755.972 6180.316 6102.277
#>  [6,] 2853.854 3647.604 4820.597 5105.502 5313.943 5682.189 5994.206 5853.499
#>  [7,] 2854.277 3638.516 4775.782 5086.329 5268.606 5497.300 5633.338 5400.892
#>  [8,] 2674.340 3347.073 4328.768 4588.388 4722.675 4866.761 4958.484 4781.373
#>  [9,] 1896.988 2331.284 3003.008 3198.226 3319.856 3446.214 3566.063 3520.855
#> [10,] 1322.556 1642.386 2114.065 2292.915 2392.080 2431.502 2473.161 2443.302
#>           [,9]    [,10]
#>  [1,] 2505.838 2100.187
#>  [2,] 3233.349 2696.198
#>  [3,] 4385.149 3590.987
#>  [4,] 4732.250 3876.614
#>  [5,] 4822.183 3966.645
#>  [6,] 4630.999 3790.863
#>  [7,] 4243.002 3421.605
#>  [8,] 3772.306 3006.114
#>  [9,] 2818.957 2250.767
#> [10,] 1949.437 1513.065